Network.AWS.ElastiCache.AddTagsToResource

Description

The AddTagsToResource action adds up to 10 cost allocation tags to the named resource. A cost allocation tag is a key-value pair where the key and value are case-sensitive. Cost allocation tags can be used to categorize and track your AWS costs.

When you apply tags to your ElastiCache resources, AWS generates a cost allocation report as a comma-separated value (CSV) file with your usage and costs aggregated by your tags. You can apply tags that represent business categories (such as cost centers, application names, or owners) to organize your costs across multiple services. For more information, see Using CostAllocation Tags in Amazon ElastiCache.

http://docs.aws.amazon.com/AmazonElastiCache/latest/APIReference/API_AddTagsToResource.html

Request

data AddTagsToResource

Request constructor

addTagsToResource

Arguments

:: Text
attrResourceName

-> AddTagsToResource

AddTagsToResource constructor.

The fields accessible through corresponding lenses are:

attrResourceName :: Text

attrTags :: [Tag]

Request lenses

attrResourceName :: Lens' AddTagsToResource Text

The name of the resource to which the tags are to be added, for example 'arn:aws:elasticache:us-west-2:0123456789:cluster:myCluster'.

attrTags :: Lens' AddTagsToResource [Tag]

A list of cost allocation tags to be added to this resource. A tag is a key-value pair. A tag key must be accompanied by a tag value.

Response

data AddTagsToResourceResponse

Response constructor

addTagsToResourceResponse :: AddTagsToResourceResponse

AddTagsToResourceResponse constructor.

The fields accessible through corresponding lenses are:

attrrTagList :: [Tag]

Response lenses

attrrTagList :: Lens' AddTagsToResourceResponse [Tag]

A list of cost allocation tags as key-value pairs.