Network.AWS.ElastiCache.AuthorizeCacheSecurityGroupIngress

Description

The AuthorizeCacheSecurityGroupIngress action allows network ingress to a cache security group. Applications using ElastiCache must be running on Amazon EC2, and Amazon EC2 security groups are used as the authorization mechanism.

You cannot authorize ingress from an Amazon EC2 security group in one region to an ElastiCache cluster in another region.

http://docs.aws.amazon.com/AmazonElastiCache/latest/APIReference/API_AuthorizeCacheSecurityGroupIngress.html

Request

data AuthorizeCacheSecurityGroupIngress

Request constructor

authorizeCacheSecurityGroupIngress

Arguments

:: Text
acsgiCacheSecurityGroupName

-> Text
acsgiEC2SecurityGroupName

-> Text
acsgiEC2SecurityGroupOwnerId

-> AuthorizeCacheSecurityGroupIngress

AuthorizeCacheSecurityGroupIngress constructor.

The fields accessible through corresponding lenses are:

acsgiCacheSecurityGroupName :: Text

acsgiEC2SecurityGroupName :: Text

acsgiEC2SecurityGroupOwnerId :: Text

Request lenses

acsgiCacheSecurityGroupName :: Lens' AuthorizeCacheSecurityGroupIngress Text

The cache security group which will allow network ingress.

acsgiEC2SecurityGroupName :: Lens' AuthorizeCacheSecurityGroupIngress Text

The Amazon EC2 security group to be authorized for ingress to the cache security group.

acsgiEC2SecurityGroupOwnerId :: Lens' AuthorizeCacheSecurityGroupIngress Text

The AWS account number of the Amazon EC2 security group owner. Note that this is not the same thing as an AWS access key ID - you must provide a valid AWS account number for this parameter.

Response

data AuthorizeCacheSecurityGroupIngressResponse

Response constructor

authorizeCacheSecurityGroupIngressResponse :: AuthorizeCacheSecurityGroupIngressResponse

AuthorizeCacheSecurityGroupIngressResponse constructor.

The fields accessible through corresponding lenses are:

acsgirCacheSecurityGroup :: Maybe CacheSecurityGroup

Response lenses

acsgirCacheSecurityGroup :: Lens' AuthorizeCacheSecurityGroupIngressResponse (Maybe CacheSecurityGroup)