Network.AWS.ElastiCache.CreateCacheCluster

Description

The CreateCacheCluster operation creates a cache cluster. All nodes in the cache cluster run the same protocol-compliant cache engine software, either Memcached or Redis.

http://docs.aws.amazon.com/AmazonElastiCache/latest/APIReference/API_CreateCacheCluster.html

Request

data CreateCacheCluster

Request constructor

createCacheCluster

Arguments

:: Text
cccCacheClusterId

-> CreateCacheCluster

CreateCacheCluster constructor.

The fields accessible through corresponding lenses are:

cccAZMode :: Maybe Text

cccAutoMinorVersionUpgrade :: Maybe Bool

cccCacheClusterId :: Text

cccCacheNodeType :: Maybe Text

cccCacheParameterGroupName :: Maybe Text

cccCacheSecurityGroupNames :: [Text]

cccCacheSubnetGroupName :: Maybe Text

cccEngine :: Maybe Text

cccEngineVersion :: Maybe Text

cccNotificationTopicArn :: Maybe Text

cccNumCacheNodes :: Maybe Int

cccPort :: Maybe Int

cccPreferredAvailabilityZone :: Maybe Text

cccPreferredAvailabilityZones :: [Text]

cccPreferredMaintenanceWindow :: Maybe Text

cccReplicationGroupId :: Maybe Text

cccSecurityGroupIds :: [Text]

cccSnapshotArns :: [Text]

cccSnapshotName :: Maybe Text

cccSnapshotRetentionLimit :: Maybe Int

cccSnapshotWindow :: Maybe Text

Request lenses

cccAZMode :: Lens' CreateCacheCluster (Maybe Text)

Specifies whether the nodes in this Memcached node group are created in a single Availability Zone or created across multiple Availability Zones in the cluster's region. This parameter is only supported for Memcached cache clusters. If the AZMode and PreferredAvailabilityZones are not specified, ElastiCache assumes single-az mode.

cccAutoMinorVersionUpgrade :: Lens' CreateCacheCluster (Maybe Bool)

Determines whether minor engine upgrades will be applied automatically to the node group during the maintenance window. A value of true allows these upgrades to occur; false disables automatic upgrades. Default: true.

cccCacheClusterId :: Lens' CreateCacheCluster Text

The node group identifier. This parameter is stored as a lowercase string. Constraints: A name must contain from 1 to 20 alphanumeric characters or hyphens. The first character must be a letter. A name cannot end with a hyphen or contain two consecutive hyphens.

cccCacheNodeType :: Lens' CreateCacheCluster (Maybe Text)

The compute and memory capacity of the nodes in the node group. Valid node types are as follows: General purpose: Current generation: cache.t2.micro, cache.t2.small, cache.t2.medium, cache.m3.medium, cache.m3.large, cache.m3.xlarge, cache.m3.2xlarge Previous generation: cache.t1.micro, cache.m1.small, cache.m1.medium, cache.m1.large, cache.m1.xlarge Compute optimized: cache.c1.xlarge Memory optimized Current generation: cache.r3.large, cache.r3.xlarge, cache.r3.2xlarge, cache.r3.4xlarge, cache.r3.8xlarge Previous generation: cache.m2.xlarge, cache.m2.2xlarge, cache.m2.4xlarge Notes: All t2 instances are created in an Amazon Virtual Private Cloud (VPC). Redis backup/restore is not supported for t2 instances. Redis Append-only files (AOF) functionality is not supported for t1 or t2 instances. For a complete listing of cache node types and specifications, see Amazon ElastiCache Product Features and Details and Cache Node Type-Specific Parameters for Memcached or Cache Node Type-Specific Parameters for Redis.

cccCacheParameterGroupName :: Lens' CreateCacheCluster (Maybe Text)

The name of the parameter group to associate with this cache cluster. If this argument is omitted, the default parameter group for the specified engine is used.

cccCacheSecurityGroupNames :: Lens' CreateCacheCluster [Text]

A list of security group names to associate with this cache cluster. Use this parameter only when you are creating a cache cluster outside of an Amazon Virtual Private Cloud (VPC).

cccCacheSubnetGroupName :: Lens' CreateCacheCluster (Maybe Text)

The name of the subnet group to be used for the cache cluster. Use this parameter only when you are creating a cache cluster in an Amazon Virtual Private Cloud (VPC).

cccEngine :: Lens' CreateCacheCluster (Maybe Text)

The name of the cache engine to be used for this cache cluster. Valid values for this parameter are: memcached | redis.

cccEngineVersion :: Lens' CreateCacheCluster (Maybe Text)

The version number of the cache engine to be used for this cache cluster. To view the supported cache engine versions, use the DescribeCacheEngineVersions operation.

cccNotificationTopicArn :: Lens' CreateCacheCluster (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon Simple Notification Service (SNS) topic to which notifications will be sent.

cccNumCacheNodes :: Lens' CreateCacheCluster (Maybe Int)

The initial number of cache nodes that the cache cluster will have. For Memcached, valid values are between 1 and 20. If you need to exceed this limit, please fill out the ElastiCache Limit Increase Request form at http:/aws.amazon.comcontact-uselasticache-node-limit-request. For Redis, only single-node cache cluster are supported at this time, so the value for this parameter must be 1.

cccPort :: Lens' CreateCacheCluster (Maybe Int)

The port number on which each of the cache nodes will accept connections.

cccPreferredAvailabilityZone :: Lens' CreateCacheCluster (Maybe Text)

The EC2 Availability Zone in which the cache cluster will be created. All nodes belonging to this Memcached cache cluster are placed in the preferred Availability Zone. If you want to create your nodes across multiple Availability Zones, use PreferredAvailabilityZones. Default: System chosen Availability Zone.

cccPreferredAvailabilityZones :: Lens' CreateCacheCluster [Text]

A list of the Availability Zones in which cache nodes will be created. The order of the zones in the list is not important. This option is only supported on Memcached. If you want all the nodes in the same Availability Zone, use PreferredAvailabilityZone instead, or repeat the Availability Zone multiple times in the list. Default: System chosen Availability Zones. Example: One Memcached node in each of three different Availability Zones: PreferredAvailabilityZones.member.1=us-east-1a&amp;PreferredAvailabilityZones.member.2=us-east-1b&amp;PreferredAvailabilityZones.member.3=us-east-1d Example: All three Memcached nodes in one Availability Zone: PreferredAvailabilityZones.member.1=us-east-1a&amp;PreferredAvailabilityZones.member.2=us-east-1a&amp;PreferredAvailabilityZones.member.3=us-east-1a.

cccPreferredMaintenanceWindow :: Lens' CreateCacheCluster (Maybe Text)

The weekly time range (in UTC) during which system maintenance can occur. Example: sun:05:00-sun:09:00.

cccReplicationGroupId :: Lens' CreateCacheCluster (Maybe Text)

The ID of the replication group to which this cache cluster should belong. If this parameter is specified, the cache cluster will be added to the specified replication group as a read replica; otherwise, the cache cluster will be a standalone primary that is not part of any replication group. If the specified replication group is Automatic Failover enabled and the availability zone is not specified, the cache cluster will be created in availability zones that provide the best spread of read replicas across availability zones. Note: This parameter is only valid if the Engine parameter is redis.

cccSecurityGroupIds :: Lens' CreateCacheCluster [Text]

One or more VPC security groups associated with the cache cluster. Use this parameter only when you are creating a cache cluster in an Amazon Virtual Private Cloud (VPC).

cccSnapshotArns :: Lens' CreateCacheCluster [Text]

A single-element string list containing an Amazon Resource Name (ARN) that uniquely identifies a Redis RDB snapshot file stored in Amazon S3. The snapshot file will be used to populate the node group. The Amazon S3 object name in the ARN cannot contain any commas. Note: This parameter is only valid if the Engine parameter is redis. Example of an Amazon S3 ARN: arn:aws:s3:::my_bucket/snapshot1.rdb.

cccSnapshotName :: Lens' CreateCacheCluster (Maybe Text)

The name of a snapshot from which to restore data into the new node group. The snapshot status changes to restoring while the new node group is being created. Note: This parameter is only valid if the Engine parameter is redis.

cccSnapshotRetentionLimit :: Lens' CreateCacheCluster (Maybe Int)

The number of days for which ElastiCache will retain automatic snapshots before deleting them. For example, if you set SnapshotRetentionLimit to 5, then a snapshot that was taken today will be retained for 5 days before being deleted. Note: This parameter is only valid if the Engine parameter is redis. Default: 0 (i.e., automatic backups are disabled for this cache cluster).

cccSnapshotWindow :: Lens' CreateCacheCluster (Maybe Text)

The daily time range (in UTC) during which ElastiCache will begin taking a daily snapshot of your node group. Example: 05:00-09:00 If you do not specify this parameter, then ElastiCache will automatically choose an appropriate time range. Note: This parameter is only valid if the Engine parameter is redis.

Response

data CreateCacheClusterResponse

Response constructor

createCacheClusterResponse :: CreateCacheClusterResponse

CreateCacheClusterResponse constructor.

The fields accessible through corresponding lenses are:

cccrCacheCluster :: Maybe CacheCluster

Response lenses

cccrCacheCluster :: Lens' CreateCacheClusterResponse (Maybe CacheCluster)