Network.AWS.ElastiCache.CreateCacheParameterGroup

Description

The CreateCacheParameterGroup action creates a new cache parameter group. A cache parameter group is a collection of parameters that you apply to all of the nodes in a cache cluster.

http://docs.aws.amazon.com/AmazonElastiCache/latest/APIReference/API_CreateCacheParameterGroup.html

Request

data CreateCacheParameterGroup

Request constructor

createCacheParameterGroup

Arguments

:: Text
ccpgCacheParameterGroupName

-> Text
ccpgCacheParameterGroupFamily

-> Text
ccpgDescription

-> CreateCacheParameterGroup

CreateCacheParameterGroup constructor.

The fields accessible through corresponding lenses are:

ccpgCacheParameterGroupFamily :: Text

ccpgCacheParameterGroupName :: Text

ccpgDescription :: Text

Request lenses

ccpgCacheParameterGroupFamily :: Lens' CreateCacheParameterGroup Text

The name of the cache parameter group family the cache parameter group can be used with.

Valid values are: 'memcached1.4' | 'redis2.6' | 'redis2.8'

ccpgCacheParameterGroupName :: Lens' CreateCacheParameterGroup Text

A user-specified name for the cache parameter group.

ccpgDescription :: Lens' CreateCacheParameterGroup Text

A user-specified description for the cache parameter group.

Response

data CreateCacheParameterGroupResponse

Response constructor

createCacheParameterGroupResponse :: CreateCacheParameterGroupResponse

CreateCacheParameterGroupResponse constructor.

The fields accessible through corresponding lenses are:

ccpgrCacheParameterGroup :: Maybe CacheParameterGroup

Response lenses

ccpgrCacheParameterGroup :: Lens' CreateCacheParameterGroupResponse (Maybe CacheParameterGroup)