Network.AWS.ElastiCache.CreateCacheSecurityGroup

Description

The CreateCacheSecurityGroup operation creates a new cache security group. Use a cache security group to control access to one or more cache clusters. Cache security groups are only used when you are creating a cluster outside of an Amazon Virtual Private Cloud (VPC). If you are creating a cluster inside of a VPC, use a cache subnet group instead. For more information, see CreateCacheSubnetGroup. https:/elasticache.us-east-1.amazonaws.com ?Action=CreateCacheSecurityGroup &CacheSecurityGroupName=mycachesecuritygroup &Description=My%20cache%20security%20group &Version=2014-03-24 &SignatureVersion=4 &SignatureMethod=HmacSHA256 &Timestamp=20140401T192317Z &X-Amz-Credential= mycachesecuritygroup 123456789012 My cache security group 2b1c8035-b7fa-11e0-9326-b7275b9d4a6c.

Request

data CreateCacheSecurityGroup

Represents the input of a CreateCacheSecurityGroup operation.

Request constructor

createCacheSecurityGroup

Arguments

:: Text
ccsgCacheSecurityGroupName

-> Text
ccsgDescription

-> CreateCacheSecurityGroup

Smart constructor for the minimum required parameters to construct a valid CreateCacheSecurityGroup request.

The fields accessible through corresponding lenses are:

CacheSecurityGroupName :: Text

Description :: Text

Request lenses

ccsgCacheSecurityGroupName :: Lens' CreateCacheSecurityGroup Text

A name for the cache security group. This value is stored as a lowercase string. Constraints: Must contain no more than 255 alphanumeric characters. Must not be the word Default. Example: mysecuritygroup.

ccsgDescription :: Lens' CreateCacheSecurityGroup Text

A description for the cache security group.

Response

data CreateCacheSecurityGroupResponse

Response constructor

createCacheSecurityGroupResponse :: CreateCacheSecurityGroupResponse

Smart constructor for the minimum required parameters to construct a valid CreateCacheSecurityGroupResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

CacheSecurityGroup :: Maybe CacheSecurityGroup

Response lenses

ccsgrCacheSecurityGroup :: Lens' CreateCacheSecurityGroupResponse (Maybe CacheSecurityGroup)

Represents the output of one of the following operations: AuthorizeCacheSecurityGroupIngress CreateCacheSecurityGroup RevokeCacheSecurityGroupIngress.