Network.AWS.ElastiCache.CreateCacheSubnetGroup

Description

The CreateCacheSubnetGroup action creates a new cache subnet group.

Use this parameter only when you are creating a cluster in an Amazon Virtual Private Cloud (VPC).

http://docs.aws.amazon.com/AmazonElastiCache/latest/APIReference/API_CreateCacheSubnetGroup.html

Request

data CreateCacheSubnetGroup

Request constructor

createCacheSubnetGroup

Arguments

:: Text
ccsgCacheSubnetGroupName

-> Text
ccsgCacheSubnetGroupDescription

-> CreateCacheSubnetGroup

CreateCacheSubnetGroup constructor.

The fields accessible through corresponding lenses are:

ccsgCacheSubnetGroupDescription :: Text

ccsgCacheSubnetGroupName :: Text

ccsgSubnetIds :: [Text]

Request lenses

ccsgCacheSubnetGroupDescription :: Lens' CreateCacheSubnetGroup Text

A description for the cache subnet group.

ccsgCacheSubnetGroupName :: Lens' CreateCacheSubnetGroup Text

A name for the cache subnet group. This value is stored as a lowercase string.

Constraints: Must contain no more than 255 alphanumeric characters or hyphens.

Example: mysubnetgroup

ccsgSubnetIds :: Lens' CreateCacheSubnetGroup [Text]

A list of VPC subnet IDs for the cache subnet group.

Response

data CreateCacheSubnetGroupResponse

Response constructor

createCacheSubnetGroupResponse :: CreateCacheSubnetGroupResponse

CreateCacheSubnetGroupResponse constructor.

The fields accessible through corresponding lenses are:

ccsgrCacheSubnetGroup :: Maybe CacheSubnetGroup

Response lenses

ccsgrCacheSubnetGroup :: Lens' CreateCacheSubnetGroupResponse (Maybe CacheSubnetGroup)