Network.AWS.ElastiCache.CreateReplicationGroup

Description

The CreateReplicationGroup action creates a replication group. A replication group is a collection of cache clusters, where one of the cache clusters is a read/write primary and the others are read-only replicas. Writes to the primary are automatically propagated to the replicas.

When you create a replication group, you must specify an existing cache cluster that is in the primary role. When the replication group has been successfully created, you can add one or more read replica replicas to it, up to a total of five read replicas.

Note: This action is valid only for Redis.

http://docs.aws.amazon.com/AmazonElastiCache/latest/APIReference/API_CreateReplicationGroup.html

Request

data CreateReplicationGroup

Request constructor

createReplicationGroup

Arguments

:: Text
crgReplicationGroupId

-> Text
crgReplicationGroupDescription

-> CreateReplicationGroup

CreateReplicationGroup constructor.

The fields accessible through corresponding lenses are:

crgAutoMinorVersionUpgrade :: Maybe Bool

crgAutomaticFailoverEnabled :: Maybe Bool

crgCacheNodeType :: Maybe Text

crgCacheParameterGroupName :: Maybe Text

crgCacheSecurityGroupNames :: [Text]

crgCacheSubnetGroupName :: Maybe Text

crgEngine :: Maybe Text

crgEngineVersion :: Maybe Text

crgNotificationTopicArn :: Maybe Text

crgNumCacheClusters :: Maybe Int

crgPort :: Maybe Int

crgPreferredCacheClusterAZs :: [Text]

crgPreferredMaintenanceWindow :: Maybe Text

crgPrimaryClusterId :: Maybe Text

crgReplicationGroupDescription :: Text

crgReplicationGroupId :: Text

crgSecurityGroupIds :: [Text]

crgSnapshotArns :: [Text]

crgSnapshotName :: Maybe Text

crgSnapshotRetentionLimit :: Maybe Int

crgSnapshotWindow :: Maybe Text

crgTags :: [Tag]

Request lenses

crgAutoMinorVersionUpgrade :: Lens' CreateReplicationGroup (Maybe Bool)

This parameter is currently disabled.

crgAutomaticFailoverEnabled :: Lens' CreateReplicationGroup (Maybe Bool)

Specifies whether a read-only replica will be automatically promoted to read/write primary if the existing primary fails.

If true, Multi-AZ is enabled for this replication group. If false, Multi-AZ is disabled for this replication group.

Default: false

ElastiCache Multi-AZ replication groups is not supported on:

Redis versions earlier than 2.8.6. T1 and T2 cache node types.

crgCacheNodeType :: Lens' CreateReplicationGroup (Maybe Text)

The compute and memory capacity of the nodes in the node group.

Valid node types are as follows:

General purpose: Current generation: 'cache.t2.micro', 'cache.t2.small', 'cache.t2.medium', 'cache.m3.medium', 'cache.m3.large', 'cache.m3.xlarge', 'cache.m3.2xlarge' Previous generation: 'cache.t1.micro', 'cache.m1.small', 'cache.m1.medium', 'cache.m1.large', 'cache.m1.xlarge' Compute optimized: 'cache.c1.xlarge' Memory optimized Current generation: 'cache.r3.large', 'cache.r3.xlarge', 'cache.r3.2xlarge', 'cache.r3.4xlarge', 'cache.r3.8xlarge' Previous generation: 'cache.m2.xlarge', 'cache.m2.2xlarge', 'cache.m2.4xlarge' Notes:

All t2 instances are created in an Amazon Virtual Private Cloud (VPC). Redis backup/restore is not supported for t2 instances. Redis Append-only files (AOF) functionality is not supported for t1 or t2 instances. For a complete listing of cache node types and specifications, see Amazon ElastiCache Product Features and Details and Cache NodeType-Specific Parameters for Memcached or Cache Node Type-Specific Parametersfor Redis.

crgCacheParameterGroupName :: Lens' CreateReplicationGroup (Maybe Text)

The name of the parameter group to associate with this replication group. If this argument is omitted, the default cache parameter group for the specified engine is used.

crgCacheSecurityGroupNames :: Lens' CreateReplicationGroup [Text]

A list of cache security group names to associate with this replication group.

crgCacheSubnetGroupName :: Lens' CreateReplicationGroup (Maybe Text)

The name of the cache subnet group to be used for the replication group.

crgEngine :: Lens' CreateReplicationGroup (Maybe Text)

The name of the cache engine to be used for the cache clusters in this replication group.

Default: redis

crgEngineVersion :: Lens' CreateReplicationGroup (Maybe Text)

The version number of the cache engine to be used for the cache clusters in this replication group. To view the supported cache engine versions, use the DescribeCacheEngineVersions action.

crgNotificationTopicArn :: Lens' CreateReplicationGroup (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon Simple Notification Service (SNS) topic to which notifications will be sent.

The Amazon SNS topic owner must be the same as the cache cluster owner.

crgNumCacheClusters :: Lens' CreateReplicationGroup (Maybe Int)

The number of cache clusters this replication group will initially have.

If Multi-AZ is enabled, the value of this parameter must be at least 2.

The maximum permitted value for NumCacheClusters is 6 (primary plus 5 replicas). If you need to exceed this limit, please fill out the ElastiCache Limit Increase Request form at http://aws.amazon.com/contact-us/elasticache-node-limit-request.

crgPort :: Lens' CreateReplicationGroup (Maybe Int)

The port number on which each member of the replication group will accept connections.

crgPreferredCacheClusterAZs :: Lens' CreateReplicationGroup [Text]

A list of EC2 availability zones in which the replication group's cache clusters will be created. The order of the availability zones in the list is not important.

If you are creating your replication group in an Amazon VPC (recommended), you can only locate cache clusters in availability zones associated with the subnets in the selected subnet group. The number of availability zones listed must equal the value of NumCacheClusters.

Default: system chosen availability zones.

Example: One Redis cache cluster in each of three availability zones. PreferredAvailabilityZones.member.1=us-west-2a PreferredAvailabilityZones.member.2=us-west-2c PreferredAvailabilityZones.member.3=us-west-2c

crgPreferredMaintenanceWindow :: Lens' CreateReplicationGroup (Maybe Text)

Specifies the weekly time range during which maintenance on the cache cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period. Valid values for ddd are:

sun mon tue wed thu fri sat Example: 'sun:05:00-sun:09:00'

crgPrimaryClusterId :: Lens' CreateReplicationGroup (Maybe Text)

The identifier of the cache cluster that will serve as the primary for this replication group. This cache cluster must already exist and have a status of available.

This parameter is not required if NumCacheClusters is specified.

crgReplicationGroupDescription :: Lens' CreateReplicationGroup Text

A user-created description for the replication group.

crgReplicationGroupId :: Lens' CreateReplicationGroup Text

The replication group identifier. This parameter is stored as a lowercase string.

Constraints:

A name must contain from 1 to 20 alphanumeric characters or hyphens. The first character must be a letter. A name cannot end with a hyphen or contain two consecutive hyphens.

crgSecurityGroupIds :: Lens' CreateReplicationGroup [Text]

One or more Amazon VPC security groups associated with this replication group.

Use this parameter only when you are creating a replication group in an Amazon Virtual Private Cloud (VPC).

crgSnapshotArns :: Lens' CreateReplicationGroup [Text]

A single-element string list containing an Amazon Resource Name (ARN) that uniquely identifies a Redis RDB snapshot file stored in Amazon S3. The snapshot file will be used to populate the node group. The Amazon S3 object name in the ARN cannot contain any commas.

Note: This parameter is only valid if the Engine parameter is redis.

Example of an Amazon S3 ARN: 'arn:aws:s3:::my_bucket/snapshot1.rdb'

crgSnapshotName :: Lens' CreateReplicationGroup (Maybe Text)

The name of a snapshot from which to restore data into the new node group. The snapshot status changes to restoring while the new node group is being created.

Note: This parameter is only valid if the Engine parameter is redis.

crgSnapshotRetentionLimit :: Lens' CreateReplicationGroup (Maybe Int)

The number of days for which ElastiCache will retain automatic snapshots before deleting them. For example, if you set SnapshotRetentionLimit to 5, then a snapshot that was taken today will be retained for 5 days before being deleted.

Note: This parameter is only valid if the Engine parameter is redis.

Default: 0 (i.e., automatic backups are disabled for this cache cluster).

crgSnapshotWindow :: Lens' CreateReplicationGroup (Maybe Text)

The daily time range (in UTC) during which ElastiCache will begin taking a daily snapshot of your node group.

Example: '05:00-09:00'

If you do not specify this parameter, then ElastiCache will automatically choose an appropriate time range.

Note: This parameter is only valid if the Engine parameter is redis.

crgTags :: Lens' CreateReplicationGroup [Tag]

A list of cost allocation tags to be added to this resource. A tag is a key-value pair. A tag key must be accompanied by a tag value.

Response

data CreateReplicationGroupResponse

Response constructor

createReplicationGroupResponse :: CreateReplicationGroupResponse

CreateReplicationGroupResponse constructor.

The fields accessible through corresponding lenses are:

crgrReplicationGroup :: Maybe ReplicationGroup

Response lenses

crgrReplicationGroup :: Lens' CreateReplicationGroupResponse (Maybe ReplicationGroup)