Network.AWS.ElastiCache.CreateSnapshot

Description

The CreateSnapshot action creates a copy of an entire cache cluster at a specific moment in time.

http://docs.aws.amazon.com/AmazonElastiCache/latest/APIReference/API_CreateSnapshot.html

Request

data CreateSnapshot

Request constructor

createSnapshot

Arguments

:: Text
csCacheClusterId

-> Text
csSnapshotName

-> CreateSnapshot

CreateSnapshot constructor.

The fields accessible through corresponding lenses are:

csCacheClusterId :: Text

csSnapshotName :: Text

Request lenses

csCacheClusterId :: Lens' CreateSnapshot Text

The identifier of an existing cache cluster. The snapshot will be created from this cache cluster.

csSnapshotName :: Lens' CreateSnapshot Text

A name for the snapshot being created.

Response

data CreateSnapshotResponse

Response constructor

createSnapshotResponse :: CreateSnapshotResponse

CreateSnapshotResponse constructor.

The fields accessible through corresponding lenses are:

csr1Snapshot :: Maybe Snapshot

Response lenses

csr1Snapshot :: Lens' CreateSnapshotResponse (Maybe Snapshot)