Network.AWS.ElastiCache.DeleteCacheCluster

Description

Deletes a previously provisioned cache cluster. DeleteCacheCluster deletes all associated cache nodes, node endpoints and the cache cluster itself. When you receive a successful response from this operation, Amazon ElastiCache immediately begins deleting the cache cluster; you cannot cancel or revert this operation.

This operation cannot be used to delete a cache cluster that is the last read replica of a replication group or node group (shard) that has Multi-AZ mode enabled or a cache cluster from a Redis (cluster mode enabled) replication group.

Important: Due to current limitations on Redis (cluster mode disabled), this operation or parameter is not supported on Redis (cluster mode enabled) replication groups.

Creating a Request

deleteCacheCluster

Arguments

:: Text
dccCacheClusterId

-> DeleteCacheCluster

Creates a value of DeleteCacheCluster with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dccFinalSnapshotIdentifier - The user-supplied name of a final cache cluster snapshot. This is the unique name that identifies the snapshot. ElastiCache creates the snapshot, and then deletes the cache cluster immediately afterward.

dccCacheClusterId - The cache cluster identifier for the cluster to be deleted. This parameter is not case sensitive.

data DeleteCacheCluster

Represents the input of a DeleteCacheCluster operation.

See: deleteCacheCluster smart constructor.

Request Lenses

dccFinalSnapshotIdentifier :: Lens' DeleteCacheCluster (Maybe Text)

The user-supplied name of a final cache cluster snapshot. This is the unique name that identifies the snapshot. ElastiCache creates the snapshot, and then deletes the cache cluster immediately afterward.

dccCacheClusterId :: Lens' DeleteCacheCluster Text

The cache cluster identifier for the cluster to be deleted. This parameter is not case sensitive.

Destructuring the Response

deleteCacheClusterResponse

Arguments

:: Int
dccrsResponseStatus

-> DeleteCacheClusterResponse

Creates a value of DeleteCacheClusterResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dccrsCacheCluster - Undocumented member.

dccrsResponseStatus - -- | The response status code.

data DeleteCacheClusterResponse

See: deleteCacheClusterResponse smart constructor.

Response Lenses

dccrsCacheCluster :: Lens' DeleteCacheClusterResponse (Maybe CacheCluster)

Undocumented member.

dccrsResponseStatus :: Lens' DeleteCacheClusterResponse Int

- | The response status code.