Network.AWS.ElastiCache.DeleteCacheParameterGroup

Description

The DeleteCacheParameterGroup action deletes the specified cache parameter group. You cannot delete a cache parameter group if it is associated with any cache clusters.

http://docs.aws.amazon.com/AmazonElastiCache/latest/APIReference/API_DeleteCacheParameterGroup.html

Request

data DeleteCacheParameterGroup

Request constructor

deleteCacheParameterGroup

Arguments

:: Text
dcpg1CacheParameterGroupName

-> DeleteCacheParameterGroup

DeleteCacheParameterGroup constructor.

The fields accessible through corresponding lenses are:

dcpg1CacheParameterGroupName :: Text

Request lenses

dcpg1CacheParameterGroupName :: Lens' DeleteCacheParameterGroup Text

The name of the cache parameter group to delete.

The specified cache security group must not be associated with any cache clusters.

Response

data DeleteCacheParameterGroupResponse

Response constructor

deleteCacheParameterGroupResponse :: DeleteCacheParameterGroupResponse

DeleteCacheParameterGroupResponse constructor.