Network.AWS.ElastiCache.DeleteCacheSecurityGroup

Description

The DeleteCacheSecurityGroup operation deletes a cache security group. You cannot delete a cache security group if it is associated with any cache clusters. https:/elasticache.us-east-1.amazonaws.com ?Action=DeleteCacheSecurityGroup &CacheSecurityGroupName=mycachesecuritygroup3 &Version=2014-03-24 &SignatureVersion=4 &SignatureMethod=HmacSHA256 &Timestamp=20140401T192317Z &X-Amz-Credential= c130cfb7-3650-11e0-ae57-f96cfe56749c.

Request

data DeleteCacheSecurityGroup

Represents the input of a DeleteCacheSecurityGroup operation.

Request constructor

deleteCacheSecurityGroup

Arguments

:: Text
dcsgCacheSecurityGroupName

-> DeleteCacheSecurityGroup

Smart constructor for the minimum required parameters to construct a valid DeleteCacheSecurityGroup request.

The fields accessible through corresponding lenses are:

CacheSecurityGroupName :: Text

Request lenses

dcsgCacheSecurityGroupName :: Lens' DeleteCacheSecurityGroup Text

The name of the cache security group to delete. You cannot delete the default security group.

Response

data DeleteCacheSecurityGroupResponse

Response constructor

deleteCacheSecurityGroupResponse :: DeleteCacheSecurityGroupResponse

Smart constructor for the minimum required parameters to construct a valid DeleteCacheSecurityGroupResponse response.

This constructor is provided for convenience and testing purposes.