Network.AWS.ElastiCache.DescribeCacheClusters

Description

The DescribeCacheClusters action returns information about all provisioned cache clusters if no cache cluster identifier is specified, or about a specific cache cluster if a cache cluster identifier is supplied.

By default, abbreviated information about the cache clusters(s) will be returned. You can use the optional ShowDetails flag to retrieve detailed information about the cache nodes associated with the cache clusters. These details include the DNS address and port for the cache node endpoint.

If the cluster is in the CREATING state, only cluster level information will be displayed until all of the nodes are successfully provisioned.

If the cluster is in the DELETING state, only cluster level information will be displayed.

If cache nodes are currently being added to the cache cluster, node endpoint information and creation time for the additional nodes will not be displayed until they are completely provisioned. When the cache cluster state is available, the cluster is ready for use.

If cache nodes are currently being removed from the cache cluster, no endpoint information for the removed nodes is displayed.

http://docs.aws.amazon.com/AmazonElastiCache/latest/APIReference/API_DescribeCacheClusters.html

Request

data DescribeCacheClusters

Request constructor

describeCacheClusters :: DescribeCacheClusters

DescribeCacheClusters constructor.

The fields accessible through corresponding lenses are:

dcc1CacheClusterId :: Maybe Text

dcc1Marker :: Maybe Text

dcc1MaxRecords :: Maybe Int

dcc1ShowCacheNodeInfo :: Maybe Bool

Request lenses

dcc1CacheClusterId :: Lens' DescribeCacheClusters (Maybe Text)

The user-supplied cluster identifier. If this parameter is specified, only information about that specific cache cluster is returned. This parameter isn't case sensitive.

dcc1Marker :: Lens' DescribeCacheClusters (Maybe Text)

An optional marker returned from a prior request. Use this marker for pagination of results from this action. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

dcc1MaxRecords :: Lens' DescribeCacheClusters (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: minimum 20; maximum 100.

dcc1ShowCacheNodeInfo :: Lens' DescribeCacheClusters (Maybe Bool)

An optional flag that can be included in the DescribeCacheCluster request to retrieve information about the individual cache nodes.

Response

data DescribeCacheClustersResponse

Response constructor

describeCacheClustersResponse :: DescribeCacheClustersResponse

DescribeCacheClustersResponse constructor.

The fields accessible through corresponding lenses are:

dccrCacheClusters :: [CacheCluster]

dccrMarker :: Maybe Text

Response lenses

dccrCacheClusters :: Lens' DescribeCacheClustersResponse [CacheCluster]

A list of cache clusters. Each item in the list contains detailed information about one cache cluster.

dccrMarker :: Lens' DescribeCacheClustersResponse (Maybe Text)

Provides an identifier to allow retrieval of paginated results.