Network.AWS.ElastiCache.DescribeCacheEngineVersions

Description

The DescribeCacheEngineVersions operation returns a list of the available cache engines and their versions. https:/elasticache.us-east-1.amazonaws.com ?Action=DescribeCacheEngineVersions &MaxRecords=100 &Version=2014-03-24 &SignatureVersion=4 &SignatureMethod=HmacSHA256 &Timestamp=20140401T192317Z &X-Amz-Credential= memcached1.4 memcached memcached version 1.4.14 memcached 1.4.14 memcached1.4 memcached memcached version 1.4.5 memcached 1.4.5 a6ac9ad2-f8a4-11e1-a4d1-a345e5370093.

Request

data DescribeCacheEngineVersions

Represents the input of a DescribeCacheEngineVersions operation.

Request constructor

describeCacheEngineVersions :: DescribeCacheEngineVersions

Smart constructor for the minimum required parameters to construct a valid DescribeCacheEngineVersions request.

The fields accessible through corresponding lenses are:

Engine :: Maybe Text

EngineVersion :: Maybe Text

CacheParameterGroupFamily :: Maybe Text

MaxRecords :: Maybe Integer

Marker :: Maybe Text

DefaultOnly :: Maybe Bool

Request lenses

dcevEngine :: Lens' DescribeCacheEngineVersions (Maybe Text)

The cache engine to return. Valid values: memcached | redis.

dcevEngineVersion :: Lens' DescribeCacheEngineVersions (Maybe Text)

The cache engine version to return. Example: 1.4.14.

dcevCacheParameterGroupFamily :: Lens' DescribeCacheEngineVersions (Maybe Text)

The name of a specific cache parameter group family to return details for. Constraints: Must be 1 to 255 alphanumeric characters First character must be a letter Cannot end with a hyphen or contain two consecutive hyphens.

dcevMaxRecords :: Lens' DescribeCacheEngineVersions (Maybe Integer)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: minimum 20; maximum 100.

dcevMarker :: Lens' DescribeCacheEngineVersions (Maybe Text)

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

dcevDefaultOnly :: Lens' DescribeCacheEngineVersions (Maybe Bool)

If true, specifies that only the default version of the specified engine or engine and major version combination is to be returned.

Response

data DescribeCacheEngineVersionsResponse

Represents the output of a DescribeCacheEngineVersions operation.

Response constructor

describeCacheEngineVersionsResponse :: DescribeCacheEngineVersionsResponse

Smart constructor for the minimum required parameters to construct a valid DescribeCacheEngineVersionsResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Marker :: Maybe Text

CacheEngineVersions :: [CacheEngineVersion]

Response lenses

dcevrMarker :: Lens' DescribeCacheEngineVersionsResponse (Maybe Text)

Provides an identifier to allow retrieval of paginated results.

dcevrCacheEngineVersions :: Lens' DescribeCacheEngineVersionsResponse [CacheEngineVersion]

A list of cache engine version details. Each element in the list contains detailed information about once cache engine version.