Network.AWS.ElastiCache.DescribeCacheParameterGroups

Description

The DescribeCacheParameterGroups action returns a list of cache parameter group descriptions. If a cache parameter group name is specified, the list will contain only the descriptions for that group.

http://docs.aws.amazon.com/AmazonElastiCache/latest/APIReference/API_DescribeCacheParameterGroups.html

Request

data DescribeCacheParameterGroups

Request constructor

describeCacheParameterGroups :: DescribeCacheParameterGroups

DescribeCacheParameterGroups constructor.

The fields accessible through corresponding lenses are:

dcpgCacheParameterGroupName :: Maybe Text

dcpgMarker :: Maybe Text

dcpgMaxRecords :: Maybe Int

Request lenses

dcpgCacheParameterGroupName :: Lens' DescribeCacheParameterGroups (Maybe Text)

The name of a specific cache parameter group to return details for.

dcpgMarker :: Lens' DescribeCacheParameterGroups (Maybe Text)

An optional marker returned from a prior request. Use this marker for pagination of results from this action. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

dcpgMaxRecords :: Lens' DescribeCacheParameterGroups (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: minimum 20; maximum 100.

Response

data DescribeCacheParameterGroupsResponse

Response constructor

describeCacheParameterGroupsResponse :: DescribeCacheParameterGroupsResponse

DescribeCacheParameterGroupsResponse constructor.

The fields accessible through corresponding lenses are:

dcpgrCacheParameterGroups :: [CacheParameterGroup]

dcpgrMarker :: Maybe Text

Response lenses

dcpgrCacheParameterGroups :: Lens' DescribeCacheParameterGroupsResponse [CacheParameterGroup]

A list of cache parameter groups. Each element in the list contains detailed information about one cache parameter group.

dcpgrMarker :: Lens' DescribeCacheParameterGroupsResponse (Maybe Text)

Provides an identifier to allow retrieval of paginated results.