Network.AWS.ElastiCache.DescribeCacheParameters

Description

The DescribeCacheParameters action returns the detailed parameter list for a particular cache parameter group.

http://docs.aws.amazon.com/AmazonElastiCache/latest/APIReference/API_DescribeCacheParameters.html

Request

data DescribeCacheParameters

Request constructor

describeCacheParameters

Arguments

:: Text
dcpCacheParameterGroupName

-> DescribeCacheParameters

DescribeCacheParameters constructor.

The fields accessible through corresponding lenses are:

dcpCacheParameterGroupName :: Text

dcpMarker :: Maybe Text

dcpMaxRecords :: Maybe Int

dcpSource :: Maybe Text

Request lenses

dcpCacheParameterGroupName :: Lens' DescribeCacheParameters Text

The name of a specific cache parameter group to return details for.

dcpMarker :: Lens' DescribeCacheParameters (Maybe Text)

An optional marker returned from a prior request. Use this marker for pagination of results from this action. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

dcpMaxRecords :: Lens' DescribeCacheParameters (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: minimum 20; maximum 100.

dcpSource :: Lens' DescribeCacheParameters (Maybe Text)

The parameter types to return.

Valid values: user | system | 'engine-default'

Response

data DescribeCacheParametersResponse

Response constructor

describeCacheParametersResponse :: DescribeCacheParametersResponse

DescribeCacheParametersResponse constructor.

The fields accessible through corresponding lenses are:

dcprCacheNodeTypeSpecificParameters :: [CacheNodeTypeSpecificParameter]

dcprMarker :: Maybe Text

dcprParameters :: [Parameter]

Response lenses

dcprCacheNodeTypeSpecificParameters :: Lens' DescribeCacheParametersResponse [CacheNodeTypeSpecificParameter]

A list of parameters specific to a particular cache node type. Each element in the list contains detailed information about one parameter.

dcprMarker :: Lens' DescribeCacheParametersResponse (Maybe Text)

Provides an identifier to allow retrieval of paginated results.

dcprParameters :: Lens' DescribeCacheParametersResponse [Parameter]

A list of Parameter instances.