Network.AWS.ElastiCache.DescribeCacheSecurityGroups

Description

The DescribeCacheSecurityGroups action returns a list of cache security group descriptions. If a cache security group name is specified, the list will contain only the description of that group.

http://docs.aws.amazon.com/AmazonElastiCache/latest/APIReference/API_DescribeCacheSecurityGroups.html

Request

data DescribeCacheSecurityGroups

Request constructor

describeCacheSecurityGroups :: DescribeCacheSecurityGroups

DescribeCacheSecurityGroups constructor.

The fields accessible through corresponding lenses are:

dcsg1CacheSecurityGroupName :: Maybe Text

dcsg1Marker :: Maybe Text

dcsg1MaxRecords :: Maybe Int

Request lenses

dcsg1CacheSecurityGroupName :: Lens' DescribeCacheSecurityGroups (Maybe Text)

The name of the cache security group to return details for.

dcsg1Marker :: Lens' DescribeCacheSecurityGroups (Maybe Text)

An optional marker returned from a prior request. Use this marker for pagination of results from this action. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

dcsg1MaxRecords :: Lens' DescribeCacheSecurityGroups (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: minimum 20; maximum 100.

Response

data DescribeCacheSecurityGroupsResponse

Response constructor

describeCacheSecurityGroupsResponse :: DescribeCacheSecurityGroupsResponse

DescribeCacheSecurityGroupsResponse constructor.

The fields accessible through corresponding lenses are:

dcsgr1CacheSecurityGroups :: [CacheSecurityGroup]

dcsgr1Marker :: Maybe Text

Response lenses

dcsgr1CacheSecurityGroups :: Lens' DescribeCacheSecurityGroupsResponse [CacheSecurityGroup]

A list of cache security groups. Each element in the list contains detailed information about one group.

dcsgr1Marker :: Lens' DescribeCacheSecurityGroupsResponse (Maybe Text)

Provides an identifier to allow retrieval of paginated results.