Network.AWS.ElastiCache.DescribeCacheSubnetGroups

Description

The DescribeCacheSubnetGroups action returns a list of cache subnet group descriptions. If a subnet group name is specified, the list will contain only the description of that group.

http://docs.aws.amazon.com/AmazonElastiCache/latest/APIReference/API_DescribeCacheSubnetGroups.html

Request

data DescribeCacheSubnetGroups

Request constructor

describeCacheSubnetGroups :: DescribeCacheSubnetGroups

DescribeCacheSubnetGroups constructor.

The fields accessible through corresponding lenses are:

dcsgCacheSubnetGroupName :: Maybe Text

dcsgMarker :: Maybe Text

dcsgMaxRecords :: Maybe Int

Request lenses

dcsgCacheSubnetGroupName :: Lens' DescribeCacheSubnetGroups (Maybe Text)

The name of the cache subnet group to return details for.

dcsgMarker :: Lens' DescribeCacheSubnetGroups (Maybe Text)

An optional marker returned from a prior request. Use this marker for pagination of results from this action. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

dcsgMaxRecords :: Lens' DescribeCacheSubnetGroups (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: minimum 20; maximum 100.

Response

data DescribeCacheSubnetGroupsResponse

Response constructor

describeCacheSubnetGroupsResponse :: DescribeCacheSubnetGroupsResponse

DescribeCacheSubnetGroupsResponse constructor.

The fields accessible through corresponding lenses are:

dcsgrCacheSubnetGroups :: [CacheSubnetGroup]

dcsgrMarker :: Maybe Text

Response lenses

dcsgrCacheSubnetGroups :: Lens' DescribeCacheSubnetGroupsResponse [CacheSubnetGroup]

A list of cache subnet groups. Each element in the list contains detailed information about one group.

dcsgrMarker :: Lens' DescribeCacheSubnetGroupsResponse (Maybe Text)

Provides an identifier to allow retrieval of paginated results.