Network.AWS.ElastiCache.DescribeEngineDefaultParameters

Description

The DescribeEngineDefaultParameters operation returns the default engine and system parameter information for the specified cache engine. Some of the output has been omitted for brevity. https:/elasticache.us-east-1.amazonaws.com ?Action=DescribeEngineDefaultParameters &CacheParameterGroupFamily=memcached1.4 &MaxRecords=100 &Version=2014-03-24 &SignatureVersion=4 &SignatureMethod=HmacSHA256 &Timestamp=20140401T192317Z &X-Amz-Credential= memcached1.4 1024 integer system false The backlog queue limit. 1-10000 backlog_queue_limit 1.4.5 (...output omitted...) cache.c1.xlarge 6000 (...output omitted...) integer system false The maximum configurable amount of memory to use to store items, in megabytes. 1-100000 max_cache_memory 1.4.5 (...output omitted...) 061282fe-b7fd-11e0-9326-b7275b9d4a6c.

Request

data DescribeEngineDefaultParameters

Represents the input of a DescribeEngineDefaultParameters operation.

Request constructor

describeEngineDefaultParameters

Arguments

:: Text
dedpCacheParameterGroupFamily

-> DescribeEngineDefaultParameters

Smart constructor for the minimum required parameters to construct a valid DescribeEngineDefaultParameters request.

The fields accessible through corresponding lenses are:

CacheParameterGroupFamily :: Text

MaxRecords :: Maybe Integer

Marker :: Maybe Text

Request lenses

dedpCacheParameterGroupFamily :: Lens' DescribeEngineDefaultParameters Text

The name of the cache parameter group family. Valid values are: memcached1.4 | redis2.6 | redis2.8.

dedpMaxRecords :: Lens' DescribeEngineDefaultParameters (Maybe Integer)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: minimum 20; maximum 100.

dedpMarker :: Lens' DescribeEngineDefaultParameters (Maybe Text)

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Response

data DescribeEngineDefaultParametersResponse

Response constructor

describeEngineDefaultParametersResponse

Arguments

:: EngineDefaults
dedprEngineDefaults

-> DescribeEngineDefaultParametersResponse

Smart constructor for the minimum required parameters to construct a valid DescribeEngineDefaultParametersResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

EngineDefaults :: EngineDefaults

Response lenses

dedprEngineDefaults :: Lens' DescribeEngineDefaultParametersResponse EngineDefaults

Represents the output of a DescribeEngineDefaultParameters operation.