Network.AWS.ElastiCache.DescribeEvents

Description

The DescribeEvents action returns events related to cache clusters, cache security groups, and cache parameter groups. You can obtain events specific to a particular cache cluster, cache security group, or cache parameter group by providing the name as a parameter.

By default, only the events occurring within the last hour are returned; however, you can retrieve up to 14 days' worth of events if necessary.

http://docs.aws.amazon.com/AmazonElastiCache/latest/APIReference/API_DescribeEvents.html

Request

data DescribeEvents

Request constructor

describeEvents :: DescribeEvents

DescribeEvents constructor.

The fields accessible through corresponding lenses are:

deDuration :: Maybe Int

deEndTime :: Maybe UTCTime

deMarker :: Maybe Text

deMaxRecords :: Maybe Int

deSourceIdentifier :: Maybe Text

deSourceType :: Maybe SourceType

deStartTime :: Maybe UTCTime

Request lenses

deDuration :: Lens' DescribeEvents (Maybe Int)

The number of minutes' worth of events to retrieve.

deEndTime :: Lens' DescribeEvents (Maybe UTCTime)

The end of the time interval for which to retrieve events, specified in ISO 8601 format.

deMarker :: Lens' DescribeEvents (Maybe Text)

An optional marker returned from a prior request. Use this marker for pagination of results from this action. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

deMaxRecords :: Lens' DescribeEvents (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: minimum 20; maximum 100.

deSourceIdentifier :: Lens' DescribeEvents (Maybe Text)

The identifier of the event source for which events will be returned. If not specified, then all sources are included in the response.

deSourceType :: Lens' DescribeEvents (Maybe SourceType)

The event source to retrieve events for. If no value is specified, all events are returned.

Valid values are: 'cache-cluster' | 'cache-parameter-group' | 'cache-security-group' | 'cache-subnet-group'

deStartTime :: Lens' DescribeEvents (Maybe UTCTime)

The beginning of the time interval to retrieve events for, specified in ISO 8601 format.

Response

data DescribeEventsResponse

Response constructor

describeEventsResponse :: DescribeEventsResponse

DescribeEventsResponse constructor.

The fields accessible through corresponding lenses are:

derEvents :: [Event]

derMarker :: Maybe Text

Response lenses

derEvents :: Lens' DescribeEventsResponse [Event]

A list of events. Each element in the list contains detailed information about one event.

derMarker :: Lens' DescribeEventsResponse (Maybe Text)

Provides an identifier to allow retrieval of paginated results.