Network.AWS.ElastiCache.DescribeReservedCacheNodes

Description

The DescribeReservedCacheNodes action returns information about reserved cache nodes for this account, or about a specified reserved cache node.

http://docs.aws.amazon.com/AmazonElastiCache/latest/APIReference/API_DescribeReservedCacheNodes.html

Request

data DescribeReservedCacheNodes

Request constructor

describeReservedCacheNodes :: DescribeReservedCacheNodes

DescribeReservedCacheNodes constructor.

The fields accessible through corresponding lenses are:

drcnCacheNodeType :: Maybe Text

drcnDuration :: Maybe Text

drcnMarker :: Maybe Text

drcnMaxRecords :: Maybe Int

drcnOfferingType :: Maybe Text

drcnProductDescription :: Maybe Text

drcnReservedCacheNodeId :: Maybe Text

drcnReservedCacheNodesOfferingId :: Maybe Text

Request lenses

drcnCacheNodeType :: Lens' DescribeReservedCacheNodes (Maybe Text)

The cache node type filter value. Use this parameter to show only those reservations matching the specified cache node type.

Valid node types are as follows:

General purpose: Current generation: 'cache.t2.micro', 'cache.t2.small', 'cache.t2.medium', 'cache.m3.medium', 'cache.m3.large', 'cache.m3.xlarge', 'cache.m3.2xlarge' Previous generation: 'cache.t1.micro', 'cache.m1.small', 'cache.m1.medium', 'cache.m1.large', 'cache.m1.xlarge' Compute optimized: 'cache.c1.xlarge' Memory optimized Current generation: 'cache.r3.large', 'cache.r3.xlarge', 'cache.r3.2xlarge', 'cache.r3.4xlarge', 'cache.r3.8xlarge' Previous generation: 'cache.m2.xlarge', 'cache.m2.2xlarge', 'cache.m2.4xlarge' Notes:

All t2 instances are created in an Amazon Virtual Private Cloud (VPC). Redis backup/restore is not supported for t2 instances. Redis Append-only files (AOF) functionality is not supported for t1 or t2 instances. For a complete listing of cache node types and specifications, see Amazon ElastiCache Product Features and Details and Cache NodeType-Specific Parameters for Memcached or Cache Node Type-Specific Parametersfor Redis.

drcnDuration :: Lens' DescribeReservedCacheNodes (Maybe Text)

The duration filter value, specified in years or seconds. Use this parameter to show only reservations for this duration.

Valid Values: '1 | 3 | 31536000 | 94608000'

drcnMarker :: Lens' DescribeReservedCacheNodes (Maybe Text)

An optional marker returned from a prior request. Use this marker for pagination of results from this action. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

drcnMaxRecords :: Lens' DescribeReservedCacheNodes (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: minimum 20; maximum 100.

drcnOfferingType :: Lens' DescribeReservedCacheNodes (Maybe Text)

The offering type filter value. Use this parameter to show only the available offerings matching the specified offering type.

Valid values: '"Light Utilization"|"Medium Utilization"|"Heavy Utilization"'

drcnProductDescription :: Lens' DescribeReservedCacheNodes (Maybe Text)

The product description filter value. Use this parameter to show only those reservations matching the specified product description.

drcnReservedCacheNodeId :: Lens' DescribeReservedCacheNodes (Maybe Text)

The reserved cache node identifier filter value. Use this parameter to show only the reservation that matches the specified reservation ID.

drcnReservedCacheNodesOfferingId :: Lens' DescribeReservedCacheNodes (Maybe Text)

The offering identifier filter value. Use this parameter to show only purchased reservations matching the specified offering identifier.

Response

data DescribeReservedCacheNodesResponse

Response constructor

describeReservedCacheNodesResponse :: DescribeReservedCacheNodesResponse

DescribeReservedCacheNodesResponse constructor.

The fields accessible through corresponding lenses are:

drcnrMarker :: Maybe Text

drcnrReservedCacheNodes :: [ReservedCacheNode]

Response lenses

drcnrMarker :: Lens' DescribeReservedCacheNodesResponse (Maybe Text)

Provides an identifier to allow retrieval of paginated results.

drcnrReservedCacheNodes :: Lens' DescribeReservedCacheNodesResponse [ReservedCacheNode]

A list of reserved cache nodes. Each element in the list contains detailed information about one node.