Network.AWS.ElastiCache.DescribeReservedCacheNodesOfferings

Description

The DescribeReservedCacheNodesOfferings action lists available reserved cache node offerings.

http://docs.aws.amazon.com/AmazonElastiCache/latest/APIReference/API_DescribeReservedCacheNodesOfferings.html

Request

data DescribeReservedCacheNodesOfferings

Request constructor

describeReservedCacheNodesOfferings :: DescribeReservedCacheNodesOfferings

DescribeReservedCacheNodesOfferings constructor.

The fields accessible through corresponding lenses are:

drcnoCacheNodeType :: Maybe Text

drcnoDuration :: Maybe Text

drcnoMarker :: Maybe Text

drcnoMaxRecords :: Maybe Int

drcnoOfferingType :: Maybe Text

drcnoProductDescription :: Maybe Text

drcnoReservedCacheNodesOfferingId :: Maybe Text

Request lenses

drcnoCacheNodeType :: Lens' DescribeReservedCacheNodesOfferings (Maybe Text)

The cache node type filter value. Use this parameter to show only the available offerings matching the specified cache node type.

Valid node types are as follows:

General purpose: Current generation: 'cache.t2.micro', 'cache.t2.small', 'cache.t2.medium', 'cache.m3.medium', 'cache.m3.large', 'cache.m3.xlarge', 'cache.m3.2xlarge' Previous generation: 'cache.t1.micro', 'cache.m1.small', 'cache.m1.medium', 'cache.m1.large', 'cache.m1.xlarge' Compute optimized: 'cache.c1.xlarge' Memory optimized Current generation: 'cache.r3.large', 'cache.r3.xlarge', 'cache.r3.2xlarge', 'cache.r3.4xlarge', 'cache.r3.8xlarge' Previous generation: 'cache.m2.xlarge', 'cache.m2.2xlarge', 'cache.m2.4xlarge' Notes:

All t2 instances are created in an Amazon Virtual Private Cloud (VPC). Redis backup/restore is not supported for t2 instances. Redis Append-only files (AOF) functionality is not supported for t1 or t2 instances. For a complete listing of cache node types and specifications, see Amazon ElastiCache Product Features and Details and Cache NodeType-Specific Parameters for Memcached or Cache Node Type-Specific Parametersfor Redis.

drcnoDuration :: Lens' DescribeReservedCacheNodesOfferings (Maybe Text)

Duration filter value, specified in years or seconds. Use this parameter to show only reservations for a given duration.

Valid Values: '1 | 3 | 31536000 | 94608000'

drcnoMarker :: Lens' DescribeReservedCacheNodesOfferings (Maybe Text)

An optional marker returned from a prior request. Use this marker for pagination of results from this action. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

drcnoMaxRecords :: Lens' DescribeReservedCacheNodesOfferings (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: minimum 20; maximum 100.

drcnoOfferingType :: Lens' DescribeReservedCacheNodesOfferings (Maybe Text)

The offering type filter value. Use this parameter to show only the available offerings matching the specified offering type.

Valid Values: '"Light Utilization"|"Medium Utilization"|"Heavy Utilization"'

drcnoProductDescription :: Lens' DescribeReservedCacheNodesOfferings (Maybe Text)

The product description filter value. Use this parameter to show only the available offerings matching the specified product description.

drcnoReservedCacheNodesOfferingId :: Lens' DescribeReservedCacheNodesOfferings (Maybe Text)

The offering identifier filter value. Use this parameter to show only the available offering that matches the specified reservation identifier.

Example: '438012d3-4052-4cc7-b2e3-8d3372e0e706'

Response

data DescribeReservedCacheNodesOfferingsResponse

Response constructor

describeReservedCacheNodesOfferingsResponse :: DescribeReservedCacheNodesOfferingsResponse

DescribeReservedCacheNodesOfferingsResponse constructor.

The fields accessible through corresponding lenses are:

drcnorMarker :: Maybe Text

drcnorReservedCacheNodesOfferings :: [ReservedCacheNodesOffering]

Response lenses

drcnorMarker :: Lens' DescribeReservedCacheNodesOfferingsResponse (Maybe Text)

Provides an identifier to allow retrieval of paginated results.

drcnorReservedCacheNodesOfferings :: Lens' DescribeReservedCacheNodesOfferingsResponse [ReservedCacheNodesOffering]

A list of reserved cache node offerings. Each element in the list contains detailed information about one offering.