Network.AWS.ElastiCache.DescribeSnapshots

Description

The DescribeSnapshots action returns information about cache cluster snapshots. By default, DescribeSnapshots lists all of your snapshots; it can optionally describe a single snapshot, or just the snapshots associated with a particular cache cluster.

http://docs.aws.amazon.com/AmazonElastiCache/latest/APIReference/API_DescribeSnapshots.html

Request

data DescribeSnapshots

Request constructor

describeSnapshots :: DescribeSnapshots

DescribeSnapshots constructor.

The fields accessible through corresponding lenses are:

dsCacheClusterId :: Maybe Text

dsMarker :: Maybe Text

dsMaxRecords :: Maybe Int

dsSnapshotName :: Maybe Text

dsSnapshotSource :: Maybe Text

Request lenses

dsCacheClusterId :: Lens' DescribeSnapshots (Maybe Text)

A user-supplied cluster identifier. If this parameter is specified, only snapshots associated with that specific cache cluster will be described.

dsMarker :: Lens' DescribeSnapshots (Maybe Text)

An optional marker returned from a prior request. Use this marker for pagination of results from this action. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

dsMaxRecords :: Lens' DescribeSnapshots (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved.

Default: 50

Constraints: minimum 20; maximum 50.

dsSnapshotName :: Lens' DescribeSnapshots (Maybe Text)

A user-supplied name of the snapshot. If this parameter is specified, only this snapshot will be described.

dsSnapshotSource :: Lens' DescribeSnapshots (Maybe Text)

If set to system, the output shows snapshots that were automatically created by ElastiCache. If set to user the output shows snapshots that were manually created. If omitted, the output shows both automatically and manually created snapshots.

Response

data DescribeSnapshotsResponse

Response constructor

describeSnapshotsResponse :: DescribeSnapshotsResponse

DescribeSnapshotsResponse constructor.

The fields accessible through corresponding lenses are:

dsrMarker :: Maybe Text

dsrSnapshots :: [Snapshot]

Response lenses

dsrMarker :: Lens' DescribeSnapshotsResponse (Maybe Text)

An optional marker returned from a prior request. Use this marker for pagination of results from this action. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

dsrSnapshots :: Lens' DescribeSnapshotsResponse [Snapshot]

A list of snapshots. Each item in the list contains detailed information about one snapshot.