Network.AWS.ElastiCache.ListAllowedNodeTypeModifications

Description

Lists all available node types that you can scale your Redis cluster's or replication group's current node type up to.

When you use the ModifyCacheCluster or ModifyReplicationGroup operations to scale up your cluster or replication group, the value of the CacheNodeType parameter must be one of the node types returned by this operation.

Creating a Request

listAllowedNodeTypeModifications :: ListAllowedNodeTypeModifications

Creates a value of ListAllowedNodeTypeModifications with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lantmCacheClusterId - The name of the cache cluster you want to scale up to a larger node instanced type. ElastiCache uses the cluster id to identify the current node type of this cluster and from that to create a list of node types you can scale up to. Important: You must provide a value for either the CacheClusterId or the ReplicationGroupId .

lantmReplicationGroupId - The name of the replication group want to scale up to a larger node type. ElastiCache uses the replication group id to identify the current node type being used by this replication group, and from that to create a list of node types you can scale up to. Important: You must provide a value for either the CacheClusterId or the ReplicationGroupId .

data ListAllowedNodeTypeModifications

The input parameters for the ListAllowedNodeTypeModifications operation.

See: listAllowedNodeTypeModifications smart constructor.

Request Lenses

lantmCacheClusterId :: Lens' ListAllowedNodeTypeModifications (Maybe Text)

The name of the cache cluster you want to scale up to a larger node instanced type. ElastiCache uses the cluster id to identify the current node type of this cluster and from that to create a list of node types you can scale up to. Important: You must provide a value for either the CacheClusterId or the ReplicationGroupId .

lantmReplicationGroupId :: Lens' ListAllowedNodeTypeModifications (Maybe Text)

The name of the replication group want to scale up to a larger node type. ElastiCache uses the replication group id to identify the current node type being used by this replication group, and from that to create a list of node types you can scale up to. Important: You must provide a value for either the CacheClusterId or the ReplicationGroupId .

Destructuring the Response

listAllowedNodeTypeModificationsResponse

Arguments

:: Int
lantmrsResponseStatus

-> ListAllowedNodeTypeModificationsResponse

Creates a value of ListAllowedNodeTypeModificationsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lantmrsScaleUpModifications - A string list, each element of which specifies a cache node type which you can use to scale your cache cluster or replication group. When scaling up a Redis cluster or replication group using ModifyCacheCluster or ModifyReplicationGroup , use a value from this list for the CacheNodeType parameter.

lantmrsResponseStatus - -- | The response status code.

data ListAllowedNodeTypeModificationsResponse

Represents the allowed node types you can use to modify your cache cluster or replication group.

See: listAllowedNodeTypeModificationsResponse smart constructor.

Response Lenses

lantmrsScaleUpModifications :: Lens' ListAllowedNodeTypeModificationsResponse [Text]

A string list, each element of which specifies a cache node type which you can use to scale your cache cluster or replication group. When scaling up a Redis cluster or replication group using ModifyCacheCluster or ModifyReplicationGroup , use a value from this list for the CacheNodeType parameter.

lantmrsResponseStatus :: Lens' ListAllowedNodeTypeModificationsResponse Int

- | The response status code.