Network.AWS.ElastiCache.ListTagsForResource

Description

The ListTagsForResource action lists all cost allocation tags currently on the named resource. A cost allocation tag is a key-value pair where the key is case-sensitive and the value is optional. Cost allocation tags can be used to categorize and track your AWS costs.

You can have a maximum of 10 cost allocation tags on an ElastiCache resource. For more information, see Using Cost Allocation Tags in AmazonElastiCache.

http://docs.aws.amazon.com/AmazonElastiCache/latest/APIReference/API_ListTagsForResource.html

Request

data ListTagsForResource

Request constructor

listTagsForResource

Arguments

:: Text
ltfrResourceName

-> ListTagsForResource

ListTagsForResource constructor.

The fields accessible through corresponding lenses are:

ltfrResourceName :: Text

Request lenses

ltfrResourceName :: Lens' ListTagsForResource Text

The name of the resource for which you want the list of tags, for example 'arn:aws:elasticache:us-west-2:0123456789:cluster:myCluster'.

Response

data ListTagsForResourceResponse

Response constructor

listTagsForResourceResponse :: ListTagsForResourceResponse

ListTagsForResourceResponse constructor.

The fields accessible through corresponding lenses are:

ltfrrTagList :: [Tag]

Response lenses

ltfrrTagList :: Lens' ListTagsForResourceResponse [Tag]

A list of cost allocation tags as key-value pairs.