Network.AWS.ElastiCache.PurchaseReservedCacheNodesOffering

Description

Allows you to purchase a reserved cache node offering.

Creating a Request

purchaseReservedCacheNodesOffering

Arguments

:: Text
prcnoReservedCacheNodesOfferingId

-> PurchaseReservedCacheNodesOffering

Creates a value of PurchaseReservedCacheNodesOffering with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

prcnoCacheNodeCount - The number of cache node instances to reserve. Default: 1

prcnoReservedCacheNodeId - A customer-specified identifier to track this reservation. Example: myreservationID

prcnoReservedCacheNodesOfferingId - The ID of the reserved cache node offering to purchase. Example: 438012d3-4052-4cc7-b2e3-8d3372e0e706

data PurchaseReservedCacheNodesOffering

Represents the input of a PurchaseReservedCacheNodesOffering operation.

See: purchaseReservedCacheNodesOffering smart constructor.

Request Lenses

prcnoCacheNodeCount :: Lens' PurchaseReservedCacheNodesOffering (Maybe Int)

The number of cache node instances to reserve. Default: 1

prcnoReservedCacheNodeId :: Lens' PurchaseReservedCacheNodesOffering (Maybe Text)

A customer-specified identifier to track this reservation. Example: myreservationID

prcnoReservedCacheNodesOfferingId :: Lens' PurchaseReservedCacheNodesOffering Text

The ID of the reserved cache node offering to purchase. Example: 438012d3-4052-4cc7-b2e3-8d3372e0e706

Destructuring the Response

purchaseReservedCacheNodesOfferingResponse

Arguments

:: Int
prcnorsResponseStatus

-> PurchaseReservedCacheNodesOfferingResponse

Creates a value of PurchaseReservedCacheNodesOfferingResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

prcnorsReservedCacheNode - Undocumented member.

prcnorsResponseStatus - -- | The response status code.

data PurchaseReservedCacheNodesOfferingResponse

See: purchaseReservedCacheNodesOfferingResponse smart constructor.

Response Lenses

prcnorsReservedCacheNode :: Lens' PurchaseReservedCacheNodesOfferingResponse (Maybe ReservedCacheNode)

Undocumented member.

prcnorsResponseStatus :: Lens' PurchaseReservedCacheNodesOfferingResponse Int

- | The response status code.