Network.AWS.ElastiCache.RebootCacheCluster

Description

The RebootCacheCluster action reboots some, or all, of the cache nodes within a provisioned cache cluster. This API will apply any modified cache parameter groups to the cache cluster. The reboot action takes place as soon as possible, and results in a momentary outage to the cache cluster. During the reboot, the cache cluster status is set to REBOOTING.

The reboot causes the contents of the cache (for each cache node being rebooted) to be lost.

When the reboot is complete, a cache cluster event is created.

http://docs.aws.amazon.com/AmazonElastiCache/latest/APIReference/API_RebootCacheCluster.html

Request

data RebootCacheCluster

Request constructor

rebootCacheCluster

Arguments

:: Text
rccCacheClusterId

-> RebootCacheCluster

RebootCacheCluster constructor.

The fields accessible through corresponding lenses are:

rccCacheClusterId :: Text

rccCacheNodeIdsToReboot :: [Text]

Request lenses

rccCacheClusterId :: Lens' RebootCacheCluster Text

The cache cluster identifier. This parameter is stored as a lowercase string.

rccCacheNodeIdsToReboot :: Lens' RebootCacheCluster [Text]

A list of cache node IDs to reboot. A node ID is a numeric identifier (0001, 0002, etc.). To reboot an entire cache cluster, specify all of the cache node IDs.

Response

data RebootCacheClusterResponse

Response constructor

rebootCacheClusterResponse :: RebootCacheClusterResponse

RebootCacheClusterResponse constructor.

The fields accessible through corresponding lenses are:

rccrCacheCluster :: Maybe CacheCluster

Response lenses

rccrCacheCluster :: Lens' RebootCacheClusterResponse (Maybe CacheCluster)