Network.AWS.ElastiCache.RemoveTagsFromResource

Description

The RemoveTagsFromResource action removes the tags identified by the TagKeys list from the named resource.

http://docs.aws.amazon.com/AmazonElastiCache/latest/APIReference/API_RemoveTagsFromResource.html

Request

data RemoveTagsFromResource

Request constructor

removeTagsFromResource

Arguments

:: Text
rtfrResourceName

-> RemoveTagsFromResource

RemoveTagsFromResource constructor.

The fields accessible through corresponding lenses are:

rtfrResourceName :: Text

rtfrTagKeys :: [Text]

Request lenses

rtfrResourceName :: Lens' RemoveTagsFromResource Text

The name of the ElastiCache resource from which you want the listed tags removed, for example 'arn:aws:elasticache:us-west-2:0123456789:cluster:myCluster'.

rtfrTagKeys :: Lens' RemoveTagsFromResource [Text]

A list of TagKeys identifying the tags you want removed from the named resource. For example, 'TagKeys.member.1=Region' removes the cost allocation tag with the key name Region from the resource named by the ResourceName parameter.

Response

data RemoveTagsFromResourceResponse

Response constructor

removeTagsFromResourceResponse :: RemoveTagsFromResourceResponse

RemoveTagsFromResourceResponse constructor.

The fields accessible through corresponding lenses are:

rtfrrTagList :: [Tag]

Response lenses

rtfrrTagList :: Lens' RemoveTagsFromResourceResponse [Tag]

A list of cost allocation tags as key-value pairs.