Network.AWS.ElastiCache.ResetCacheParameterGroup

Description

The ResetCacheParameterGroup action modifies the parameters of a cache parameter group to the engine or system default value. You can reset specific parameters by submitting a list of parameter names. To reset the entire cache parameter group, specify the ResetAllParameters and CacheParameterGroupName parameters.

http://docs.aws.amazon.com/AmazonElastiCache/latest/APIReference/API_ResetCacheParameterGroup.html

Request

data ResetCacheParameterGroup

Request constructor

resetCacheParameterGroup

Arguments

:: Text
rcpgCacheParameterGroupName

-> ResetCacheParameterGroup

ResetCacheParameterGroup constructor.

The fields accessible through corresponding lenses are:

rcpgCacheParameterGroupName :: Text

rcpgParameterNameValues :: [ParameterNameValue]

rcpgResetAllParameters :: Maybe Bool

Request lenses

rcpgCacheParameterGroupName :: Lens' ResetCacheParameterGroup Text

The name of the cache parameter group to reset.

rcpgParameterNameValues :: Lens' ResetCacheParameterGroup [ParameterNameValue]

An array of parameter names to be reset. If you are not resetting the entire cache parameter group, you must specify at least one parameter name.

rcpgResetAllParameters :: Lens' ResetCacheParameterGroup (Maybe Bool)

If true, all parameters in the cache parameter group will be reset to default values. If false, no such action occurs.

Valid values: true | false

Response

data ResetCacheParameterGroupResponse

Response constructor

resetCacheParameterGroupResponse :: ResetCacheParameterGroupResponse

ResetCacheParameterGroupResponse constructor.

The fields accessible through corresponding lenses are:

rcpgrCacheParameterGroupName :: Maybe Text

Response lenses

rcpgrCacheParameterGroupName :: Lens' ResetCacheParameterGroupResponse (Maybe Text)

The name of the cache parameter group.