Network.AWS.ElastiCache.RevokeCacheSecurityGroupIngress

Description

The RevokeCacheSecurityGroupIngress operation revokes ingress from a cache security group. Use this operation to disallow access from an Amazon EC2 security group that had been previously authorized. https:/elasticache.us-east-1.amazonaws.com ?Action=RevokeCacheSecurityGroupIngress &EC2SecurityGroupName=default &CacheSecurityGroupName=mygroup &EC2SecurityGroupOwnerId=1234-5678-1234 &Version=2014-03-24 &SignatureVersion=4 &SignatureMethod=HmacSHA256 &Timestamp=20140401T192317Z &X-Amz-Credential= revoking default 123456781234 mygroup 123456789012 My security group 02ae3699-3650-11e0-a564-8f11342c56b0.

Request

data RevokeCacheSecurityGroupIngress

Represents the input of a RevokeCacheSecurityGroupIngress operation.

Request constructor

revokeCacheSecurityGroupIngress

Arguments

:: Text
rcsgiCacheSecurityGroupName

-> Text
rcsgiEC2SecurityGroupName

-> Text
rcsgiEC2SecurityGroupOwnerId

-> RevokeCacheSecurityGroupIngress

Smart constructor for the minimum required parameters to construct a valid RevokeCacheSecurityGroupIngress request.

The fields accessible through corresponding lenses are:

CacheSecurityGroupName :: Text

EC2SecurityGroupName :: Text

EC2SecurityGroupOwnerId :: Text

Request lenses

rcsgiCacheSecurityGroupName :: Lens' RevokeCacheSecurityGroupIngress Text

The name of the cache security group to revoke ingress from.

rcsgiEC2SecurityGroupName :: Lens' RevokeCacheSecurityGroupIngress Text

The name of the Amazon EC2 security group to revoke access from.

rcsgiEC2SecurityGroupOwnerId :: Lens' RevokeCacheSecurityGroupIngress Text

The AWS account number of the Amazon EC2 security group owner. Note that this is not the same thing as an AWS access key ID - you must provide a valid AWS account number for this parameter.

Response

data RevokeCacheSecurityGroupIngressResponse

Response constructor

revokeCacheSecurityGroupIngressResponse :: RevokeCacheSecurityGroupIngressResponse

Smart constructor for the minimum required parameters to construct a valid RevokeCacheSecurityGroupIngressResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

CacheSecurityGroup :: Maybe CacheSecurityGroup

Response lenses

rcsgirCacheSecurityGroup :: Lens' RevokeCacheSecurityGroupIngressResponse (Maybe CacheSecurityGroup)

Represents the output of one of the following operations: AuthorizeCacheSecurityGroupIngress CreateCacheSecurityGroup RevokeCacheSecurityGroupIngress.