Network.AWS.ElastiCache.TestFailover

Description

Represents the input of a TestFailover operation which test automatic failover on a specified node group (called shard in the console) in a replication group (called cluster in the console).

Note the following

A customer can use this operation to test automatic failover on up to 5 shards (called node groups in the ElastiCache API and AWS CLI) in any rolling 24-hour period.

If calling this operation on shards in different clusters (called replication groups in the API and CLI), the calls can be made concurrently.

If calling this operation multiple times on different shards in the same Redis (cluster mode enabled) replication group, the first node replacement must complete before a subsequent call can be made.

To determine whether the node replacement is complete you can check Events using the Amazon ElastiCache console, the AWS CLI, or the ElastiCache API. Look for the following automatic failover related events, listed here in order of occurrance:

Replication group message: Test Failover API called for node group node-group-id

Cache cluster message: Failover from master node primary-node-id to replica node node-id completed

Replication group message: Failover from master node primary-node-id to replica node node-id completed

Cache cluster message: Recovering cache nodes node-id

Cache cluster message: Finished recovery for cache nodes node-id

For more information see:

Viewing ElastiCache Events in the ElastiCache User Guide

DescribeEvents in the ElastiCache API Reference

Also see, Testing Multi-AZ with Automatic Failover in the ElastiCache User Guide .

Creating a Request

testFailover

Arguments

:: Text
tfReplicationGroupId

-> Text
tfNodeGroupId

-> TestFailover

Creates a value of TestFailover with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tfReplicationGroupId - The name of the replication group (console: cluster) whose automatic failover is being tested by this operation.

tfNodeGroupId - The name of the node group (called shard in the console) in this replication group on which automatic failover is to be tested. You may test automatic failover on up to 5 node groups in any rolling 24-hour period.

data TestFailover

See: testFailover smart constructor.

Request Lenses

tfReplicationGroupId :: Lens' TestFailover Text

The name of the replication group (console: cluster) whose automatic failover is being tested by this operation.

tfNodeGroupId :: Lens' TestFailover Text

The name of the node group (called shard in the console) in this replication group on which automatic failover is to be tested. You may test automatic failover on up to 5 node groups in any rolling 24-hour period.

Destructuring the Response

testFailoverResponse

Arguments

:: Int
tfrsResponseStatus

-> TestFailoverResponse

Creates a value of TestFailoverResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tfrsReplicationGroup - Undocumented member.

tfrsResponseStatus - -- | The response status code.

data TestFailoverResponse

See: testFailoverResponse smart constructor.

Response Lenses

tfrsReplicationGroup :: Lens' TestFailoverResponse (Maybe ReplicationGroup)

Undocumented member.

tfrsResponseStatus :: Lens' TestFailoverResponse Int

- | The response status code.