Network.AWS.ElastiCache.Types

Service

data ElastiCache

Version 2015-02-02 of the Amazon ElastiCache service.

Error

data RESTError :: *

XML

ns :: Text

NodeSnapshot

data NodeSnapshot

nodeSnapshot :: NodeSnapshot

NodeSnapshot constructor.

The fields accessible through corresponding lenses are:

nsCacheNodeCreateTime :: Maybe UTCTime

nsCacheNodeId :: Maybe Text

nsCacheSize :: Maybe Text

nsSnapshotCreateTime :: Maybe UTCTime

nsCacheNodeCreateTime :: Lens' NodeSnapshot (Maybe UTCTime)

The date and time when the cache node was created in the source cache cluster.

nsCacheNodeId :: Lens' NodeSnapshot (Maybe Text)

The cache node identifier for the node in the source cache cluster.

nsCacheSize :: Lens' NodeSnapshot (Maybe Text)

The size of the cache on the source cache node.

nsSnapshotCreateTime :: Lens' NodeSnapshot (Maybe UTCTime)

The date and time when the source node's metadata and cache data set was obtained for the snapshot.

Snapshot

data Snapshot

snapshot :: Snapshot

Snapshot constructor.

The fields accessible through corresponding lenses are:

sAutoMinorVersionUpgrade :: Maybe Bool

sCacheClusterCreateTime :: Maybe UTCTime

sCacheClusterId :: Maybe Text

sCacheNodeType :: Maybe Text

sCacheParameterGroupName :: Maybe Text

sCacheSubnetGroupName :: Maybe Text

sEngine :: Maybe Text

sEngineVersion :: Maybe Text

sNodeSnapshots :: [NodeSnapshot]

sNumCacheNodes :: Maybe Int

sPort :: Maybe Int

sPreferredAvailabilityZone :: Maybe Text

sPreferredMaintenanceWindow :: Maybe Text

sSnapshotName :: Maybe Text

sSnapshotRetentionLimit :: Maybe Int

sSnapshotSource :: Maybe Text

sSnapshotStatus :: Maybe Text

sSnapshotWindow :: Maybe Text

sTopicArn :: Maybe Text

sVpcId :: Maybe Text

sAutoMinorVersionUpgrade :: Lens' Snapshot (Maybe Bool)

This parameter is currently disabled.

sCacheClusterCreateTime :: Lens' Snapshot (Maybe UTCTime)

The date and time when the source cache cluster was created.

sCacheClusterId :: Lens' Snapshot (Maybe Text)

The user-supplied identifier of the source cache cluster.

sCacheNodeType :: Lens' Snapshot (Maybe Text)

The name of the compute and memory capacity node type for the source cache cluster.

Valid node types are as follows:

General purpose: Current generation: 'cache.t2.micro', 'cache.t2.small', 'cache.t2.medium', 'cache.m3.medium', 'cache.m3.large', 'cache.m3.xlarge', 'cache.m3.2xlarge' Previous generation: 'cache.t1.micro', 'cache.m1.small', 'cache.m1.medium', 'cache.m1.large', 'cache.m1.xlarge' Compute optimized: 'cache.c1.xlarge' Memory optimized Current generation: 'cache.r3.large', 'cache.r3.xlarge', 'cache.r3.2xlarge', 'cache.r3.4xlarge', 'cache.r3.8xlarge' Previous generation: 'cache.m2.xlarge', 'cache.m2.2xlarge', 'cache.m2.4xlarge' Notes:

All t2 instances are created in an Amazon Virtual Private Cloud (VPC). Redis backup/restore is not supported for t2 instances. Redis Append-only files (AOF) functionality is not supported for t1 or t2 instances. For a complete listing of cache node types and specifications, see Amazon ElastiCache Product Features and Details and Cache NodeType-Specific Parameters for Memcached or Cache Node Type-Specific Parametersfor Redis.

sCacheParameterGroupName :: Lens' Snapshot (Maybe Text)

The cache parameter group that is associated with the source cache cluster.

sCacheSubnetGroupName :: Lens' Snapshot (Maybe Text)

The name of the cache subnet group associated with the source cache cluster.

sEngine :: Lens' Snapshot (Maybe Text)

The name of the cache engine (memcached or redis) used by the source cache cluster.

sEngineVersion :: Lens' Snapshot (Maybe Text)

The version of the cache engine version that is used by the source cache cluster.

sNodeSnapshots :: Lens' Snapshot [NodeSnapshot]

A list of the cache nodes in the source cache cluster.

sNumCacheNodes :: Lens' Snapshot (Maybe Int)

The number of cache nodes in the source cache cluster.

For clusters running Redis, this value must be 1. For clusters running Memcached, this value must be between 1 and 20.

sPort :: Lens' Snapshot (Maybe Int)

The port number used by each cache nodes in the source cache cluster.

sPreferredAvailabilityZone :: Lens' Snapshot (Maybe Text)

The name of the Availability Zone in which the source cache cluster is located.

sPreferredMaintenanceWindow :: Lens' Snapshot (Maybe Text)

Specifies the weekly time range during which maintenance on the cache cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period. Valid values for ddd are:

sun mon tue wed thu fri sat Example: 'sun:05:00-sun:09:00'

sSnapshotName :: Lens' Snapshot (Maybe Text)

The name of a snapshot. For an automatic snapshot, the name is system-generated; for a manual snapshot, this is the user-provided name.

sSnapshotRetentionLimit :: Lens' Snapshot (Maybe Int)

For an automatic snapshot, the number of days for which ElastiCache will retain the snapshot before deleting it.

For manual snapshots, this field reflects the SnapshotRetentionLimit for the source cache cluster when the snapshot was created. This field is otherwise ignored: Manual snapshots do not expire, and can only be deleted using the DeleteSnapshot action.

Important If the value of SnapshotRetentionLimit is set to zero (0), backups are turned off.

sSnapshotSource :: Lens' Snapshot (Maybe Text)

Indicates whether the snapshot is from an automatic backup (automated) or was created manually (manual).

sSnapshotStatus :: Lens' Snapshot (Maybe Text)

The status of the snapshot. Valid values: creating | available | restoring | copying | deleting.

sSnapshotWindow :: Lens' Snapshot (Maybe Text)

The daily time range during which ElastiCache takes daily snapshots of the source cache cluster.

sTopicArn :: Lens' Snapshot (Maybe Text)

The Amazon Resource Name (ARN) for the topic used by the source cache cluster for publishing notifications.

sVpcId :: Lens' Snapshot (Maybe Text)

The Amazon Virtual Private Cloud identifier (VPC ID) of the cache subnet group for the source cache cluster.

Event

data Event

event :: Event

Event constructor.

The fields accessible through corresponding lenses are:

eDate :: Maybe UTCTime

eMessage :: Maybe Text

eSourceIdentifier :: Maybe Text

eSourceType :: Maybe SourceType

eDate :: Lens' Event (Maybe UTCTime)

The date and time when the event occurred.

eMessage :: Lens' Event (Maybe Text)

The text of the event.

eSourceIdentifier :: Lens' Event (Maybe Text)

The identifier for the source of the event. For example, if the event occurred at the cache cluster level, the identifier would be the name of the cache cluster.

eSourceType :: Lens' Event (Maybe SourceType)

Specifies the origin of this event - a cache cluster, a parameter group, a security group, etc.

NodeGroup

data NodeGroup

nodeGroup :: NodeGroup

NodeGroup constructor.

The fields accessible through corresponding lenses are:

ngNodeGroupId :: Maybe Text

ngNodeGroupMembers :: [NodeGroupMember]

ngPrimaryEndpoint :: Maybe Endpoint

ngStatus :: Maybe Text

ngNodeGroupId :: Lens' NodeGroup (Maybe Text)

The identifier for the node group. A replication group contains only one node group; therefore, the node group ID is 0001.

ngNodeGroupMembers :: Lens' NodeGroup [NodeGroupMember]

A list containing information about individual nodes within the node group.

ngPrimaryEndpoint :: Lens' NodeGroup (Maybe Endpoint)

ngStatus :: Lens' NodeGroup (Maybe Text)

The current state of this replication group - creating, available, etc.

CacheNodeTypeSpecificValue

data CacheNodeTypeSpecificValue

cacheNodeTypeSpecificValue :: CacheNodeTypeSpecificValue

CacheNodeTypeSpecificValue constructor.

The fields accessible through corresponding lenses are:

cntsvCacheNodeType :: Maybe Text

cntsvValue :: Maybe Text

cntsvCacheNodeType :: Lens' CacheNodeTypeSpecificValue (Maybe Text)

The cache node type for which this value applies.

cntsvValue :: Lens' CacheNodeTypeSpecificValue (Maybe Text)

The value for the cache node type.

Tag

data Tag

tag :: Tag

Tag constructor.

The fields accessible through corresponding lenses are:

tagKey :: Maybe Text

tagValue :: Maybe Text

tagKey :: Lens' Tag (Maybe Text)

The key for the tag.

tagValue :: Lens' Tag (Maybe Text)

The tag's value. May not be null.

PendingAutomaticFailoverStatus

data PendingAutomaticFailoverStatus

Constructors

Disabled
disabled

Enabled
enabled

NotificationConfiguration

data NotificationConfiguration

notificationConfiguration :: NotificationConfiguration

NotificationConfiguration constructor.

The fields accessible through corresponding lenses are:

ncTopicArn :: Maybe Text

ncTopicStatus :: Maybe Text

ncTopicArn :: Lens' NotificationConfiguration (Maybe Text)

The Amazon Resource Name (ARN) that identifies the topic.

ncTopicStatus :: Lens' NotificationConfiguration (Maybe Text)

The current state of the topic.

ReplicationGroupPendingModifiedValues

data ReplicationGroupPendingModifiedValues

replicationGroupPendingModifiedValues :: ReplicationGroupPendingModifiedValues

ReplicationGroupPendingModifiedValues constructor.

The fields accessible through corresponding lenses are:

rgpmvAutomaticFailoverStatus :: Maybe PendingAutomaticFailoverStatus

rgpmvPrimaryClusterId :: Maybe Text

rgpmvAutomaticFailoverStatus :: Lens' ReplicationGroupPendingModifiedValues (Maybe PendingAutomaticFailoverStatus)

Indicates the status of Multi-AZ for this replication group.

ElastiCache Multi-AZ replication groups are not supported on:

Redis versions earlier than 2.8.6. T1 and T2 cache node types.

rgpmvPrimaryClusterId :: Lens' ReplicationGroupPendingModifiedValues (Maybe Text)

The primary cluster ID which will be applied immediately (if '--apply-immediately' was specified), or during the next maintenance window.

EC2SecurityGroup

data EC2SecurityGroup

ec2SecurityGroup :: EC2SecurityGroup

EC2SecurityGroup constructor.

The fields accessible through corresponding lenses are:

ecsgEC2SecurityGroupName :: Maybe Text

ecsgEC2SecurityGroupOwnerId :: Maybe Text

ecsgStatus :: Maybe Text

ecsgEC2SecurityGroupName :: Lens' EC2SecurityGroup (Maybe Text)

The name of the Amazon EC2 security group.

ecsgEC2SecurityGroupOwnerId :: Lens' EC2SecurityGroup (Maybe Text)

The AWS account ID of the Amazon EC2 security group owner.

ecsgStatus :: Lens' EC2SecurityGroup (Maybe Text)

The status of the Amazon EC2 security group.

ParameterNameValue

data ParameterNameValue

parameterNameValue :: ParameterNameValue

ParameterNameValue constructor.

The fields accessible through corresponding lenses are:

pnvParameterName :: Maybe Text

pnvParameterValue :: Maybe Text

pnvParameterName :: Lens' ParameterNameValue (Maybe Text)

The name of the parameter.

pnvParameterValue :: Lens' ParameterNameValue (Maybe Text)

The value of the parameter.

SourceType

data SourceType

Constructors

STCacheCluster
cache-cluster

STCacheParameterGroup
cache-parameter-group

STCacheSecurityGroup
cache-security-group

STCacheSubnetGroup
cache-subnet-group

CacheSubnetGroup

data CacheSubnetGroup

cacheSubnetGroup :: CacheSubnetGroup

CacheSubnetGroup constructor.

The fields accessible through corresponding lenses are:

csgCacheSubnetGroupDescription :: Maybe Text

csgCacheSubnetGroupName :: Maybe Text

csgSubnets :: [Subnet]

csgVpcId :: Maybe Text

csgCacheSubnetGroupDescription :: Lens' CacheSubnetGroup (Maybe Text)

The description of the cache subnet group.

csgCacheSubnetGroupName :: Lens' CacheSubnetGroup (Maybe Text)

The name of the cache subnet group.

csgSubnets :: Lens' CacheSubnetGroup [Subnet]

A list of subnets associated with the cache subnet group.

csgVpcId :: Lens' CacheSubnetGroup (Maybe Text)

The Amazon Virtual Private Cloud identifier (VPC ID) of the cache subnet group.

ReservedCacheNode

data ReservedCacheNode

reservedCacheNode :: ReservedCacheNode

ReservedCacheNode constructor.

The fields accessible through corresponding lenses are:

rcnCacheNodeCount :: Maybe Int

rcnCacheNodeType :: Maybe Text

rcnDuration :: Maybe Int

rcnFixedPrice :: Maybe Double

rcnOfferingType :: Maybe Text

rcnProductDescription :: Maybe Text

rcnRecurringCharges :: [RecurringCharge]

rcnReservedCacheNodeId :: Maybe Text

rcnReservedCacheNodesOfferingId :: Maybe Text

rcnStartTime :: Maybe UTCTime

rcnState :: Maybe Text

rcnUsagePrice :: Maybe Double

rcnCacheNodeCount :: Lens' ReservedCacheNode (Maybe Int)

The number of cache nodes that have been reserved.

rcnCacheNodeType :: Lens' ReservedCacheNode (Maybe Text)

The cache node type for the reserved cache nodes.

Valid node types are as follows:

General purpose: Current generation: 'cache.t2.micro', 'cache.t2.small', 'cache.t2.medium', 'cache.m3.medium', 'cache.m3.large', 'cache.m3.xlarge', 'cache.m3.2xlarge' Previous generation: 'cache.t1.micro', 'cache.m1.small', 'cache.m1.medium', 'cache.m1.large', 'cache.m1.xlarge' Compute optimized: 'cache.c1.xlarge' Memory optimized Current generation: 'cache.r3.large', 'cache.r3.xlarge', 'cache.r3.2xlarge', 'cache.r3.4xlarge', 'cache.r3.8xlarge' Previous generation: 'cache.m2.xlarge', 'cache.m2.2xlarge', 'cache.m2.4xlarge' Notes:

All t2 instances are created in an Amazon Virtual Private Cloud (VPC). Redis backup/restore is not supported for t2 instances. Redis Append-only files (AOF) functionality is not supported for t1 or t2 instances. For a complete listing of cache node types and specifications, see Amazon ElastiCache Product Features and Details and Cache NodeType-Specific Parameters for Memcached or Cache Node Type-Specific Parametersfor Redis.

rcnDuration :: Lens' ReservedCacheNode (Maybe Int)

The duration of the reservation in seconds.

rcnFixedPrice :: Lens' ReservedCacheNode (Maybe Double)

The fixed price charged for this reserved cache node.

rcnOfferingType :: Lens' ReservedCacheNode (Maybe Text)

The offering type of this reserved cache node.

rcnProductDescription :: Lens' ReservedCacheNode (Maybe Text)

The description of the reserved cache node.

rcnRecurringCharges :: Lens' ReservedCacheNode [RecurringCharge]

The recurring price charged to run this reserved cache node.

rcnReservedCacheNodeId :: Lens' ReservedCacheNode (Maybe Text)

The unique identifier for the reservation.

rcnReservedCacheNodesOfferingId :: Lens' ReservedCacheNode (Maybe Text)

The offering identifier.

rcnStartTime :: Lens' ReservedCacheNode (Maybe UTCTime)

The time the reservation started.

rcnState :: Lens' ReservedCacheNode (Maybe Text)

The state of the reserved cache node.

rcnUsagePrice :: Lens' ReservedCacheNode (Maybe Double)

The hourly price charged for this reserved cache node.

Subnet

data Subnet

subnet :: Subnet

Subnet constructor.

The fields accessible through corresponding lenses are:

sSubnetAvailabilityZone :: Maybe AvailabilityZone

sSubnetIdentifier :: Maybe Text

sSubnetAvailabilityZone :: Lens' Subnet (Maybe AvailabilityZone)

The Availability Zone associated with the subnet.

sSubnetIdentifier :: Lens' Subnet (Maybe Text)

The unique identifier for the subnet.

SecurityGroupMembership

data SecurityGroupMembership

securityGroupMembership :: SecurityGroupMembership

SecurityGroupMembership constructor.

The fields accessible through corresponding lenses are:

sgmSecurityGroupId :: Maybe Text

sgmStatus :: Maybe Text

sgmSecurityGroupId :: Lens' SecurityGroupMembership (Maybe Text)

The identifier of the cache security group.

sgmStatus :: Lens' SecurityGroupMembership (Maybe Text)

The status of the cache security group membership. The status changes whenever a cache security group is modified, or when the cache security groups assigned to a cache cluster are modified.

CacheCluster

data CacheCluster

cacheCluster :: CacheCluster

CacheCluster constructor.

The fields accessible through corresponding lenses are:

ccAutoMinorVersionUpgrade :: Maybe Bool

ccCacheClusterCreateTime :: Maybe UTCTime

ccCacheClusterId :: Maybe Text

ccCacheClusterStatus :: Maybe Text

ccCacheNodeType :: Maybe Text

ccCacheNodes :: [CacheNode]

ccCacheParameterGroup :: Maybe CacheParameterGroupStatus

ccCacheSecurityGroups :: [CacheSecurityGroupMembership]

ccCacheSubnetGroupName :: Maybe Text

ccClientDownloadLandingPage :: Maybe Text

ccConfigurationEndpoint :: Maybe Endpoint

ccEngine :: Maybe Text

ccEngineVersion :: Maybe Text

ccNotificationConfiguration :: Maybe NotificationConfiguration

ccNumCacheNodes :: Maybe Int

ccPendingModifiedValues :: Maybe PendingModifiedValues

ccPreferredAvailabilityZone :: Maybe Text

ccPreferredMaintenanceWindow :: Maybe Text

ccReplicationGroupId :: Maybe Text

ccSecurityGroups :: [SecurityGroupMembership]

ccSnapshotRetentionLimit :: Maybe Int

ccSnapshotWindow :: Maybe Text

ccAutoMinorVersionUpgrade :: Lens' CacheCluster (Maybe Bool)

This parameter is currently disabled.

ccCacheClusterCreateTime :: Lens' CacheCluster (Maybe UTCTime)

The date and time when the cache cluster was created.

ccCacheClusterId :: Lens' CacheCluster (Maybe Text)

The user-supplied identifier of the cache cluster. This identifier is a unique key that identifies a cache cluster.

ccCacheClusterStatus :: Lens' CacheCluster (Maybe Text)

The current state of this cache cluster, one of the following values: available, creating, deleted, deleting, incompatible-network, modifying, rebootingcache cluster nodes, restore-failed, or snapshotting.

ccCacheNodeType :: Lens' CacheCluster (Maybe Text)

The name of the compute and memory capacity node type for the cache cluster.

Valid node types are as follows:

General purpose: Current generation: 'cache.t2.micro', 'cache.t2.small', 'cache.t2.medium', 'cache.m3.medium', 'cache.m3.large', 'cache.m3.xlarge', 'cache.m3.2xlarge' Previous generation: 'cache.t1.micro', 'cache.m1.small', 'cache.m1.medium', 'cache.m1.large', 'cache.m1.xlarge' Compute optimized: 'cache.c1.xlarge' Memory optimized Current generation: 'cache.r3.large', 'cache.r3.xlarge', 'cache.r3.2xlarge', 'cache.r3.4xlarge', 'cache.r3.8xlarge' Previous generation: 'cache.m2.xlarge', 'cache.m2.2xlarge', 'cache.m2.4xlarge' Notes:

All t2 instances are created in an Amazon Virtual Private Cloud (VPC). Redis backup/restore is not supported for t2 instances. Redis Append-only files (AOF) functionality is not supported for t1 or t2 instances. For a complete listing of cache node types and specifications, see Amazon ElastiCache Product Features and Details and Cache NodeType-Specific Parameters for Memcached or Cache Node Type-Specific Parametersfor Redis.

ccCacheNodes :: Lens' CacheCluster [CacheNode]

A list of cache nodes that are members of the cache cluster.

ccCacheParameterGroup :: Lens' CacheCluster (Maybe CacheParameterGroupStatus)

ccCacheSecurityGroups :: Lens' CacheCluster [CacheSecurityGroupMembership]

A list of cache security group elements, composed of name and status sub-elements.

ccCacheSubnetGroupName :: Lens' CacheCluster (Maybe Text)

The name of the cache subnet group associated with the cache cluster.

ccClientDownloadLandingPage :: Lens' CacheCluster (Maybe Text)

The URL of the web page where you can download the latest ElastiCache client library.

ccConfigurationEndpoint :: Lens' CacheCluster (Maybe Endpoint)

ccEngine :: Lens' CacheCluster (Maybe Text)

The name of the cache engine (memcached or redis) to be used for this cache cluster.

ccEngineVersion :: Lens' CacheCluster (Maybe Text)

The version of the cache engine version that is used in this cache cluster.

ccNotificationConfiguration :: Lens' CacheCluster (Maybe NotificationConfiguration)

ccNumCacheNodes :: Lens' CacheCluster (Maybe Int)

The number of cache nodes in the cache cluster.

For clusters running Redis, this value must be 1. For clusters running Memcached, this value must be between 1 and 20.

ccPendingModifiedValues :: Lens' CacheCluster (Maybe PendingModifiedValues)

ccPreferredAvailabilityZone :: Lens' CacheCluster (Maybe Text)

The name of the Availability Zone in which the cache cluster is located or Multiple if the cache nodes are located in different Availability Zones.

ccPreferredMaintenanceWindow :: Lens' CacheCluster (Maybe Text)

Specifies the weekly time range during which maintenance on the cache cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period. Valid values for ddd are:

sun mon tue wed thu fri sat Example: 'sun:05:00-sun:09:00'

ccReplicationGroupId :: Lens' CacheCluster (Maybe Text)

The replication group to which this cache cluster belongs. If this field is empty, the cache cluster is not associated with any replication group.

ccSecurityGroups :: Lens' CacheCluster [SecurityGroupMembership]

A list of VPC Security Groups associated with the cache cluster.

ccSnapshotRetentionLimit :: Lens' CacheCluster (Maybe Int)

The number of days for which ElastiCache will retain automatic cache cluster snapshots before deleting them. For example, if you set SnapshotRetentionLimit to 5, then a snapshot that was taken today will be retained for 5 days before being deleted.

Important If the value of SnapshotRetentionLimit is set to zero (0), backups are turned off.

ccSnapshotWindow :: Lens' CacheCluster (Maybe Text)

The daily time range (in UTC) during which ElastiCache will begin taking a daily snapshot of your cache cluster.

Example: '05:00-09:00'

EngineDefaults

data EngineDefaults

engineDefaults :: EngineDefaults

EngineDefaults constructor.

The fields accessible through corresponding lenses are:

edCacheNodeTypeSpecificParameters :: [CacheNodeTypeSpecificParameter]

edCacheParameterGroupFamily :: Maybe Text

edMarker :: Maybe Text

edParameters :: [Parameter]

edCacheNodeTypeSpecificParameters :: Lens' EngineDefaults [CacheNodeTypeSpecificParameter]

A list of parameters specific to a particular cache node type. Each element in the list contains detailed information about one parameter.

edCacheParameterGroupFamily :: Lens' EngineDefaults (Maybe Text)

Specifies the name of the cache parameter group family to which the engine default parameters apply.

edMarker :: Lens' EngineDefaults (Maybe Text)

Provides an identifier to allow retrieval of paginated results.

edParameters :: Lens' EngineDefaults [Parameter]

Contains a list of engine default parameters.

CacheParameterGroupStatus

data CacheParameterGroupStatus

cacheParameterGroupStatus :: CacheParameterGroupStatus

CacheParameterGroupStatus constructor.

The fields accessible through corresponding lenses are:

cpgsCacheNodeIdsToReboot :: [Text]

cpgsCacheParameterGroupName :: Maybe Text

cpgsParameterApplyStatus :: Maybe Text

cpgsCacheNodeIdsToReboot :: Lens' CacheParameterGroupStatus [Text]

A list of the cache node IDs which need to be rebooted for parameter changes to be applied. A node ID is a numeric identifier (0001, 0002, etc.).

cpgsCacheParameterGroupName :: Lens' CacheParameterGroupStatus (Maybe Text)

The name of the cache parameter group.

cpgsParameterApplyStatus :: Lens' CacheParameterGroupStatus (Maybe Text)

The status of parameter updates.

CacheNode

data CacheNode

cacheNode :: CacheNode

CacheNode constructor.

The fields accessible through corresponding lenses are:

cnCacheNodeCreateTime :: Maybe UTCTime

cnCacheNodeId :: Maybe Text

cnCacheNodeStatus :: Maybe Text

cnCustomerAvailabilityZone :: Maybe Text

cnEndpoint :: Maybe Endpoint

cnParameterGroupStatus :: Maybe Text

cnSourceCacheNodeId :: Maybe Text

cnCacheNodeCreateTime :: Lens' CacheNode (Maybe UTCTime)

The date and time when the cache node was created.

cnCacheNodeId :: Lens' CacheNode (Maybe Text)

The cache node identifier. A node ID is a numeric identifier (0001, 0002, etc.). The combination of cluster ID and node ID uniquely identifies every cache node used in a customer's AWS account.

cnCacheNodeStatus :: Lens' CacheNode (Maybe Text)

The current state of this cache node.

cnCustomerAvailabilityZone :: Lens' CacheNode (Maybe Text)

The Availability Zone where this node was created and now resides.

cnEndpoint :: Lens' CacheNode (Maybe Endpoint)

The hostname for connecting to this cache node.

cnParameterGroupStatus :: Lens' CacheNode (Maybe Text)

The status of the parameter group applied to this cache node.

cnSourceCacheNodeId :: Lens' CacheNode (Maybe Text)

The ID of the primary node to which this read replica node is synchronized. If this field is empty, then this node is not associated with a primary cache cluster.

CacheSecurityGroupMembership

data CacheSecurityGroupMembership

cacheSecurityGroupMembership :: CacheSecurityGroupMembership

CacheSecurityGroupMembership constructor.

The fields accessible through corresponding lenses are:

csgmCacheSecurityGroupName :: Maybe Text

csgmStatus :: Maybe Text

csgmCacheSecurityGroupName :: Lens' CacheSecurityGroupMembership (Maybe Text)

The name of the cache security group.

csgmStatus :: Lens' CacheSecurityGroupMembership (Maybe Text)

The membership status in the cache security group. The status changes when a cache security group is modified, or when the cache security groups assigned to a cache cluster are modified.

AvailabilityZone

data AvailabilityZone

availabilityZone :: AvailabilityZone

AvailabilityZone constructor.

The fields accessible through corresponding lenses are:

azName :: Maybe Text

azName :: Lens' AvailabilityZone (Maybe Text)

The name of the Availability Zone.

NodeGroupMember

data NodeGroupMember

nodeGroupMember :: NodeGroupMember

NodeGroupMember constructor.

The fields accessible through corresponding lenses are:

ngmCacheClusterId :: Maybe Text

ngmCacheNodeId :: Maybe Text

ngmCurrentRole :: Maybe Text

ngmPreferredAvailabilityZone :: Maybe Text

ngmReadEndpoint :: Maybe Endpoint

ngmCacheClusterId :: Lens' NodeGroupMember (Maybe Text)

The ID of the cache cluster to which the node belongs.

ngmCacheNodeId :: Lens' NodeGroupMember (Maybe Text)

The ID of the node within its cache cluster. A node ID is a numeric identifier (0001, 0002, etc.).

ngmCurrentRole :: Lens' NodeGroupMember (Maybe Text)

The role that is currently assigned to the node - primary or replica.

ngmPreferredAvailabilityZone :: Lens' NodeGroupMember (Maybe Text)

The name of the Availability Zone in which the node is located.

ngmReadEndpoint :: Lens' NodeGroupMember (Maybe Endpoint)

CacheParameterGroup

data CacheParameterGroup

cacheParameterGroup :: CacheParameterGroup

CacheParameterGroup constructor.

The fields accessible through corresponding lenses are:

cpgCacheParameterGroupFamily :: Maybe Text

cpgCacheParameterGroupName :: Maybe Text

cpgDescription :: Maybe Text

cpgCacheParameterGroupFamily :: Lens' CacheParameterGroup (Maybe Text)

The name of the cache parameter group family that this cache parameter group is compatible with.

cpgCacheParameterGroupName :: Lens' CacheParameterGroup (Maybe Text)

The name of the cache parameter group.

cpgDescription :: Lens' CacheParameterGroup (Maybe Text)

The description for this cache parameter group.

AutomaticFailoverStatus

data AutomaticFailoverStatus

Constructors

AFSDisabled
disabled

AFSDisabling
disabling

AFSEnabled
enabled

AFSEnabling
enabling

CacheSecurityGroup

data CacheSecurityGroup

cacheSecurityGroup :: CacheSecurityGroup

CacheSecurityGroup constructor.

The fields accessible through corresponding lenses are:

csgCacheSecurityGroupName :: Maybe Text

csgDescription :: Maybe Text

csgEC2SecurityGroups :: [EC2SecurityGroup]

csgOwnerId :: Maybe Text

csgCacheSecurityGroupName :: Lens' CacheSecurityGroup (Maybe Text)

The name of the cache security group.

csgDescription :: Lens' CacheSecurityGroup (Maybe Text)

The description of the cache security group.

csgEC2SecurityGroups :: Lens' CacheSecurityGroup [EC2SecurityGroup]

A list of Amazon EC2 security groups that are associated with this cache security group.

csgOwnerId :: Lens' CacheSecurityGroup (Maybe Text)

The AWS account ID of the cache security group owner.

CacheNodeTypeSpecificParameter

data CacheNodeTypeSpecificParameter

cacheNodeTypeSpecificParameter :: CacheNodeTypeSpecificParameter

CacheNodeTypeSpecificParameter constructor.

The fields accessible through corresponding lenses are:

cntspAllowedValues :: Maybe Text

cntspCacheNodeTypeSpecificValues :: [CacheNodeTypeSpecificValue]

cntspDataType :: Maybe Text

cntspDescription :: Maybe Text

cntspIsModifiable :: Maybe Bool

cntspMinimumEngineVersion :: Maybe Text

cntspParameterName :: Maybe Text

cntspSource :: Maybe Text

cntspAllowedValues :: Lens' CacheNodeTypeSpecificParameter (Maybe Text)

The valid range of values for the parameter.

cntspCacheNodeTypeSpecificValues :: Lens' CacheNodeTypeSpecificParameter [CacheNodeTypeSpecificValue]

A list of cache node types and their corresponding values for this parameter.

cntspDataType :: Lens' CacheNodeTypeSpecificParameter (Maybe Text)

The valid data type for the parameter.

cntspDescription :: Lens' CacheNodeTypeSpecificParameter (Maybe Text)

A description of the parameter.

cntspIsModifiable :: Lens' CacheNodeTypeSpecificParameter (Maybe Bool)

Indicates whether (true) or not (false) the parameter can be modified. Some parameters have security or operational implications that prevent them from being changed.

cntspMinimumEngineVersion :: Lens' CacheNodeTypeSpecificParameter (Maybe Text)

The earliest cache engine version to which the parameter can apply.

cntspParameterName :: Lens' CacheNodeTypeSpecificParameter (Maybe Text)

The name of the parameter.

cntspSource :: Lens' CacheNodeTypeSpecificParameter (Maybe Text)

The source of the parameter value.

AZMode

data AZMode

Constructors

CrossAz
cross-az

SingleAz
single-az

CacheEngineVersion

data CacheEngineVersion

cacheEngineVersion :: CacheEngineVersion

CacheEngineVersion constructor.

The fields accessible through corresponding lenses are:

cevCacheEngineDescription :: Maybe Text

cevCacheEngineVersionDescription :: Maybe Text

cevCacheParameterGroupFamily :: Maybe Text

cevEngine :: Maybe Text

cevEngineVersion :: Maybe Text

cevCacheEngineDescription :: Lens' CacheEngineVersion (Maybe Text)

The description of the cache engine.

cevCacheEngineVersionDescription :: Lens' CacheEngineVersion (Maybe Text)

The description of the cache engine version.

cevCacheParameterGroupFamily :: Lens' CacheEngineVersion (Maybe Text)

The name of the cache parameter group family associated with this cache engine.

cevEngine :: Lens' CacheEngineVersion (Maybe Text)

The name of the cache engine.

cevEngineVersion :: Lens' CacheEngineVersion (Maybe Text)

The version number of the cache engine.

ReplicationGroup

data ReplicationGroup

replicationGroup :: ReplicationGroup

ReplicationGroup constructor.

The fields accessible through corresponding lenses are:

rgAutomaticFailover :: Maybe AutomaticFailoverStatus

rgDescription :: Maybe Text

rgMemberClusters :: [Text]

rgNodeGroups :: [NodeGroup]

rgPendingModifiedValues :: Maybe ReplicationGroupPendingModifiedValues

rgReplicationGroupId :: Maybe Text

rgSnapshottingClusterId :: Maybe Text

rgStatus :: Maybe Text

rgAutomaticFailover :: Lens' ReplicationGroup (Maybe AutomaticFailoverStatus)

Indicates the status of Multi-AZ for this replication group.

ElastiCache Multi-AZ replication groups are not supported on:

Redis versions earlier than 2.8.6. T1 and T2 cache node types.

rgDescription :: Lens' ReplicationGroup (Maybe Text)

The description of the replication group.

rgMemberClusters :: Lens' ReplicationGroup [Text]

The names of all the cache clusters that are part of this replication group.

rgNodeGroups :: Lens' ReplicationGroup [NodeGroup]

A single element list with information about the nodes in the replication group.

rgPendingModifiedValues :: Lens' ReplicationGroup (Maybe ReplicationGroupPendingModifiedValues)

A group of settings to be applied to the replication group, either immediately or during the next maintenance window.

rgReplicationGroupId :: Lens' ReplicationGroup (Maybe Text)

The identifier for the replication group.

rgSnapshottingClusterId :: Lens' ReplicationGroup (Maybe Text)

The cache cluster ID that is used as the daily snapshot source for the replication group.

rgStatus :: Lens' ReplicationGroup (Maybe Text)

The current state of this replication group - creating, available, etc.

RecurringCharge

data RecurringCharge

recurringCharge :: RecurringCharge

RecurringCharge constructor.

The fields accessible through corresponding lenses are:

rcRecurringChargeAmount :: Maybe Double

rcRecurringChargeFrequency :: Maybe Text

rcRecurringChargeAmount :: Lens' RecurringCharge (Maybe Double)

The monetary amount of the recurring charge.

rcRecurringChargeFrequency :: Lens' RecurringCharge (Maybe Text)

The frequency of the recurring charge.

ReservedCacheNodesOffering

data ReservedCacheNodesOffering

reservedCacheNodesOffering :: ReservedCacheNodesOffering

ReservedCacheNodesOffering constructor.

The fields accessible through corresponding lenses are:

rcnoCacheNodeType :: Maybe Text

rcnoDuration :: Maybe Int

rcnoFixedPrice :: Maybe Double

rcnoOfferingType :: Maybe Text

rcnoProductDescription :: Maybe Text

rcnoRecurringCharges :: [RecurringCharge]

rcnoReservedCacheNodesOfferingId :: Maybe Text

rcnoUsagePrice :: Maybe Double

rcnoCacheNodeType :: Lens' ReservedCacheNodesOffering (Maybe Text)

The cache node type for the reserved cache node.

Valid node types are as follows:

General purpose: Current generation: 'cache.t2.micro', 'cache.t2.small', 'cache.t2.medium', 'cache.m3.medium', 'cache.m3.large', 'cache.m3.xlarge', 'cache.m3.2xlarge' Previous generation: 'cache.t1.micro', 'cache.m1.small', 'cache.m1.medium', 'cache.m1.large', 'cache.m1.xlarge' Compute optimized: 'cache.c1.xlarge' Memory optimized Current generation: 'cache.r3.large', 'cache.r3.xlarge', 'cache.r3.2xlarge', 'cache.r3.4xlarge', 'cache.r3.8xlarge' Previous generation: 'cache.m2.xlarge', 'cache.m2.2xlarge', 'cache.m2.4xlarge' Notes:

All t2 instances are created in an Amazon Virtual Private Cloud (VPC). Redis backup/restore is not supported for t2 instances. Redis Append-only files (AOF) functionality is not supported for t1 or t2 instances. For a complete listing of cache node types and specifications, see Amazon ElastiCache Product Features and Details and Cache NodeType-Specific Parameters for Memcached or Cache Node Type-Specific Parametersfor Redis.

rcnoDuration :: Lens' ReservedCacheNodesOffering (Maybe Int)

The duration of the offering. in seconds.

rcnoFixedPrice :: Lens' ReservedCacheNodesOffering (Maybe Double)

The fixed price charged for this offering.

rcnoOfferingType :: Lens' ReservedCacheNodesOffering (Maybe Text)

The offering type.

rcnoProductDescription :: Lens' ReservedCacheNodesOffering (Maybe Text)

The cache engine used by the offering.

rcnoRecurringCharges :: Lens' ReservedCacheNodesOffering [RecurringCharge]

The recurring price charged to run this reserved cache node.

rcnoReservedCacheNodesOfferingId :: Lens' ReservedCacheNodesOffering (Maybe Text)

A unique identifier for the reserved cache node offering.

rcnoUsagePrice :: Lens' ReservedCacheNodesOffering (Maybe Double)

The hourly price charged for this offering.

TagListMessage

data TagListMessage

tagListMessage :: TagListMessage

TagListMessage constructor.

The fields accessible through corresponding lenses are:

tlmTagList :: [Tag]

tlmTagList :: Lens' TagListMessage [Tag]

A list of cost allocation tags as key-value pairs.

Endpoint

data Endpoint

endpoint :: Endpoint

Endpoint constructor.

The fields accessible through corresponding lenses are:

eAddress :: Maybe Text

ePort :: Maybe Int

eAddress :: Lens' Endpoint (Maybe Text)

The DNS hostname of the cache node.

ePort :: Lens' Endpoint (Maybe Int)

The port number that the cache engine is listening on.

PendingModifiedValues

data PendingModifiedValues

pendingModifiedValues :: PendingModifiedValues

PendingModifiedValues constructor.

The fields accessible through corresponding lenses are:

pmvCacheNodeIdsToRemove :: [Text]

pmvEngineVersion :: Maybe Text

pmvNumCacheNodes :: Maybe Int

pmvCacheNodeIdsToRemove :: Lens' PendingModifiedValues [Text]

A list of cache node IDs that are being removed (or will be removed) from the cache cluster. A node ID is a numeric identifier (0001, 0002, etc.).

pmvEngineVersion :: Lens' PendingModifiedValues (Maybe Text)

The new cache engine version that the cache cluster will run.

pmvNumCacheNodes :: Lens' PendingModifiedValues (Maybe Int)

The new number of cache nodes for the cache cluster.

For clusters running Redis, this value must be 1. For clusters running Memcached, this value must be between 1 and 20.

CacheParameterGroupNameMessage

data CacheParameterGroupNameMessage

cacheParameterGroupNameMessage :: CacheParameterGroupNameMessage

CacheParameterGroupNameMessage constructor.

The fields accessible through corresponding lenses are:

cpgnmCacheParameterGroupName :: Maybe Text

cpgnmCacheParameterGroupName :: Lens' CacheParameterGroupNameMessage (Maybe Text)

The name of the cache parameter group.

Parameter

data Parameter

parameter :: Parameter

Parameter constructor.

The fields accessible through corresponding lenses are:

pAllowedValues :: Maybe Text

pDataType :: Maybe Text

pDescription :: Maybe Text

pIsModifiable :: Maybe Bool

pMinimumEngineVersion :: Maybe Text

pParameterName :: Maybe Text

pParameterValue :: Maybe Text

pSource :: Maybe Text

pAllowedValues :: Lens' Parameter (Maybe Text)

The valid range of values for the parameter.

pDataType :: Lens' Parameter (Maybe Text)

The valid data type for the parameter.

pDescription :: Lens' Parameter (Maybe Text)

A description of the parameter.

pIsModifiable :: Lens' Parameter (Maybe Bool)

Indicates whether (true) or not (false) the parameter can be modified. Some parameters have security or operational implications that prevent them from being changed.

pMinimumEngineVersion :: Lens' Parameter (Maybe Text)

The earliest cache engine version to which the parameter can apply.

pParameterName :: Lens' Parameter (Maybe Text)

The name of the parameter.

pParameterValue :: Lens' Parameter (Maybe Text)

The value of the parameter.

pSource :: Lens' Parameter (Maybe Text)

The source of the parameter.