Network.AWS.ElastiCache.V2014_07_15

Description

ElastiCache is a web service that makes it easy to deploy, operate, and scale an in-memory cache in the cloud. The service improves the performance of web applications by allowing you to retrieve information from fast, managed, in-memory caches, instead of relying entirely on slower disk-based databases. Amazon ElastiCache automatically detects and replaces failed nodes, reducing the overhead associated with self-managed infrastructures and provides a resilient system that mitigates the risk of overloaded databases, which slow website and application load times. Through integration with Amazon CloudWatch, Amazon ElastiCache provides enhanced visibility into key performance metrics associated with your Memcached or Redis nodes.

Documentation

module Network.AWS.ElastiCache.V2014_07_15.AuthorizeCacheSecurityGroupIngress

module Network.AWS.ElastiCache.V2014_07_15.CopySnapshot

module Network.AWS.ElastiCache.V2014_07_15.CreateCacheCluster

module Network.AWS.ElastiCache.V2014_07_15.CreateCacheParameterGroup

module Network.AWS.ElastiCache.V2014_07_15.CreateCacheSecurityGroup

module Network.AWS.ElastiCache.V2014_07_15.CreateCacheSubnetGroup

module Network.AWS.ElastiCache.V2014_07_15.CreateReplicationGroup

module Network.AWS.ElastiCache.V2014_07_15.CreateSnapshot

module Network.AWS.ElastiCache.V2014_07_15.DeleteCacheCluster

module Network.AWS.ElastiCache.V2014_07_15.DeleteCacheParameterGroup

module Network.AWS.ElastiCache.V2014_07_15.DeleteCacheSecurityGroup

module Network.AWS.ElastiCache.V2014_07_15.DeleteCacheSubnetGroup

module Network.AWS.ElastiCache.V2014_07_15.DeleteReplicationGroup

module Network.AWS.ElastiCache.V2014_07_15.DeleteSnapshot

module Network.AWS.ElastiCache.V2014_07_15.DescribeCacheClusters

module Network.AWS.ElastiCache.V2014_07_15.DescribeCacheEngineVersions

module Network.AWS.ElastiCache.V2014_07_15.DescribeCacheParameterGroups

module Network.AWS.ElastiCache.V2014_07_15.DescribeCacheParameters

module Network.AWS.ElastiCache.V2014_07_15.DescribeCacheSecurityGroups

module Network.AWS.ElastiCache.V2014_07_15.DescribeCacheSubnetGroups

module Network.AWS.ElastiCache.V2014_07_15.DescribeEngineDefaultParameters

module Network.AWS.ElastiCache.V2014_07_15.DescribeEvents

module Network.AWS.ElastiCache.V2014_07_15.DescribeReplicationGroups

module Network.AWS.ElastiCache.V2014_07_15.DescribeReservedCacheNodes

module Network.AWS.ElastiCache.V2014_07_15.DescribeReservedCacheNodesOfferings

module Network.AWS.ElastiCache.V2014_07_15.DescribeSnapshots

module Network.AWS.ElastiCache.V2014_07_15.ModifyCacheCluster

module Network.AWS.ElastiCache.V2014_07_15.ModifyCacheParameterGroup

module Network.AWS.ElastiCache.V2014_07_15.ModifyCacheSubnetGroup

module Network.AWS.ElastiCache.V2014_07_15.ModifyReplicationGroup

module Network.AWS.ElastiCache.V2014_07_15.PurchaseReservedCacheNodesOffering

module Network.AWS.ElastiCache.V2014_07_15.RebootCacheCluster

module Network.AWS.ElastiCache.V2014_07_15.ResetCacheParameterGroup

module Network.AWS.ElastiCache.V2014_07_15.RevokeCacheSecurityGroupIngress

module Network.AWS.ElastiCache.V2014_07_15.Types