Network.AWS.ElastiCache.V2014_07_15.CreateCacheParameterGroup

Description

The CreateCacheParameterGroup operation creates a new cache parameter group. A cache parameter group is a collection of parameters that you apply to all of the nodes in a cache cluster. https:/elasticache.us-east-1.amazonaws.com ?Action=CreateCacheParameterGroup &Description=My%20first%20cache%20parameter%20group &CacheParameterGroupFamily=memcached1.4 &CacheParameterGroupName=mycacheparametergroup1 &Version=2014-03-24 &SignatureVersion=4 &SignatureMethod=HmacSHA256 &Timestamp=20140401T192317Z &X-Amz-Credential= mycacheparametergroup3 memcached1.4 My first cache parameter group 05699541-b7f9-11e0-9326-b7275b9d4a6c.

Request

data CreateCacheParameterGroup

Represents the input of a CreateCacheParameterGroup operation.

Request constructor

mkCreateCacheParameterGroup

Arguments

:: Text
ccpgCacheParameterGroupName

-> Text
ccpgCacheParameterGroupFamily

-> Text
ccpgDescription

-> CreateCacheParameterGroup

Smart constructor for the minimum required parameters to construct a valid CreateCacheParameterGroup request.

The fields accessible through corresponding lenses are:

CacheParameterGroupName :: Text

CacheParameterGroupFamily :: Text

Description :: Text

Request lenses

ccpgCacheParameterGroupName :: Lens' CreateCacheParameterGroup Text

A user-specified name for the cache parameter group.

ccpgCacheParameterGroupFamily :: Lens' CreateCacheParameterGroup Text

The name of the cache parameter group family the cache parameter group can be used with. Valid values are: memcached1.4 | redis2.6 | redis2.8.

ccpgDescription :: Lens' CreateCacheParameterGroup Text

A user-specified description for the cache parameter group.

Response

data CreateCacheParameterGroupResponse

Response constructor

mkCreateCacheParameterGroupResponse :: CreateCacheParameterGroupResponse

Smart constructor for the minimum required parameters to construct a valid CreateCacheParameterGroupResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

CacheParameterGroup :: Maybe CacheParameterGroup

Response lenses

ccpgrCacheParameterGroup :: Lens' CreateCacheParameterGroupResponse (Maybe CacheParameterGroup)

Represents the output of a CreateCacheParameterGroup operation.