Network.AWS.ElastiCache.V2014_07_15.CreateCacheSubnetGroup

Description

The CreateCacheSubnetGroup operation creates a new cache subnet group. Use this parameter only when you are creating a cluster in an Amazon Virtual Private Cloud (VPC). https:/elasticache.amazonaws.com ?Action=CreateCacheSubnetGroup &CacheSubnetGroupName=myCachesubnetgroup &CacheSubnetGroupDescription=My%20new%20CacheSubnetGroup &Version=2014-03-24 &SignatureVersion=4 &SignatureMethod=HmacSHA256 &Timestamp=20140401T192317Z &X-Amz-Credential= 990524496922 My new CacheSubnetGroup myCachesubnetgroup Active subnet-7c5b4115 us-east-1c Active subnet-7b5b4112 us-east-1b Active subnet-3ea6bd57 us-east-1d ed662948-a57b-11df-9e38-7ffab86c801f.

Documentation

data CreateCacheSubnetGroup

Constructors

CreateCacheSubnetGroup

Fields

_ccsgnCacheSubnetGroupName :: Text

A name for the cache subnet group. This value is stored as a lowercase string. Constraints: Must contain no more than 255 alphanumeric characters or hyphens. Example: mysubnetgroup.

_ccsgnCacheSubnetGroupDescription :: Text

A description for the cache subnet group.

_ccsgnSubnetIds :: [Text]

A list of VPC subnet IDs for the cache subnet group.

ccsgnSubnetIds :: Lens' CreateCacheSubnetGroup [Text]

ccsgnCacheSubnetGroupName :: Lens' CreateCacheSubnetGroup Text

ccsgnCacheSubnetGroupDescription :: Lens' CreateCacheSubnetGroup Text

data CreateCacheSubnetGroupResponse

Constructors

CreateCacheSubnetGroupResponse

Fields

_csgzCacheSubnetGroup :: Maybe CacheSubnetGroup

Represents the output of one of the following operations: CreateCacheSubnetGroup ModifyCacheSubnetGroup.

csgzCacheSubnetGroup :: Iso' CreateCacheSubnetGroupResponse (Maybe CacheSubnetGroup)