Network.AWS.ElastiCache.V2014_07_15.CreateSnapshot

Description

The CreateSnapshot operation creates a copy of an entire cache cluster at a specific moment in time. https:/elasticache.us-east-1.amazonaws.com ?Action=CreateSnapshot &CacheClusterId=my-redis-primary &SnapshotName=my-manual-snapshot &Version=2014-03-24 &SignatureVersion=4 &SignatureMethod=HmacSHA256 &Timestamp=20140401T192317Z &X-Amz-Credential= my-redis-primary 6379 cache.m1.small default.redis2.8 redis us-east-1d 2014-04-01T18:46:57.972Z 2.8.6 manual true wed:09:00-wed:10:00 my-manual-snapshot 5 2014-04-01T18:46:57.972Z 0001 creating 1 07:30-08:30 faf5a232-b9ce-11e3-8a16-7978bb24ffdf.

Documentation

data CreateSnapshot

Constructors

CreateSnapshot

Fields

_csoCacheClusterId :: Text

The identifier of an existing cache cluster. The snapshot will be created from this cache cluster.

_csoSnapshotName :: Text

A name for the snapshot being created.

csoSnapshotName :: Lens' CreateSnapshot Text

csoCacheClusterId :: Lens' CreateSnapshot Text

data CreateSnapshotResponse

Constructors

CreateSnapshotResponse

Fields

_sssssssssssssssssssssrSnapshot :: Maybe Snapshot

Represents a copy of an entire cache cluster as of the time when the snapshot was taken.

sssssssssssssssssssssrSnapshot :: Iso' CreateSnapshotResponse (Maybe Snapshot)