Network.AWS.ElastiCache.V2014_07_15.DeleteCacheCluster

Description

The DeleteCacheCluster operation deletes a previously provisioned cache cluster. DeleteCacheCluster deletes all associated cache nodes, node endpoints and the cache cluster itself. When you receive a successful response from this operation, Amazon ElastiCache immediately begins deleting the cache cluster; you cannot cancel or revert this operation. https:/elasticache.us-east-1.amazonaws.com ?Action=DeleteCacheCluster &CacheClusterId=simcoprod43&Version=2014-03-24 &Version=2014-03-24 &SignatureVersion=4 &SignatureMethod=HmacSHA256 &Timestamp=20140401T192317Z &X-Amz-Credential= in-sync default.memcached1.4 simcoprod43 deleting 11211 simcoprod43.m2st2p.cfg.cache.amazonaws.com cache.m1.large memcached us-east-1b 2014-03-27T02:18:26.497Z 1.4.5 true mon:05:00-mon:05:30 default active 3 ab84aa7e-b7fa-11e0-9b0b-a9261be2b354.

Documentation

deleteCacheCluster

Arguments

:: Text
_dccmCacheClusterId

-> DeleteCacheCluster

Minimum specification for a DeleteCacheCluster request.

data DeleteCacheCluster

Constructors

DeleteCacheCluster

Fields

_dccmCacheClusterId :: Text

The cache cluster identifier for the cluster to be deleted. This parameter is not case sensitive.

_dccmFinalSnapshotIdentifier :: Maybe Text

The name of a final cache cluster snapshot. ElastiCache creates the snapshot, and then deletes the cache cluster immediately afterward.

dccmFinalSnapshotIdentifier :: Lens' DeleteCacheCluster (Maybe Text)

dccmCacheClusterId :: Lens' DeleteCacheCluster Text

data DeleteCacheClusterResponse

Constructors

DeleteCacheClusterResponse

Fields

_ccwCacheCluster :: Maybe CacheCluster

Contains all of the attributes of a specific cache cluster.

ccwCacheCluster :: Iso' DeleteCacheClusterResponse (Maybe CacheCluster)