Network.AWS.ElastiCache.V2014_07_15.DeleteCacheParameterGroup

Description

The DeleteCacheParameterGroup operation deletes the specified cache parameter group. You cannot delete a cache parameter group if it is associated with any cache clusters. https:/elasticache.us-east-1.amazonaws.com ?Action=DeleteCacheParameterGroup &CacheParameterGroupName=myparametergroup &Version=2014-03-24 &SignatureVersion=4 &SignatureMethod=HmacSHA256 &Timestamp=20140401T192317Z &X-Amz-Credential= d0a417cb-575b-11e0-8869-cd22b4f9d96f.

Request

data DeleteCacheParameterGroup

Represents the input of a DeleteCacheParameterGroup operation.

Request constructor

mkDeleteCacheParameterGroup

Arguments

:: Text
dcpgCacheParameterGroupName

-> DeleteCacheParameterGroup

Smart constructor for the minimum required parameters to construct a valid DeleteCacheParameterGroup request.

Request lenses

dcpgCacheParameterGroupName :: Lens' DeleteCacheParameterGroup Text

The name of the cache parameter group to delete. The specified cache security group must not be associated with any cache clusters.

Response

data DeleteCacheParameterGroupResponse