Network.AWS.ElastiCache.V2014_07_15.DeleteCacheSubnetGroup

Description

The DeleteCacheSubnetGroup operation deletes a cache subnet group. You cannot delete a cache subnet group if it is associated with any cache clusters. https:/elasticache.amazonaws.com ?Action=DeleteCacheSubnetGroup &CacheSubnetGroupName=mysubnetgroup &Version=2014-03-24 &SignatureVersion=4 &SignatureMethod=HmacSHA256 &Timestamp=20140401T192317Z &X-Amz-Credential= 5d013245-4172-11df-8520-e7e1e602a915.

Documentation

data DeleteCacheSubnetGroup

Constructors

DeleteCacheSubnetGroup

Fields

_dcsgpCacheSubnetGroupName :: Text

The name of the cache subnet group to delete. Constraints: Must contain no more than 255 alphanumeric characters or hyphens.

dcsgpCacheSubnetGroupName :: Iso' DeleteCacheSubnetGroup Text

data DeleteCacheSubnetGroupResponse

Constructors

DeleteCacheSubnetGroupResponse