Network.AWS.ElastiCache.V2014_07_15.DeleteReplicationGroup

Description

The DeleteReplicationGroup operation deletes an existing replication group. By default, this operation deletes the entire replication group, including the primary cache cluster and all of the read replicas. You can optionally delete only the read replicas, while retaining the primary cache cluster. When you receive a successful response from this operation, Amazon ElastiCache immediately begins deleting the selected resources; you cannot cancel or revert this operation. https:/elasticache.us-east-1.amazonaws.com ?Action=DeleteReplicationGroup &RetainPrimaryCluster=false &FinalSnapshotIdentifier=my-final-snapshot &ReplicationGroupId=my-repgroup &Version=2014-03-24 &SignatureVersion=4 &SignatureMethod=HmacSHA256 &Timestamp=20140401T192317Z &X-Amz-Credential= my-redis-primary my-repgroup deleting My replication group 93eb37db-b9d7-11e3-8a16-7978bb24ffdf.

Documentation

deleteReplicationGroup

Arguments

:: Text
_drgmReplicationGroupId

-> DeleteReplicationGroup

Minimum specification for a DeleteReplicationGroup request.

data DeleteReplicationGroup

Constructors

DeleteReplicationGroup

Fields

_drgmReplicationGroupId :: Text

The identifier for the replication group to be deleted. This parameter is not case sensitive.

_drgmRetainPrimaryCluster :: Maybe Bool

If set to true, all of the read replicas will be deleted, but the primary cache cluster will be retained.

_drgmFinalSnapshotIdentifier :: Maybe Text

The name of a final cache cluster snapshot. ElastiCache creates the snapshot from the primary cluster in the replication group, rather than one of the replicas; this is to ensure that it captures the freshest data. After the final snapshot is taken, the replication group is deleted immediately afterward.

drgmRetainPrimaryCluster :: Lens' DeleteReplicationGroup (Maybe Bool)

drgmReplicationGroupId :: Lens' DeleteReplicationGroup Text

drgmFinalSnapshotIdentifier :: Lens' DeleteReplicationGroup (Maybe Text)

data DeleteReplicationGroupResponse

Constructors

DeleteReplicationGroupResponse

Fields

_rgxReplicationGroup :: Maybe ReplicationGroup

Contains all of the attributes of a specific replication group.

rgxReplicationGroup :: Iso' DeleteReplicationGroupResponse (Maybe ReplicationGroup)