Network.AWS.ElastiCache.V2014_07_15.DeleteSnapshot

Description

The DeleteSnapshot operation deletes an existing snapshot. When you receive a successful response from this operation, ElastiCache immediately begins deleting the snapshot; you cannot cancel or revert this operation. https:/elasticache.us-east-1.amazonaws.com ?Action=DeleteSnapshot &SnapshotName=my-manual-snapshot &Version=2014-03-24 &SignatureVersion=4 &SignatureMethod=HmacSHA256 &Timestamp=20140401T192317Z &X-Amz-Credential= my-redis-primary 6379 cache.m1.small default.redis2.8 redis us-east-1d 2014-04-01T18:46:57.972Z 2.8.6 manual true wed:09:00-wed:10:00 my-manual-snapshot 5 2014-04-01T18:54:12Z 2014-04-01T18:46:57.972Z 0001 3 MB deleting 1 07:30-08:30 694d7017-b9d2-11e3-8a16-7978bb24ffdf.

Request

data DeleteSnapshot

Represents the input of a DeleteSnapshotMessage operation.

Request constructor

mkDeleteSnapshot

Arguments

:: Text
dsSnapshotName

-> DeleteSnapshot

Smart constructor for the minimum required parameters to construct a valid DeleteSnapshot request.

Request lenses

dsSnapshotName :: Lens' DeleteSnapshot Text

The name of the snapshot to be deleted.

Response

data DeleteSnapshotResponse

Response lenses

dsrSnapshot :: Lens' DeleteSnapshotResponse (Maybe Snapshot)

Represents a copy of an entire cache cluster as of the time when the snapshot was taken.