Network.AWS.ElastiCache.V2014_07_15.DescribeCacheEngineVersions

Description

The DescribeCacheEngineVersions operation returns a list of the available cache engines and their versions. https:/elasticache.us-east-1.amazonaws.com ?Action=DescribeCacheEngineVersions &MaxRecords=100 &Version=2014-03-24 &SignatureVersion=4 &SignatureMethod=HmacSHA256 &Timestamp=20140401T192317Z &X-Amz-Credential= memcached1.4 memcached memcached version 1.4.14 memcached 1.4.14 memcached1.4 memcached memcached version 1.4.5 memcached 1.4.5 a6ac9ad2-f8a4-11e1-a4d1-a345e5370093.

Documentation

describeCacheEngineVersions :: DescribeCacheEngineVersions

Minimum specification for a DescribeCacheEngineVersions request.

data DescribeCacheEngineVersions

Constructors

DescribeCacheEngineVersions

Fields

_dcevmDefaultOnly :: Maybe Bool

If true, specifies that only the default version of the specified engine or engine and major version combination is to be returned.

_dcevmMaxRecords :: Maybe Integer

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: minimum 20; maximum 100.

_dcevmEngineVersion :: Maybe Text

The cache engine version to return. Example: 1.4.14.

_dcevmCacheParameterGroupFamily :: Maybe Text

The name of a specific cache parameter group family to return details for. Constraints: Must be 1 to 255 alphanumeric characters First character must be a letter Cannot end with a hyphen or contain two consecutive hyphens.

_dcevmEngine :: Maybe Text

The cache engine to return. Valid values: memcached | redis.

_dcevmMarker :: Maybe Text

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

dcevmMaxRecords :: Lens' DescribeCacheEngineVersions (Maybe Integer)

dcevmMarker :: Lens' DescribeCacheEngineVersions (Maybe Text)

dcevmEngineVersion :: Lens' DescribeCacheEngineVersions (Maybe Text)

dcevmEngine :: Lens' DescribeCacheEngineVersions (Maybe Text)

dcevmDefaultOnly :: Lens' DescribeCacheEngineVersions (Maybe Bool)

dcevmCacheParameterGroupFamily :: Lens' DescribeCacheEngineVersions (Maybe Text)

data DescribeCacheEngineVersionsResponse

Constructors

DescribeCacheEngineVersionsResponse

Fields

_cevmCacheEngineVersions :: [CacheEngineVersion]

A list of cache engine version details. Each element in the list contains detailed information about once cache engine version.

_cevmMarker :: Maybe Text

Provides an identifier to allow retrieval of paginated results.

cevmMarker :: Lens' DescribeCacheEngineVersionsResponse (Maybe Text)

cevmCacheEngineVersions :: Lens' DescribeCacheEngineVersionsResponse [CacheEngineVersion]