Network.AWS.ElastiCache.V2014_07_15.DescribeCacheParameters

Description

The DescribeCacheParameters operation returns the detailed parameter list for a particular cache parameter group. Some of the output has been omitted for brevity. https:/elasticache.us-east-1.amazonaws.com ?Action=DescribeCacheParameters &CacheParameterGroupName=default.memcached1.4 &MaxRecords=100 &Version=2014-03-24 &SignatureVersion=4 &SignatureMethod=HmacSHA256 &Timestamp=20140401T192317Z &X-Amz-Credential= 1-07-15/"> cache.c1.xlarge 6000 (...output omitted...) integer system false The maximum configurable amount of memory to use to store items, in megabytes. 1-100000 max_cache_memory 1.4.5 (...output omitted...) 1024 integer system false The backlog queue limit. 1-10000 backlog_queue_limit 1.4.5 (...output omitted...) 0c507368-b7fe-11e0-9326-b7275b9d4a6c.

Request

data DescribeCacheParameters

Represents the input of a DescribeCacheParameters operation.

Request constructor

mkDescribeCacheParameters

Arguments

:: Text
dcpCacheParameterGroupName

-> DescribeCacheParameters

Smart constructor for the minimum required parameters to construct a valid DescribeCacheParameters request.

Request lenses

dcpCacheParameterGroupName :: Lens' DescribeCacheParameters Text

The name of a specific cache parameter group to return details for.

dcpSource :: Lens' DescribeCacheParameters (Maybe Text)

The parameter types to return. Valid values: user | system | engine-default.

dcpMaxRecords :: Lens' DescribeCacheParameters (Maybe Integer)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: minimum 20; maximum 100.

dcpMarker :: Lens' DescribeCacheParameters (Maybe Text)

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Response

data DescribeCacheParametersResponse

Represents the output of a DescribeCacheParameters operation.

Response lenses

dcprMarker :: Lens' DescribeCacheParametersResponse (Maybe Text)

Provides an identifier to allow retrieval of paginated results.

dcprParameters :: Lens' DescribeCacheParametersResponse [Parameter]

A list of Parameter instances.

dcprCacheNodeTypeSpecificParameters :: Lens' DescribeCacheParametersResponse [CacheNodeTypeSpecificParameter]

A list of parameters specific to a particular cache node type. Each element in the list contains detailed information about one parameter.