Network.AWS.ElastiCache.V2014_07_15.DescribeCacheSecurityGroups

Description

The DescribeCacheSecurityGroups operation returns a list of cache security group descriptions. If a cache security group name is specified, the list will contain only the description of that group. https:/elasticache.us-east-1.amazonaws.com ?Action=DescribeCacheSecurityGroups &MaxRecords=100 &Version=2014-03-24 &SignatureVersion=4 &SignatureMethod=HmacSHA256 &Timestamp=20140401T192317Z &X-Amz-Credential= default 123456789012 default mycachesecuritygroup 123456789012 My Security Group a95360ae-b7fc-11e0-9326-b7275b9d4a6c.

Request

data DescribeCacheSecurityGroups

Represents the input of a DescribeCacheSecurityGroups operation.

Request constructor

mkDescribeCacheSecurityGroups :: DescribeCacheSecurityGroups

Smart constructor for the minimum required parameters to construct a valid DescribeCacheSecurityGroups request.

The fields accessible through corresponding lenses are:

CacheSecurityGroupName :: Maybe Text

MaxRecords :: Maybe Integer

Marker :: Maybe Text

Request lenses

dcsg2CacheSecurityGroupName :: Lens' DescribeCacheSecurityGroups (Maybe Text)

The name of the cache security group to return details for.

dcsg2MaxRecords :: Lens' DescribeCacheSecurityGroups (Maybe Integer)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: minimum 20; maximum 100.

dcsg2Marker :: Lens' DescribeCacheSecurityGroups (Maybe Text)

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Response

data DescribeCacheSecurityGroupsResponse

Represents the output of a DescribeCacheSecurityGroups operation.

Response constructor

mkDescribeCacheSecurityGroupsResponse :: DescribeCacheSecurityGroupsResponse

Smart constructor for the minimum required parameters to construct a valid DescribeCacheSecurityGroupsResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Marker :: Maybe Text

CacheSecurityGroups :: [CacheSecurityGroup]

Response lenses

dcsgrMarker :: Lens' DescribeCacheSecurityGroupsResponse (Maybe Text)

Provides an identifier to allow retrieval of paginated results.

dcsgrCacheSecurityGroups :: Lens' DescribeCacheSecurityGroupsResponse [CacheSecurityGroup]

A list of cache security groups. Each element in the list contains detailed information about one group.