Network.AWS.ElastiCache.V2014_07_15.DescribeCacheSubnetGroups

Description

The DescribeCacheSubnetGroups operation returns a list of cache subnet group descriptions. If a subnet group name is specified, the list will contain only the description of that group. Some of the output has been omitted for brevity. https:/elasticache.amazonaws.com ?Action=DescribeCacheSubnetGroups &Version=2014-03-24 &SignatureVersion=4 &SignatureMethod=HmacSHA256 &Timestamp=20140401T192317Z &X-Amz-Credential= 990524496922 description subnet_grp1 Active subnet-7c5b4115 us-east-1c Active subnet-7b5b4112 us-east-1b Active subnet-3ea6bd57 us-east-1d (...output omitted...) 31d0faee-229b-11e1-81f1-df3a2a803dad.

Request

data DescribeCacheSubnetGroups

Represents the input of a DescribeCacheSubnetGroups operation.

Request constructor

mkDescribeCacheSubnetGroups :: DescribeCacheSubnetGroups

Smart constructor for the minimum required parameters to construct a valid DescribeCacheSubnetGroups request.

Request lenses

dcsg3CacheSubnetGroupName :: Lens' DescribeCacheSubnetGroups (Maybe Text)

The name of the cache subnet group to return details for.

dcsg3MaxRecords :: Lens' DescribeCacheSubnetGroups (Maybe Integer)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: minimum 20; maximum 100.

dcsg3Marker :: Lens' DescribeCacheSubnetGroups (Maybe Text)

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Response

data DescribeCacheSubnetGroupsResponse

Represents the output of a DescribeCacheSubnetGroups operation.

Response lenses

dcsgrsrsMarker :: Lens' DescribeCacheSubnetGroupsResponse (Maybe Text)

Provides an identifier to allow retrieval of paginated results.

dcsgrsrsCacheSubnetGroups :: Lens' DescribeCacheSubnetGroupsResponse [CacheSubnetGroup]

A list of cache subnet groups. Each element in the list contains detailed information about one group.