Network.AWS.ElastiCache.V2014_07_15.DescribeEvents

Description

The DescribeEvents operation returns events related to cache clusters, cache security groups, and cache parameter groups. You can obtain events specific to a particular cache cluster, cache security group, or cache parameter group by providing the name as a parameter. By default, only the events occurring within the last hour are returned; however, you can retrieve up to 14 days' worth of events if necessary. Some of the output has been omitted for brevity. https:/elasticache.us-east-1.amazonaws.com ?Action=DescribeEvents &MaxRecords=100 &Version=2014-03-24 &SignatureVersion=4 &SignatureMethod=HmacSHA256 &Timestamp=20140401T192317Z &X-Amz-Credential= Cache cluster created cache-cluster 2014-04-01T18:22:18.202Z my-redis-primary (...output omitted...) e21c81b4-b9cd-11e3-8a16-7978bb24ffdf.

Documentation

describeEvents :: DescribeEvents

Minimum specification for a DescribeEvents request.

data DescribeEvents

Constructors

DescribeEvents

Fields

_demMaxRecords :: Maybe Integer

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: minimum 20; maximum 100.

_demDuration :: Maybe Integer

The number of minutes' worth of events to retrieve.

_demSourceType :: Maybe SourceType

The event source to retrieve events for. If no value is specified, all events are returned. Valid values are: cache-cluster | cache-parameter-group | cache-security-group | cache-subnet-group.

_demSourceIdentifier :: Maybe Text

The identifier of the event source for which events will be returned. If not specified, then all sources are included in the response.

_demMarker :: Maybe Text

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

_demStartTime :: Maybe ISO8601

The beginning of the time interval to retrieve events for, specified in ISO 8601 format.

_demEndTime :: Maybe ISO8601

The end of the time interval for which to retrieve events, specified in ISO 8601 format.

demStartTime :: Lens' DescribeEvents (Maybe ISO8601)

demSourceType :: Lens' DescribeEvents (Maybe SourceType)

demSourceIdentifier :: Lens' DescribeEvents (Maybe Text)

demMaxRecords :: Lens' DescribeEvents (Maybe Integer)

demMarker :: Lens' DescribeEvents (Maybe Text)

demEndTime :: Lens' DescribeEvents (Maybe ISO8601)

demDuration :: Lens' DescribeEvents (Maybe Integer)

data DescribeEventsResponse

Constructors

DescribeEventsResponse

Fields

_emEvents :: [Event]

A list of events. Each element in the list contains detailed information about one event.

_emMarker :: Maybe Text

Provides an identifier to allow retrieval of paginated results.

emMarker :: Lens' DescribeEventsResponse (Maybe Text)

emEvents :: Lens' DescribeEventsResponse [Event]