Network.AWS.ElastiCache.V2014_07_15.DescribeReplicationGroups

Description

The DescribeReplicationGroups operation returns information about a particular replication group. If no identifier is specified, DescribeReplicationGroups returns information about all replication groups. https:/elasticache.us-east-1.amazonaws.com ?Action=DescribeReplicationGroups &MaxRecords=100 &Version=2014-03-24 &SignatureVersion=4 &SignatureMethod=HmacSHA256 &Timestamp=20140401T192317Z &X-Amz-Credential= my-redis-primary my-redis-primary 0001 6379 my-repgroup.q68zge.ng.0001.use1devo.elmo-dev.amazonaws.com available my-redis-primary 6379 my-redis-primary.q68zge.0001.use1devo.elmo-dev.amazonaws.com us-east-1d 0001 primary my-repgroup available My replication group 144745b0-b9d3-11e3-8a16-7978bb24ffdf.

Documentation

describeReplicationGroups :: DescribeReplicationGroups

Minimum specification for a DescribeReplicationGroups request.

data DescribeReplicationGroups

Constructors

DescribeReplicationGroups

Fields

_drgnMaxRecords :: Maybe Integer

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: minimum 20; maximum 100.

_drgnMarker :: Maybe Text

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

_drgnReplicationGroupId :: Maybe Text

The identifier for the replication group to be described. This parameter is not case sensitive. If you do not specify this parameter, information about all replication groups is returned.

drgnReplicationGroupId :: Lens' DescribeReplicationGroups (Maybe Text)

drgnMaxRecords :: Lens' DescribeReplicationGroups (Maybe Integer)

drgnMarker :: Lens' DescribeReplicationGroups (Maybe Text)

data DescribeReplicationGroupsResponse

Constructors

DescribeReplicationGroupsResponse

Fields

_rgmReplicationGroups :: [ReplicationGroup]

A list of replication groups. Each item in the list contains detailed information about one replication group.

_rgmMarker :: Maybe Text

Provides an identifier to allow retrieval of paginated results.

rgmReplicationGroups :: Lens' DescribeReplicationGroupsResponse [ReplicationGroup]

rgmMarker :: Lens' DescribeReplicationGroupsResponse (Maybe Text)