Network.AWS.ElastiCache.V2014_07_15.DescribeReservedCacheNodes

Description

The DescribeReservedCacheNodes operation returns information about reserved cache nodes for this account, or about a specified reserved cache node. https:/elasticache.amazonaws.com ?Action=DescribeReservedCacheNodes &ReservedCacheNodeId=customerSpecifiedID &Version=2014-03-24 &SignatureVersion=4 &SignatureMethod=HmacSHA256 &Timestamp=20140401T192317Z &X-Amz-Credential= Medium Utilization memcached 649fd0c8-cf6d-47a0-bfa6-060f8e75e95f payment-failed myreservationid 1 2010-12-15T00:25:14.131Z 31536000 227.5 0.046 cache.m1.small c695119b-2961-11e1-bd06-6fe008f046c3.

Documentation

describeReservedCacheNodes :: DescribeReservedCacheNodes

Minimum specification for a DescribeReservedCacheNodes request.

data DescribeReservedCacheNodes

Constructors

DescribeReservedCacheNodes

Fields

_drcnmMaxRecords :: Maybe Integer

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: minimum 20; maximum 100.

_drcnmCacheNodeType :: Maybe Text

The cache node type filter value. Use this parameter to show only those reservations matching the specified cache node type.

_drcnmProductDescription :: Maybe Text

The product description filter value. Use this parameter to show only those reservations matching the specified product description.

_drcnmMarker :: Maybe Text

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

_drcnmReservedCacheNodeId :: Maybe Text

The reserved cache node identifier filter value. Use this parameter to show only the reservation that matches the specified reservation ID.

_drcnmOfferingType :: Maybe Text

The offering type filter value. Use this parameter to show only the available offerings matching the specified offering type. Valid values: "Light Utilization" | "Medium Utilization" | "Heavy Utilization".

_drcnmDuration :: Maybe Text

The duration filter value, specified in years or seconds. Use this parameter to show only reservations for this duration. Valid Values: 1 | 3 | 31536000 | 94608000.

_drcnmReservedCacheNodesOfferingId :: Maybe Text

The offering identifier filter value. Use this parameter to show only purchased reservations matching the specified offering identifier.

drcnmReservedCacheNodesOfferingId :: Lens' DescribeReservedCacheNodes (Maybe Text)

drcnmReservedCacheNodeId :: Lens' DescribeReservedCacheNodes (Maybe Text)

drcnmProductDescription :: Lens' DescribeReservedCacheNodes (Maybe Text)

drcnmOfferingType :: Lens' DescribeReservedCacheNodes (Maybe Text)

drcnmMaxRecords :: Lens' DescribeReservedCacheNodes (Maybe Integer)

drcnmMarker :: Lens' DescribeReservedCacheNodes (Maybe Text)

drcnmDuration :: Lens' DescribeReservedCacheNodes (Maybe Text)

drcnmCacheNodeType :: Lens' DescribeReservedCacheNodes (Maybe Text)

data DescribeReservedCacheNodesResponse

Constructors

DescribeReservedCacheNodesResponse

Fields

_rcnmReservedCacheNodes :: [ReservedCacheNode]

A list of reserved cache nodes. Each element in the list contains detailed information about one node.

_rcnmMarker :: Maybe Text

Provides an identifier to allow retrieval of paginated results.

rcnmReservedCacheNodes :: Lens' DescribeReservedCacheNodesResponse [ReservedCacheNode]

rcnmMarker :: Lens' DescribeReservedCacheNodesResponse (Maybe Text)