Network.AWS.ElastiCache.V2014_07_15.DescribeReservedCacheNodesOfferings

Description

The DescribeReservedCacheNodesOfferings operation lists available reserved cache node offerings. https:/elasticache.amazonaws.com ?Action=DescribeReservedCacheNodesOfferings &ReservedCacheNodesOfferingId=438012d3-4052-4cc7-b2e3-8d3372e0e706 &Version=2014-03-24 &SignatureVersion=4 &SignatureMethod=HmacSHA256 &Timestamp=20140401T192317Z &X-Amz-Credential= 31536000 Heavy Utilization Hourly 0.123 162.0 memcached 0.0 SampleOfferingId cache.m1.small 521b420a-2961-11e1-bd06-6fe008f046c3.

Documentation

describeReservedCacheNodesOfferings :: DescribeReservedCacheNodesOfferings

Minimum specification for a DescribeReservedCacheNodesOfferings request.

data DescribeReservedCacheNodesOfferings

Constructors

DescribeReservedCacheNodesOfferings

Fields

_drcnomMaxRecords :: Maybe Integer

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: minimum 20; maximum 100.

_drcnomCacheNodeType :: Maybe Text

The cache node type filter value. Use this parameter to show only the available offerings matching the specified cache node type.

_drcnomProductDescription :: Maybe Text

The product description filter value. Use this parameter to show only the available offerings matching the specified product description.

_drcnomMarker :: Maybe Text

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

_drcnomOfferingType :: Maybe Text

The offering type filter value. Use this parameter to show only the available offerings matching the specified offering type. Valid Values: "Light Utilization" | "Medium Utilization" | "Heavy Utilization".

_drcnomDuration :: Maybe Text

Duration filter value, specified in years or seconds. Use this parameter to show only reservations for a given duration. Valid Values: 1 | 3 | 31536000 | 94608000.

_drcnomReservedCacheNodesOfferingId :: Maybe Text

The offering identifier filter value. Use this parameter to show only the available offering that matches the specified reservation identifier. Example: 438012d3-4052-4cc7-b2e3-8d3372e0e706.

drcnomReservedCacheNodesOfferingId :: Lens' DescribeReservedCacheNodesOfferings (Maybe Text)

drcnomProductDescription :: Lens' DescribeReservedCacheNodesOfferings (Maybe Text)

drcnomOfferingType :: Lens' DescribeReservedCacheNodesOfferings (Maybe Text)

drcnomMaxRecords :: Lens' DescribeReservedCacheNodesOfferings (Maybe Integer)

drcnomMarker :: Lens' DescribeReservedCacheNodesOfferings (Maybe Text)

drcnomDuration :: Lens' DescribeReservedCacheNodesOfferings (Maybe Text)

drcnomCacheNodeType :: Lens' DescribeReservedCacheNodesOfferings (Maybe Text)

data DescribeReservedCacheNodesOfferingsResponse

Constructors

DescribeReservedCacheNodesOfferingsResponse

Fields

_rcnomReservedCacheNodesOfferings :: [ReservedCacheNodesOffering]

A list of reserved cache node offerings. Each element in the list contains detailed information about one offering.

_rcnomMarker :: Maybe Text

Provides an identifier to allow retrieval of paginated results.

rcnomReservedCacheNodesOfferings :: Lens' DescribeReservedCacheNodesOfferingsResponse [ReservedCacheNodesOffering]

rcnomMarker :: Lens' DescribeReservedCacheNodesOfferingsResponse (Maybe Text)