Network.AWS.ElastiCache.V2014_07_15.ModifyCacheSubnetGroup

Description

The ModifyCacheSubnetGroup operation modifies an existing cache subnet group. https:/elasticache.amazonaws.com ?Action=ModifyCacheSubnetGroup &CacheSubnetGroupName=myCachesubnetgroup &CacheSubnetGroupDescription=My%20modified%20CacheSubnetGroup &Version=2014-03-24 &SignatureVersion=4 &SignatureMethod=HmacSHA256 &Timestamp=20140401T192317Z &X-Amz-Credential= 990524496922 My modified CacheSubnetGroup myCachesubnetgroup Active subnet-7c5b4115 us-east-1c Active subnet-7b5b4112 us-east-1b Active subnet-3ea6bd57 us-east-1d ed662948-a57b-11df-9e38-7ffab86c801f.

Request

data ModifyCacheSubnetGroup

Represents the input of a ModifyCacheSubnetGroup operation.

Request constructor

mkModifyCacheSubnetGroup

Arguments

:: Text
mcsgCacheSubnetGroupName

-> ModifyCacheSubnetGroup

Smart constructor for the minimum required parameters to construct a valid ModifyCacheSubnetGroup request.

The fields accessible through corresponding lenses are:

CacheSubnetGroupName :: Text

CacheSubnetGroupDescription :: Maybe Text

SubnetIds :: [Text]

Request lenses

mcsgCacheSubnetGroupName :: Lens' ModifyCacheSubnetGroup Text

The name for the cache subnet group. This value is stored as a lowercase string. Constraints: Must contain no more than 255 alphanumeric characters or hyphens. Example: mysubnetgroup.

mcsgCacheSubnetGroupDescription :: Lens' ModifyCacheSubnetGroup (Maybe Text)

A description for the cache subnet group.

mcsgSubnetIds :: Lens' ModifyCacheSubnetGroup [Text]

The EC2 subnet IDs for the cache subnet group.

Response

data ModifyCacheSubnetGroupResponse

Response constructor

mkModifyCacheSubnetGroupResponse :: ModifyCacheSubnetGroupResponse

Smart constructor for the minimum required parameters to construct a valid ModifyCacheSubnetGroupResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

CacheSubnetGroup :: Maybe CacheSubnetGroup

Response lenses

mcsgrCacheSubnetGroup :: Lens' ModifyCacheSubnetGroupResponse (Maybe CacheSubnetGroup)

Represents the output of one of the following operations: CreateCacheSubnetGroup ModifyCacheSubnetGroup.