Network.AWS.ElastiCache.V2014_07_15.PurchaseReservedCacheNodesOffering

Description

The PurchaseReservedCacheNodesOffering operation allows you to purchase a reserved cache node offering. https:/elasticache.amazonaws.com ?Action=PurchaseReservedCacheNodesOffering &ReservedCacheNodeId=myreservationID &ReservedCacheNodesOfferingId=438012d3-4052-4cc7-b2e3-8d3372e0e706 &CacheNodeCount=1 &Version=2014-03-24 &SignatureVersion=4 &SignatureMethod=HmacSHA256 &Timestamp=20140401T192317Z &X-Amz-Credential= Medium Utilization memcached 438012d3-4052-4cc7-b2e3-8d3372e0e706 payment-pending myreservationID 10 2014-03-18T23:24:56.577Z 31536000 123.0 0.123 cache.m1.small 7f099901-29cf-11e1-bd06-6fe008f046c3.

Request

data PurchaseReservedCacheNodesOffering

Represents the input of a PurchaseReservedCacheNodesOffering operation.

Request constructor

mkPurchaseReservedCacheNodesOffering

Arguments

:: Text
prcnoReservedCacheNodesOfferingId

-> PurchaseReservedCacheNodesOffering

Smart constructor for the minimum required parameters to construct a valid PurchaseReservedCacheNodesOffering request.

Request lenses

prcnoReservedCacheNodesOfferingId :: Lens' PurchaseReservedCacheNodesOffering Text

The ID of the reserved cache node offering to purchase. Example: 438012d3-4052-4cc7-b2e3-8d3372e0e706.

prcnoReservedCacheNodeId :: Lens' PurchaseReservedCacheNodesOffering (Maybe Text)

A customer-specified identifier to track this reservation. Example: myreservationID.

prcnoCacheNodeCount :: Lens' PurchaseReservedCacheNodesOffering (Maybe Integer)

The number of cache node instances to reserve. Default: 1.

Response

data PurchaseReservedCacheNodesOfferingResponse

Response lenses

prcnorsReservedCacheNode :: Lens' PurchaseReservedCacheNodesOfferingResponse (Maybe ReservedCacheNode)

Represents the output of a PurchaseReservedCacheNodesOffering operation.