Network.AWS.ElastiCache.V2014_07_15.RebootCacheCluster

Description

The RebootCacheCluster operation reboots some, or all, of the cache nodes within a provisioned cache cluster. This API will apply any modified cache parameter groups to the cache cluster. The reboot action takes place as soon as possible, and results in a momentary outage to the cache cluster. During the reboot, the cache cluster status is set to REBOOTING. The reboot causes the contents of the cache (for each cache node being rebooted) to be lost. When the reboot is complete, a cache cluster event is created. https:/elasticache.us-east-1.amazonaws.com ?Action=RebootCacheCluster &CacheClusterId=mycache &CacheNodeIdsToReboot.member.1=0001 &CacheNodeIdsToReboot.member.2=0002 &CacheNodeIdsToReboot.member.3=0003 &Version=2014-03-24 &SignatureVersion=4 &SignatureMethod=HmacSHA256 &Timestamp=20140401T192317Z &X-Amz-Credential= rebooting cache cluster nodes default.memcached1.4 in-sync mycache 11211 mycache.q68zge.cfg.use1devo.elmo-dev.amazonaws.com cache.m1.small memcached us-east-1b 2014-04-01T19:04:12.812Z 1.4.17 true wed:09:00-wed:10:00 https:/console.aws.amazon.comelasticache/home#client-download: default active 3 cf7e6fc4-b9d1-11e3-8a16-7978bb24ffdf.

Request

data RebootCacheCluster

Represents the input of a RebootCacheCluster operation.

Request constructor

mkRebootCacheCluster

Arguments

:: Text
rccCacheClusterId

-> [Text]
rccCacheNodeIdsToReboot

-> RebootCacheCluster

Smart constructor for the minimum required parameters to construct a valid RebootCacheCluster request.

Request lenses

rccCacheClusterId :: Lens' RebootCacheCluster Text

The cache cluster identifier. This parameter is stored as a lowercase string.

rccCacheNodeIdsToReboot :: Lens' RebootCacheCluster [Text]

A list of cache node IDs to reboot. A node ID is a numeric identifier (0001, 0002, etc.). To reboot an entire cache cluster, specify all of the cache node IDs.

Response

data RebootCacheClusterResponse

Response lenses

rccrCacheCluster :: Lens' RebootCacheClusterResponse (Maybe CacheCluster)

Contains all of the attributes of a specific cache cluster.