Network.AWS.ElastiCache.V2014_07_15.Types

Description

ElastiCache is a web service that makes it easy to deploy, operate, and scale an in-memory cache in the cloud. The service improves the performance of web applications by allowing you to retrieve information from fast, managed, in-memory caches, instead of relying entirely on slower disk-based databases. Amazon ElastiCache automatically detects and replaces failed nodes, reducing the overhead associated with self-managed infrastructures and provides a resilient system that mitigates the risk of overloaded databases, which slow website and application load times. Through integration with Amazon CloudWatch, Amazon ElastiCache provides enhanced visibility into key performance metrics associated with your Memcached or Redis nodes.

Service

data ElastiCache

Supported version (2014-07-15) of the Amazon ElastiCache service.

XML

xmlOptions :: Tagged a XMLOptions

SourceType

data SourceType

Constructors

SourceTypeCacheCluster
cache-cluster

SourceTypeCacheParameterGroup
cache-parameter-group

SourceTypeCacheSecurityGroup
cache-security-group

SourceTypeCacheSubnetGroup
cache-subnet-group

AvailabilityZone

data AvailabilityZone

The Availability Zone associated with the subnet.

mkAvailabilityZone :: AvailabilityZone

Smart constructor for the minimum required fields to construct a valid AvailabilityZone data type to populate a request.

The fields accessible through corresponding lenses are:

Name :: Maybe Text

azName :: Lens' AvailabilityZone (Maybe Text)

The name of the Availability Zone.

ReplicationGroupPendingModifiedValues

data ReplicationGroupPendingModifiedValues

A group of settings to be applied to the replication group, either immediately or during the next maintenance window.

mkReplicationGroupPendingModifiedValues :: ReplicationGroupPendingModifiedValues

Smart constructor for the minimum required fields to construct a valid ReplicationGroupPendingModifiedValues data type to populate a request.

The fields accessible through corresponding lenses are:

PrimaryClusterId :: Maybe Text

rgpmvPrimaryClusterId :: Lens' ReplicationGroupPendingModifiedValues (Maybe Text)

The primary cluster ID which will be applied immediately (if --apply-immediately was specified), or during the next maintenance window.

CacheCluster

data CacheCluster

Contains all of the attributes of a specific cache cluster.

mkCacheCluster :: CacheCluster

Smart constructor for the minimum required fields to construct a valid CacheCluster data type.

CacheCluster is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

CacheClusterId :: Maybe Text

ConfigurationEndpoint :: Maybe Endpoint

ClientDownloadLandingPage :: Maybe Text

CacheNodeType :: Maybe Text

Engine :: Maybe Text

EngineVersion :: Maybe Text

CacheClusterStatus :: Maybe Text

NumCacheNodes :: Maybe Integer

PreferredAvailabilityZone :: Maybe Text

CacheClusterCreateTime :: Maybe ISO8601

PreferredMaintenanceWindow :: Maybe Text

PendingModifiedValues :: Maybe PendingModifiedValues

NotificationConfiguration :: Maybe NotificationConfiguration

CacheSecurityGroups :: [CacheSecurityGroupMembership]

CacheParameterGroup :: Maybe CacheParameterGroupStatus

CacheSubnetGroupName :: Maybe Text

CacheNodes :: [CacheNode]

AutoMinorVersionUpgrade :: Maybe Bool

SecurityGroups :: [SecurityGroupMembership]

ReplicationGroupId :: Maybe Text

SnapshotRetentionLimit :: Maybe Integer

SnapshotWindow :: Maybe Text

ccCacheClusterId :: Lens' CacheCluster (Maybe Text)

The user-supplied identifier of the cache cluster. This is a unique key that identifies a cache cluster.

ccConfigurationEndpoint :: Lens' CacheCluster (Maybe Endpoint)

Represents the information required for client programs to connect to a cache node.

ccClientDownloadLandingPage :: Lens' CacheCluster (Maybe Text)

The URL of the web page where you can download the latest ElastiCache client library.

ccCacheNodeType :: Lens' CacheCluster (Maybe Text)

The name of the compute and memory capacity node type for the cache cluster.

ccEngine :: Lens' CacheCluster (Maybe Text)

The name of the cache engine (memcached or redis) to be used for this cache cluster.

ccEngineVersion :: Lens' CacheCluster (Maybe Text)

The version of the cache engine version that is used in this cache cluster.

ccCacheClusterStatus :: Lens' CacheCluster (Maybe Text)

The current state of this cache cluster - creating, available, etc.

ccNumCacheNodes :: Lens' CacheCluster (Maybe Integer)

The number of cache nodes in the cache cluster.

ccPreferredAvailabilityZone :: Lens' CacheCluster (Maybe Text)

The name of the Availability Zone in which the cache cluster is located or Multiple if the cache nodes are located in different Availability Zones.

ccCacheClusterCreateTime :: Lens' CacheCluster (Maybe ISO8601)

The date and time when the cache cluster was created.

ccPreferredMaintenanceWindow :: Lens' CacheCluster (Maybe Text)

The time range (in UTC) during which weekly system maintenance can occur.

ccPendingModifiedValues :: Lens' CacheCluster (Maybe PendingModifiedValues)

A group of settings that will be applied to the cache cluster in the future, or that are currently being applied.

ccNotificationConfiguration :: Lens' CacheCluster (Maybe NotificationConfiguration)

Describes a notification topic and its status. Notification topics are used for publishing ElastiCache events to subscribers using Amazon Simple Notification Service (SNS).

ccCacheSecurityGroups :: Lens' CacheCluster [CacheSecurityGroupMembership]

A list of cache security group elements, composed of name and status sub-elements.

ccCacheParameterGroup :: Lens' CacheCluster (Maybe CacheParameterGroupStatus)

The status of the cache parameter group.

ccCacheSubnetGroupName :: Lens' CacheCluster (Maybe Text)

The name of the cache subnet group associated with the cache cluster.

ccCacheNodes :: Lens' CacheCluster [CacheNode]

A list of cache nodes that are members of the cache cluster.

ccAutoMinorVersionUpgrade :: Lens' CacheCluster (Maybe Bool)

If true, then minor version patches are applied automatically; if false, then automatic minor version patches are disabled.

ccSecurityGroups :: Lens' CacheCluster [SecurityGroupMembership]

A list of VPC Security Groups associated with the cache cluster.

ccReplicationGroupId :: Lens' CacheCluster (Maybe Text)

The replication group to which this cache cluster belongs. If this field is empty, the cache cluster is not associated with any replication group.

ccSnapshotRetentionLimit :: Lens' CacheCluster (Maybe Integer)

The number of days for which ElastiCache will retain automatic cache cluster snapshots before deleting them. For example, if you set SnapshotRetentionLimit to 5, then a snapshot that was taken today will be retained for 5 days before being deleted. ImportantIf the value of SnapshotRetentionLimit is set to zero (0), backups are turned off.

ccSnapshotWindow :: Lens' CacheCluster (Maybe Text)

The daily time range (in UTC) during which ElastiCache will begin taking a daily snapshot of your cache cluster. Example: 05:00-09:00.

CacheEngineVersion

data CacheEngineVersion

Provides all of the details about a particular cache engine version.

mkCacheEngineVersion :: CacheEngineVersion

Smart constructor for the minimum required fields to construct a valid CacheEngineVersion data type.

CacheEngineVersion is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Engine :: Maybe Text

EngineVersion :: Maybe Text

CacheParameterGroupFamily :: Maybe Text

CacheEngineDescription :: Maybe Text

CacheEngineVersionDescription :: Maybe Text

cevEngine :: Lens' CacheEngineVersion (Maybe Text)

The name of the cache engine.

cevEngineVersion :: Lens' CacheEngineVersion (Maybe Text)

The version number of the cache engine.

cevCacheParameterGroupFamily :: Lens' CacheEngineVersion (Maybe Text)

The name of the cache parameter group family associated with this cache engine.

cevCacheEngineDescription :: Lens' CacheEngineVersion (Maybe Text)

The description of the cache engine.

cevCacheEngineVersionDescription :: Lens' CacheEngineVersion (Maybe Text)

The description of the cache engine version.

CacheNode

data CacheNode

Represents an individual cache node within a cache cluster. Each cache node runs its own instance of the cluster's protocol-compliant caching software - either Memcached or Redis.

mkCacheNode :: CacheNode

Smart constructor for the minimum required fields to construct a valid CacheNode data type to populate a request.

The fields accessible through corresponding lenses are:

CacheNodeId :: Maybe Text

CacheNodeStatus :: Maybe Text

CacheNodeCreateTime :: Maybe ISO8601

Endpoint :: Maybe Endpoint

ParameterGroupStatus :: Maybe Text

SourceCacheNodeId :: Maybe Text

CustomerAvailabilityZone :: Maybe Text

cnCacheNodeId :: Lens' CacheNode (Maybe Text)

The cache node identifier. A node ID is a numeric identifier (0001, 0002, etc.). The combination of cluster ID and node ID uniquely identifies every cache node used in a customer's AWS account.

cnCacheNodeStatus :: Lens' CacheNode (Maybe Text)

The current state of this cache node.

cnCacheNodeCreateTime :: Lens' CacheNode (Maybe ISO8601)

The date and time when the cache node was created.

cnEndpoint :: Lens' CacheNode (Maybe Endpoint)

The hostname and IP address for connecting to this cache node.

cnParameterGroupStatus :: Lens' CacheNode (Maybe Text)

The status of the parameter group applied to this cache node.

cnSourceCacheNodeId :: Lens' CacheNode (Maybe Text)

The ID of the primary node to which this read replica node is synchronized. If this field is empty, then this node is not associated with a primary cache cluster.

cnCustomerAvailabilityZone :: Lens' CacheNode (Maybe Text)

The Availability Zone where this node was created and now resides.

CacheNodeTypeSpecificParameter

data CacheNodeTypeSpecificParameter

A parameter that has a different value for each cache node type it is applied to. For example, in a Redis cache cluster, a cache.m1.large cache node type would have a larger maxmemory value than a cache.m1.small type.

mkCacheNodeTypeSpecificParameter :: CacheNodeTypeSpecificParameter

Smart constructor for the minimum required fields to construct a valid CacheNodeTypeSpecificParameter data type.

CacheNodeTypeSpecificParameter is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

ParameterName :: Maybe Text

Description :: Maybe Text

Source :: Maybe Text

DataType :: Maybe Text

AllowedValues :: Maybe Text

IsModifiable :: Maybe Bool

MinimumEngineVersion :: Maybe Text

CacheNodeTypeSpecificValues :: [CacheNodeTypeSpecificValue]

cntspParameterName :: Lens' CacheNodeTypeSpecificParameter (Maybe Text)

The name of the parameter.

cntspDescription :: Lens' CacheNodeTypeSpecificParameter (Maybe Text)

A description of the parameter.

cntspSource :: Lens' CacheNodeTypeSpecificParameter (Maybe Text)

The source of the parameter value.

cntspDataType :: Lens' CacheNodeTypeSpecificParameter (Maybe Text)

The valid data type for the parameter.

cntspAllowedValues :: Lens' CacheNodeTypeSpecificParameter (Maybe Text)

The valid range of values for the parameter.

cntspIsModifiable :: Lens' CacheNodeTypeSpecificParameter (Maybe Bool)

Indicates whether (true) or not (false) the parameter can be modified. Some parameters have security or operational implications that prevent them from being changed.

cntspMinimumEngineVersion :: Lens' CacheNodeTypeSpecificParameter (Maybe Text)

The earliest cache engine version to which the parameter can apply.

cntspCacheNodeTypeSpecificValues :: Lens' CacheNodeTypeSpecificParameter [CacheNodeTypeSpecificValue]

A list of cache node types and their corresponding values for this parameter.

CacheNodeTypeSpecificValue

data CacheNodeTypeSpecificValue

A value that applies only to a certain cache node type.

mkCacheNodeTypeSpecificValue :: CacheNodeTypeSpecificValue

Smart constructor for the minimum required fields to construct a valid CacheNodeTypeSpecificValue data type to populate a request.

The fields accessible through corresponding lenses are:

CacheNodeType :: Maybe Text

Value :: Maybe Text

cntsvCacheNodeType :: Lens' CacheNodeTypeSpecificValue (Maybe Text)

The cache node type for which this value applies.

cntsvValue :: Lens' CacheNodeTypeSpecificValue (Maybe Text)

The value for the cache node type.

CacheParameterGroup

data CacheParameterGroup

Represents the output of a CreateCacheParameterGroup operation.

mkCacheParameterGroup :: CacheParameterGroup

Smart constructor for the minimum required fields to construct a valid CacheParameterGroup data type.

CacheParameterGroup is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

CacheParameterGroupName :: Maybe Text

CacheParameterGroupFamily :: Maybe Text

Description :: Maybe Text

cpgCacheParameterGroupName :: Lens' CacheParameterGroup (Maybe Text)

The name of the cache parameter group.

cpgCacheParameterGroupFamily :: Lens' CacheParameterGroup (Maybe Text)

The name of the cache parameter group family that this cache parameter group is compatible with.

cpgDescription :: Lens' CacheParameterGroup (Maybe Text)

The description for this cache parameter group.

CacheParameterGroupStatus

data CacheParameterGroupStatus

The status of the cache parameter group.

mkCacheParameterGroupStatus :: CacheParameterGroupStatus

Smart constructor for the minimum required fields to construct a valid CacheParameterGroupStatus data type to populate a request.

The fields accessible through corresponding lenses are:

CacheParameterGroupName :: Maybe Text

ParameterApplyStatus :: Maybe Text

CacheNodeIdsToReboot :: [Text]

cpgsCacheParameterGroupName :: Lens' CacheParameterGroupStatus (Maybe Text)

The name of the cache parameter group.

cpgsParameterApplyStatus :: Lens' CacheParameterGroupStatus (Maybe Text)

The status of parameter updates.

cpgsCacheNodeIdsToReboot :: Lens' CacheParameterGroupStatus [Text]

A list of the cache node IDs which need to be rebooted for parameter changes to be applied. A node ID is a numeric identifier (0001, 0002, etc.).

CacheSecurityGroup

data CacheSecurityGroup

Represents the output of one of the following operations: AuthorizeCacheSecurityGroupIngress CreateCacheSecurityGroup RevokeCacheSecurityGroupIngress.

mkCacheSecurityGroup :: CacheSecurityGroup

Smart constructor for the minimum required fields to construct a valid CacheSecurityGroup data type.

CacheSecurityGroup is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

OwnerId :: Maybe Text

CacheSecurityGroupName :: Maybe Text

Description :: Maybe Text

EC2SecurityGroups :: [EC2SecurityGroup]

csgOwnerId :: Lens' CacheSecurityGroup (Maybe Text)

The AWS account ID of the cache security group owner.

csgCacheSecurityGroupName :: Lens' CacheSecurityGroup (Maybe Text)

The name of the cache security group.

csgDescription :: Lens' CacheSecurityGroup (Maybe Text)

The description of the cache security group.

csgEC2SecurityGroups :: Lens' CacheSecurityGroup [EC2SecurityGroup]

A list of Amazon EC2 security groups that are associated with this cache security group.

CacheSecurityGroupMembership

data CacheSecurityGroupMembership

Represents a cache cluster's status within a particular cache security group.

mkCacheSecurityGroupMembership :: CacheSecurityGroupMembership

Smart constructor for the minimum required fields to construct a valid CacheSecurityGroupMembership data type to populate a request.

The fields accessible through corresponding lenses are:

CacheSecurityGroupName :: Maybe Text

Status :: Maybe Text

csgmCacheSecurityGroupName :: Lens' CacheSecurityGroupMembership (Maybe Text)

The name of the cache security group.

csgmStatus :: Lens' CacheSecurityGroupMembership (Maybe Text)

The membership status in the cache security group. The status changes when a cache security group is modified, or when the cache security groups assigned to a cache cluster are modified.

CacheSubnetGroup

data CacheSubnetGroup

Represents the output of one of the following operations: CreateCacheSubnetGroup ModifyCacheSubnetGroup.

mkCacheSubnetGroup :: CacheSubnetGroup

Smart constructor for the minimum required fields to construct a valid CacheSubnetGroup data type.

CacheSubnetGroup is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

CacheSubnetGroupName :: Maybe Text

CacheSubnetGroupDescription :: Maybe Text

VpcId :: Maybe Text

Subnets :: [Subnet]

csgrCacheSubnetGroupName :: Lens' CacheSubnetGroup (Maybe Text)

The name of the cache subnet group.

csgrCacheSubnetGroupDescription :: Lens' CacheSubnetGroup (Maybe Text)

The description of the cache subnet group.

csgrVpcId :: Lens' CacheSubnetGroup (Maybe Text)

The Amazon Virtual Private Cloud identifier (VPC ID) of the cache subnet group.

csgrSubnets :: Lens' CacheSubnetGroup [Subnet]

A list of subnets associated with the cache subnet group.

EC2SecurityGroup

data EC2SecurityGroup

Provides ownership and status information for an Amazon EC2 security group.

mkEC2SecurityGroup :: EC2SecurityGroup

Smart constructor for the minimum required fields to construct a valid EC2SecurityGroup data type to populate a request.

The fields accessible through corresponding lenses are:

Status :: Maybe Text

EC2SecurityGroupName :: Maybe Text

EC2SecurityGroupOwnerId :: Maybe Text

ecsgStatus :: Lens' EC2SecurityGroup (Maybe Text)

The status of the Amazon EC2 security group.

ecsgEC2SecurityGroupName :: Lens' EC2SecurityGroup (Maybe Text)

The name of the Amazon EC2 security group.

ecsgEC2SecurityGroupOwnerId :: Lens' EC2SecurityGroup (Maybe Text)

The AWS account ID of the Amazon EC2 security group owner.

Endpoint

data Endpoint

Represents the information required for client programs to connect to a cache node.

mkEndpoint :: Endpoint

Smart constructor for the minimum required fields to construct a valid Endpoint data type to populate a request.

The fields accessible through corresponding lenses are:

Address :: Maybe Text

Port :: Maybe Integer

eAddress :: Lens' Endpoint (Maybe Text)

The DNS hostname of the cache node.

ePort :: Lens' Endpoint (Maybe Integer)

The port number that the cache engine is listening on.

EngineDefaults

data EngineDefaults

Represents the output of a DescribeEngineDefaultParameters operation.

mkEngineDefaults :: EngineDefaults

Smart constructor for the minimum required fields to construct a valid EngineDefaults data type.

EngineDefaults is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

CacheParameterGroupFamily :: Maybe Text

Marker :: Maybe Text

Parameters :: [Parameter]

CacheNodeTypeSpecificParameters :: [CacheNodeTypeSpecificParameter]

edCacheParameterGroupFamily :: Lens' EngineDefaults (Maybe Text)

Specifies the name of the cache parameter group family to which the engine default parameters apply.

edMarker :: Lens' EngineDefaults (Maybe Text)

Provides an identifier to allow retrieval of paginated results.

edParameters :: Lens' EngineDefaults [Parameter]

Contains a list of engine default parameters.

edCacheNodeTypeSpecificParameters :: Lens' EngineDefaults [CacheNodeTypeSpecificParameter]

A list of parameters specific to a particular cache node type. Each element in the list contains detailed information about one parameter.

Event

data Event

Represents a single occurrence of something interesting within the system. Some examples of events are creating a cache cluster, adding or removing a cache node, or rebooting a node.

mkEvent :: Event

Smart constructor for the minimum required fields to construct a valid Event data type.

Event is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

SourceIdentifier :: Maybe Text

SourceType :: Maybe SourceType

Message :: Maybe Text

Date :: Maybe ISO8601

erSourceIdentifier :: Lens' Event (Maybe Text)

The identifier for the source of the event. For example, if the event occurred at the cache cluster level, the identifier would be the name of the cache cluster.

erSourceType :: Lens' Event (Maybe SourceType)

Specifies the origin of this event - a cache cluster, a parameter group, a security group, etc.

erMessage :: Lens' Event (Maybe Text)

The text of the event.

erDate :: Lens' Event (Maybe ISO8601)

The date and time when the event occurred.

NodeGroup

data NodeGroup

Represents a collection of cache nodes in a replication group.

mkNodeGroup :: NodeGroup

Smart constructor for the minimum required fields to construct a valid NodeGroup data type to populate a request.

The fields accessible through corresponding lenses are:

NodeGroupId :: Maybe Text

Status :: Maybe Text

PrimaryEndpoint :: Maybe Endpoint

NodeGroupMembers :: [NodeGroupMember]

ngNodeGroupId :: Lens' NodeGroup (Maybe Text)

The identifier for the node group. A replication group contains only one node group; therefore, the node group ID is 0001.

ngStatus :: Lens' NodeGroup (Maybe Text)

The current state of this replication group - creating, available, etc.

ngPrimaryEndpoint :: Lens' NodeGroup (Maybe Endpoint)

Represents the information required for client programs to connect to a cache node.

ngNodeGroupMembers :: Lens' NodeGroup [NodeGroupMember]

A list containing information about individual nodes within the node group.

NodeGroupMember

data NodeGroupMember

Represents a single node within a node group.

mkNodeGroupMember :: NodeGroupMember

Smart constructor for the minimum required fields to construct a valid NodeGroupMember data type to populate a request.

The fields accessible through corresponding lenses are:

CacheClusterId :: Maybe Text

CacheNodeId :: Maybe Text

ReadEndpoint :: Maybe Endpoint

PreferredAvailabilityZone :: Maybe Text

CurrentRole :: Maybe Text

ngmCacheClusterId :: Lens' NodeGroupMember (Maybe Text)

The ID of the cache cluster to which the node belongs.

ngmCacheNodeId :: Lens' NodeGroupMember (Maybe Text)

The ID of the node within its cache cluster. A node ID is a numeric identifier (0001, 0002, etc.).

ngmReadEndpoint :: Lens' NodeGroupMember (Maybe Endpoint)

Represents the information required for client programs to connect to a cache node.

ngmPreferredAvailabilityZone :: Lens' NodeGroupMember (Maybe Text)

The name of the Availability Zone in which the node is located.

ngmCurrentRole :: Lens' NodeGroupMember (Maybe Text)

The role that is currently assigned to the node - primary or replica.

NodeSnapshot

data NodeSnapshot

Represents an individual cache node in a snapshot of a cache cluster.

mkNodeSnapshot :: NodeSnapshot

Smart constructor for the minimum required fields to construct a valid NodeSnapshot data type to populate a request.

The fields accessible through corresponding lenses are:

CacheNodeId :: Maybe Text

CacheSize :: Maybe Text

CacheNodeCreateTime :: Maybe ISO8601

SnapshotCreateTime :: Maybe ISO8601

nsCacheNodeId :: Lens' NodeSnapshot (Maybe Text)

The cache node identifier for the node in the source cache cluster.

nsCacheSize :: Lens' NodeSnapshot (Maybe Text)

The size of the cache on the source cache node.

nsCacheNodeCreateTime :: Lens' NodeSnapshot (Maybe ISO8601)

The date and time when the cache node was created in the source cache cluster.

nsSnapshotCreateTime :: Lens' NodeSnapshot (Maybe ISO8601)

The date and time when the source node's metadata and cache data set was obtained for the snapshot.

NotificationConfiguration

data NotificationConfiguration

Describes a notification topic and its status. Notification topics are used for publishing ElastiCache events to subscribers using Amazon Simple Notification Service (SNS).

mkNotificationConfiguration :: NotificationConfiguration

Smart constructor for the minimum required fields to construct a valid NotificationConfiguration data type to populate a request.

The fields accessible through corresponding lenses are:

TopicArn :: Maybe Text

TopicStatus :: Maybe Text

ncTopicArn :: Lens' NotificationConfiguration (Maybe Text)

The Amazon Resource Name (ARN) that identifies the topic.

ncTopicStatus :: Lens' NotificationConfiguration (Maybe Text)

The current state of the topic.

Parameter

data Parameter

Describes an individual setting that controls some aspect of ElastiCache behavior.

mkParameter :: Parameter

Smart constructor for the minimum required fields to construct a valid Parameter data type.

Parameter is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

ParameterName :: Maybe Text

ParameterValue :: Maybe Text

Description :: Maybe Text

Source :: Maybe Text

DataType :: Maybe Text

AllowedValues :: Maybe Text

IsModifiable :: Maybe Bool

MinimumEngineVersion :: Maybe Text

pParameterName :: Lens' Parameter (Maybe Text)

The name of the parameter.

pParameterValue :: Lens' Parameter (Maybe Text)

The value of the parameter.

pDescription :: Lens' Parameter (Maybe Text)

A description of the parameter.

pSource :: Lens' Parameter (Maybe Text)

The source of the parameter.

pDataType :: Lens' Parameter (Maybe Text)

The valid data type for the parameter.

pAllowedValues :: Lens' Parameter (Maybe Text)

The valid range of values for the parameter.

pIsModifiable :: Lens' Parameter (Maybe Bool)

Indicates whether (true) or not (false) the parameter can be modified. Some parameters have security or operational implications that prevent them from being changed.

pMinimumEngineVersion :: Lens' Parameter (Maybe Text)

The earliest cache engine version to which the parameter can apply.

ParameterNameValue

data ParameterNameValue

Describes a name-value pair that is used to update the value of a parameter.

mkParameterNameValue :: ParameterNameValue

Smart constructor for the minimum required fields to construct a valid ParameterNameValue data type to populate a request.

The fields accessible through corresponding lenses are:

ParameterName :: Maybe Text

ParameterValue :: Maybe Text

pnvParameterName :: Lens' ParameterNameValue (Maybe Text)

The name of the parameter.

pnvParameterValue :: Lens' ParameterNameValue (Maybe Text)

The value of the parameter.

PendingModifiedValues

data PendingModifiedValues

A group of settings that will be applied to the cache cluster in the future, or that are currently being applied.

mkPendingModifiedValues :: PendingModifiedValues

Smart constructor for the minimum required fields to construct a valid PendingModifiedValues data type to populate a request.

The fields accessible through corresponding lenses are:

NumCacheNodes :: Maybe Integer

CacheNodeIdsToRemove :: [Text]

EngineVersion :: Maybe Text

pmvNumCacheNodes :: Lens' PendingModifiedValues (Maybe Integer)

The new number of cache nodes for the cache cluster.

pmvCacheNodeIdsToRemove :: Lens' PendingModifiedValues [Text]

A list of cache node IDs that are being removed (or will be removed) from the cache cluster. A node ID is a numeric identifier (0001, 0002, etc.).

pmvEngineVersion :: Lens' PendingModifiedValues (Maybe Text)

The new cache engine version that the cache cluster will run.

RecurringCharge

data RecurringCharge

Contains the specific price and frequency of a recurring charges for a reserved cache node, or for a reserved cache node offering.

mkRecurringCharge :: RecurringCharge

Smart constructor for the minimum required fields to construct a valid RecurringCharge data type to populate a request.

The fields accessible through corresponding lenses are:

RecurringChargeAmount :: Maybe Double

RecurringChargeFrequency :: Maybe Text

rcRecurringChargeAmount :: Lens' RecurringCharge (Maybe Double)

The monetary amount of the recurring charge.

rcRecurringChargeFrequency :: Lens' RecurringCharge (Maybe Text)

The frequency of the recurring charge.

ReplicationGroup

data ReplicationGroup

Contains all of the attributes of a specific replication group.

mkReplicationGroup :: ReplicationGroup

Smart constructor for the minimum required fields to construct a valid ReplicationGroup data type.

ReplicationGroup is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

ReplicationGroupId :: Maybe Text

Description :: Maybe Text

Status :: Maybe Text

PendingModifiedValues :: Maybe ReplicationGroupPendingModifiedValues

MemberClusters :: [Text]

NodeGroups :: [NodeGroup]

SnapshottingClusterId :: Maybe Text

rgReplicationGroupId :: Lens' ReplicationGroup (Maybe Text)

The identifier for the replication group.

rgDescription :: Lens' ReplicationGroup (Maybe Text)

The description of the replication group.

rgStatus :: Lens' ReplicationGroup (Maybe Text)

The current state of this replication group - creating, available, etc.

rgPendingModifiedValues :: Lens' ReplicationGroup (Maybe ReplicationGroupPendingModifiedValues)

A group of settings to be applied to the replication group, either immediately or during the next maintenance window.

rgMemberClusters :: Lens' ReplicationGroup [Text]

The names of all the cache clusters that are part of this replication group.

rgNodeGroups :: Lens' ReplicationGroup [NodeGroup]

A single element list with information about the nodes in the replication group.

rgSnapshottingClusterId :: Lens' ReplicationGroup (Maybe Text)

The cache cluster ID that is used as the daily snapshot source for the replication group.

ReservedCacheNode

data ReservedCacheNode

Represents the output of a PurchaseReservedCacheNodesOffering operation.

mkReservedCacheNode :: ReservedCacheNode

Smart constructor for the minimum required fields to construct a valid ReservedCacheNode data type.

ReservedCacheNode is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

ReservedCacheNodeId :: Maybe Text

ReservedCacheNodesOfferingId :: Maybe Text

CacheNodeType :: Maybe Text

StartTime :: Maybe ISO8601

Duration :: Maybe Integer

FixedPrice :: Maybe Double

UsagePrice :: Maybe Double

CacheNodeCount :: Maybe Integer

ProductDescription :: Maybe Text

OfferingType :: Maybe Text

State :: Maybe Text

RecurringCharges :: [RecurringCharge]

rcnReservedCacheNodeId :: Lens' ReservedCacheNode (Maybe Text)

The unique identifier for the reservation.

rcnReservedCacheNodesOfferingId :: Lens' ReservedCacheNode (Maybe Text)

The offering identifier.

rcnCacheNodeType :: Lens' ReservedCacheNode (Maybe Text)

The cache node type for the reserved cache nodes.

rcnStartTime :: Lens' ReservedCacheNode (Maybe ISO8601)

The time the reservation started.

rcnDuration :: Lens' ReservedCacheNode (Maybe Integer)

The duration of the reservation in seconds.

rcnFixedPrice :: Lens' ReservedCacheNode (Maybe Double)

The fixed price charged for this reserved cache node.

rcnUsagePrice :: Lens' ReservedCacheNode (Maybe Double)

The hourly price charged for this reserved cache node.

rcnCacheNodeCount :: Lens' ReservedCacheNode (Maybe Integer)

The number of cache nodes that have been reserved.

rcnProductDescription :: Lens' ReservedCacheNode (Maybe Text)

The description of the reserved cache node.

rcnOfferingType :: Lens' ReservedCacheNode (Maybe Text)

The offering type of this reserved cache node.

rcnState :: Lens' ReservedCacheNode (Maybe Text)

The state of the reserved cache node.

rcnRecurringCharges :: Lens' ReservedCacheNode [RecurringCharge]

The recurring price charged to run this reserved cache node.

ReservedCacheNodesOffering

data ReservedCacheNodesOffering

Describes all of the attributes of a reserved cache node offering.

mkReservedCacheNodesOffering :: ReservedCacheNodesOffering

Smart constructor for the minimum required fields to construct a valid ReservedCacheNodesOffering data type.

ReservedCacheNodesOffering is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

ReservedCacheNodesOfferingId :: Maybe Text

CacheNodeType :: Maybe Text

Duration :: Maybe Integer

FixedPrice :: Maybe Double

UsagePrice :: Maybe Double

ProductDescription :: Maybe Text

OfferingType :: Maybe Text

RecurringCharges :: [RecurringCharge]

rcnoReservedCacheNodesOfferingId :: Lens' ReservedCacheNodesOffering (Maybe Text)

A unique identifier for the reserved cache node offering.

rcnoCacheNodeType :: Lens' ReservedCacheNodesOffering (Maybe Text)

The cache node type for the reserved cache node.

rcnoDuration :: Lens' ReservedCacheNodesOffering (Maybe Integer)

The duration of the offering. in seconds.

rcnoFixedPrice :: Lens' ReservedCacheNodesOffering (Maybe Double)

The fixed price charged for this offering.

rcnoUsagePrice :: Lens' ReservedCacheNodesOffering (Maybe Double)

The hourly price charged for this offering.

rcnoProductDescription :: Lens' ReservedCacheNodesOffering (Maybe Text)

The cache engine used by the offering.

rcnoOfferingType :: Lens' ReservedCacheNodesOffering (Maybe Text)

The offering type.

rcnoRecurringCharges :: Lens' ReservedCacheNodesOffering [RecurringCharge]

The recurring price charged to run this reserved cache node.

SecurityGroupMembership

data SecurityGroupMembership

Represents a single cache security group and its status..

mkSecurityGroupMembership :: SecurityGroupMembership

Smart constructor for the minimum required fields to construct a valid SecurityGroupMembership data type to populate a request.

The fields accessible through corresponding lenses are:

SecurityGroupId :: Maybe Text

Status :: Maybe Text

sgmSecurityGroupId :: Lens' SecurityGroupMembership (Maybe Text)

The identifier of the cache security group.

sgmStatus :: Lens' SecurityGroupMembership (Maybe Text)

The status of the cache security group membership. The status changes whenever a cache security group is modified, or when the cache security groups assigned to a cache cluster are modified.

Snapshot

data Snapshot

Represents a copy of an entire cache cluster as of the time when the snapshot was taken.

mkSnapshot :: Snapshot

Smart constructor for the minimum required fields to construct a valid Snapshot data type.

Snapshot is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

SnapshotName :: Maybe Text

CacheClusterId :: Maybe Text

SnapshotStatus :: Maybe Text

SnapshotSource :: Maybe Text

CacheNodeType :: Maybe Text

Engine :: Maybe Text

EngineVersion :: Maybe Text

NumCacheNodes :: Maybe Integer

PreferredAvailabilityZone :: Maybe Text

CacheClusterCreateTime :: Maybe ISO8601

PreferredMaintenanceWindow :: Maybe Text

TopicArn :: Maybe Text

Port :: Maybe Integer

CacheParameterGroupName :: Maybe Text

CacheSubnetGroupName :: Maybe Text

VpcId :: Maybe Text

AutoMinorVersionUpgrade :: Maybe Bool

SnapshotRetentionLimit :: Maybe Integer

SnapshotWindow :: Maybe Text

NodeSnapshots :: [NodeSnapshot]

sSnapshotName :: Lens' Snapshot (Maybe Text)

The name of a snapshot. For an automatic snapshot, the name is system-generated; for a manual snapshot, this is the user-provided name.

sCacheClusterId :: Lens' Snapshot (Maybe Text)

The user-supplied identifier of the source cache cluster.

sSnapshotStatus :: Lens' Snapshot (Maybe Text)

The status of the snapshot. Valid values: creating | available | restoring | copying | deleting.

sSnapshotSource :: Lens' Snapshot (Maybe Text)

Indicates whether the snapshot is from an automatic backup (automated) or was created manually (manual).

sCacheNodeType :: Lens' Snapshot (Maybe Text)

The name of the compute and memory capacity node type for the source cache cluster.

sEngine :: Lens' Snapshot (Maybe Text)

The name of the cache engine (memcached or redis) used by the source cache cluster.

sEngineVersion :: Lens' Snapshot (Maybe Text)

The version of the cache engine version that is used by the source cache cluster.

sNumCacheNodes :: Lens' Snapshot (Maybe Integer)

The number of cache nodes in the source cache cluster.

sPreferredAvailabilityZone :: Lens' Snapshot (Maybe Text)

The name of the Availability Zone in which the source cache cluster is located.

sCacheClusterCreateTime :: Lens' Snapshot (Maybe ISO8601)

The date and time when the source cache cluster was created.

sPreferredMaintenanceWindow :: Lens' Snapshot (Maybe Text)

The time range (in UTC) during which weekly system maintenance can occur on the source cache cluster.

sTopicArn :: Lens' Snapshot (Maybe Text)

The Amazon Resource Name (ARN) for the topic used by the source cache cluster for publishing notifications.

sPort :: Lens' Snapshot (Maybe Integer)

The port number used by each cache nodes in the source cache cluster.

sCacheParameterGroupName :: Lens' Snapshot (Maybe Text)

The cache parameter group that is associated with the source cache cluster.

sCacheSubnetGroupName :: Lens' Snapshot (Maybe Text)

The name of the cache subnet group associated with the source cache cluster.

sVpcId :: Lens' Snapshot (Maybe Text)

The Amazon Virtual Private Cloud identifier (VPC ID) of the cache subnet group for the source cache cluster.

sAutoMinorVersionUpgrade :: Lens' Snapshot (Maybe Bool)

For the source cache cluster, indicates whether minor version patches are applied automatically (true) or not (false).

sSnapshotRetentionLimit :: Lens' Snapshot (Maybe Integer)

For an automatic snapshot, the number of days for which ElastiCache will retain the snapshot before deleting it. For manual snapshots, this field reflects the SnapshotRetentionLimit for the source cache cluster when the snapshot was created. This field is otherwise ignored: Manual snapshots do not expire, and can only be deleted using the DeleteSnapshot action. ImportantIf the value of SnapshotRetentionLimit is set to zero (0), backups are turned off.

sSnapshotWindow :: Lens' Snapshot (Maybe Text)

The daily time range during which ElastiCache takes daily snapshots of the source cache cluster.

sNodeSnapshots :: Lens' Snapshot [NodeSnapshot]

A list of the cache nodes in the source cache cluster.

Subnet

data Subnet

Represents the subnet associated with a cache cluster. This parameter refers to subnets defined in Amazon Virtual Private Cloud (Amazon VPC) and used with ElastiCache.

mkSubnet :: Subnet

Smart constructor for the minimum required fields to construct a valid Subnet data type to populate a request.

The fields accessible through corresponding lenses are:

SubnetIdentifier :: Maybe Text

SubnetAvailabilityZone :: Maybe AvailabilityZone

srSubnetIdentifier :: Lens' Subnet (Maybe Text)

The unique identifier for the subnet.

srSubnetAvailabilityZone :: Lens' Subnet (Maybe AvailabilityZone)

The Availability Zone associated with the subnet.