Network.AWS.ElastiCache.AddTagsToResource

Description

Adds up to 50 cost allocation tags to the named resource. A cost allocation tag is a key-value pair where the key and value are case-sensitive. You can use cost allocation tags to categorize and track your AWS costs.

When you apply tags to your ElastiCache resources, AWS generates a cost allocation report as a comma-separated value (CSV) file with your usage and costs aggregated by your tags. You can apply tags that represent business categories (such as cost centers, application names, or owners) to organize your costs across multiple services. For more information, see Using Cost Allocation Tags in Amazon ElastiCache in the ElastiCache User Guide .

Creating a Request

addTagsToResource

Arguments

:: Text
attrResourceName

-> AddTagsToResource

Creates a value of AddTagsToResource with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

attrResourceName - The Amazon Resource Name (ARN) of the resource to which the tags are to be added, for example arn:aws:elasticache:us-west-2:0123456789:cluster:myCluster or arn:aws:elasticache:us-west-2:0123456789:snapshot:mySnapshot . For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces .

attrTags - A list of cost allocation tags to be added to this resource. A tag is a key-value pair. A tag key must be accompanied by a tag value.

data AddTagsToResource

Represents the input of an AddTagsToResource operation.

See: addTagsToResource smart constructor.

Request Lenses

attrResourceName :: Lens' AddTagsToResource Text

The Amazon Resource Name (ARN) of the resource to which the tags are to be added, for example arn:aws:elasticache:us-west-2:0123456789:cluster:myCluster or arn:aws:elasticache:us-west-2:0123456789:snapshot:mySnapshot . For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces .

attrTags :: Lens' AddTagsToResource [Tag]

A list of cost allocation tags to be added to this resource. A tag is a key-value pair. A tag key must be accompanied by a tag value.

Destructuring the Response

tagListMessage :: TagListMessage

Creates a value of TagListMessage with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tlmTagList - A list of cost allocation tags as key-value pairs.

data TagListMessage

Represents the output from the AddTagsToResource , ListTagsForResource , and RemoveTagsFromResource operations.

See: tagListMessage smart constructor.

Response Lenses

tlmTagList :: Lens' TagListMessage [Tag]

A list of cost allocation tags as key-value pairs.