Network.AWS.ElastiCache.CopySnapshot

Description

Makes a copy of an existing snapshot.

Important: Users or groups that have permissions to use the CopySnapshot operation can create their own Amazon S3 buckets and copy snapshots to it. To control access to your snapshots, use an IAM policy to control who has the ability to use the CopySnapshot operation. For more information about using IAM to control the use of ElastiCache operations, see Exporting Snapshots and Authentication & Access Control .

You could receive the following error messages.

Error Messages

Error Message: The S3 bucket %s is outside of the region.

Solution: Create an Amazon S3 bucket in the same region as your snapshot. For more information, see Step 1: Create an Amazon S3 Bucket in the ElastiCache User Guide.

Error Message: The S3 bucket %s does not exist.

Solution: Create an Amazon S3 bucket in the same region as your snapshot. For more information, see Step 1: Create an Amazon S3 Bucket in the ElastiCache User Guide.

Error Message: The S3 bucket %s is not owned by the authenticated user.

Solution: Create an Amazon S3 bucket in the same region as your snapshot. For more information, see Step 1: Create an Amazon S3 Bucket in the ElastiCache User Guide.

Error Message: The authenticated user does not have sufficient permissions to perform the desired activity.

Solution: Contact your system administrator to get the needed permissions.

Error Message: The S3 bucket %s already contains an object with key %s.

Solution: Give the TargetSnapshotName a new and unique value. If exporting a snapshot, you could alternatively create a new Amazon S3 bucket and use this same value for TargetSnapshotName .

Error Message: ElastiCache has not been granted READ permissions %s on the S3 Bucket.

Solution: Add List and Read permissions on the bucket. For more information, see Step 2: Grant ElastiCache Access to Your Amazon S3 Bucket in the ElastiCache User Guide.

Error Message: ElastiCache has not been granted WRITE permissions %s on the S3 Bucket.

Solution: Add UploadDelete permissions on the bucket. For more information, see <http:docs.aws.amazon.comAmazonElastiCachelatestUserGuide/Snapshots.Exporting.html#Snapshots.Exporting.GrantAccess Step 2: Grant ElastiCache Access to Your Amazon S3 Bucket> in the ElastiCache User Guide.

Error Message: ElastiCache has not been granted READ_ACP permissions %s on the S3 Bucket.

Solution: Add View Permissions on the bucket. For more information, see Step 2: Grant ElastiCache Access to Your Amazon S3 Bucket in the ElastiCache User Guide.

Creating a Request

copySnapshot

Arguments

:: Text
csSourceSnapshotName

-> Text
csTargetSnapshotName

-> CopySnapshot

Creates a value of CopySnapshot with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

csTargetBucket - The Amazon S3 bucket to which the snapshot is exported. This parameter is used only when exporting a snapshot for external access. When using this parameter to export a snapshot, be sure Amazon ElastiCache has the needed permissions to this S3 bucket. For more information, see Step 2: Grant ElastiCache Access to Your Amazon S3 Bucket in the Amazon ElastiCache User Guide . For more information, see Exporting a Snapshot in the Amazon ElastiCache User Guide .

csSourceSnapshotName - The name of an existing snapshot from which to make a copy.

csTargetSnapshotName - A name for the snapshot copy. ElastiCache does not permit overwriting a snapshot, therefore this name must be unique within its context - ElastiCache or an Amazon S3 bucket if exporting.

data CopySnapshot

Represents the input of a CopySnapshotMessage operation.

See: copySnapshot smart constructor.

Request Lenses

csTargetBucket :: Lens' CopySnapshot (Maybe Text)

The Amazon S3 bucket to which the snapshot is exported. This parameter is used only when exporting a snapshot for external access. When using this parameter to export a snapshot, be sure Amazon ElastiCache has the needed permissions to this S3 bucket. For more information, see Step 2: Grant ElastiCache Access to Your Amazon S3 Bucket in the Amazon ElastiCache User Guide . For more information, see Exporting a Snapshot in the Amazon ElastiCache User Guide .

csSourceSnapshotName :: Lens' CopySnapshot Text

The name of an existing snapshot from which to make a copy.

csTargetSnapshotName :: Lens' CopySnapshot Text

A name for the snapshot copy. ElastiCache does not permit overwriting a snapshot, therefore this name must be unique within its context - ElastiCache or an Amazon S3 bucket if exporting.

Destructuring the Response

copySnapshotResponse

Arguments

:: Int
csrsResponseStatus

-> CopySnapshotResponse

Creates a value of CopySnapshotResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

csrsSnapshot - Undocumented member.

csrsResponseStatus - -- | The response status code.

data CopySnapshotResponse

See: copySnapshotResponse smart constructor.

Response Lenses

csrsSnapshot :: Lens' CopySnapshotResponse (Maybe Snapshot)

Undocumented member.

csrsResponseStatus :: Lens' CopySnapshotResponse Int

- | The response status code.