Network.AWS.ElastiCache.CopySnapshot

Description

The CopySnapshot action makes a copy of an existing snapshot.

http://docs.aws.amazon.com/AmazonElastiCache/latest/APIReference/API_CopySnapshot.html

Request

data CopySnapshot

Request constructor

copySnapshot

Arguments

:: Text
csSourceSnapshotName

-> Text
csTargetSnapshotName

-> CopySnapshot

CopySnapshot constructor.

The fields accessible through corresponding lenses are:

csSourceSnapshotName :: Text

csTargetSnapshotName :: Text

Request lenses

csSourceSnapshotName :: Lens' CopySnapshot Text

The name of an existing snapshot from which to copy.

csTargetSnapshotName :: Lens' CopySnapshot Text

A name for the copied snapshot.

Response

data CopySnapshotResponse

Response constructor

copySnapshotResponse :: CopySnapshotResponse

CopySnapshotResponse constructor.

The fields accessible through corresponding lenses are:

csrSnapshot :: Maybe Snapshot

Response lenses

csrSnapshot :: Lens' CopySnapshotResponse (Maybe Snapshot)