Network.AWS.ElastiCache.CreateCacheCluster

Description

Creates a cache cluster. All nodes in the cache cluster run the same protocol-compliant cache engine software, either Memcached or Redis.

Important: Due to current limitations on Redis (cluster mode disabled), this operation or parameter is not supported on Redis (cluster mode enabled) replication groups.

Creating a Request

createCacheCluster

Arguments

:: Text
cccCacheClusterId

-> CreateCacheCluster

Creates a value of CreateCacheCluster with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cccEngineVersion - The version number of the cache engine to be used for this cache cluster. To view the supported cache engine versions, use the DescribeCacheEngineVersions operation. Important: You can upgrade to a newer engine version (see Selecting a Cache Engine and Version ), but you cannot downgrade to an earlier engine version. If you want to use an earlier engine version, you must delete the existing cache cluster or replication group and create it anew with the earlier engine version.

cccCacheNodeType - The compute and memory capacity of the nodes in the node group (shard). Valid node types are as follows: * General purpose: * Current generation: cache.t2.micro , cache.t2.small , cache.t2.medium , cache.m3.medium , cache.m3.large , cache.m3.xlarge , cache.m3.2xlarge , cache.m4.large , cache.m4.xlarge , cache.m4.2xlarge , cache.m4.4xlarge , cache.m4.10xlarge * Previous generation: cache.t1.micro , cache.m1.small , cache.m1.medium , cache.m1.large , cache.m1.xlarge * Compute optimized: cache.c1.xlarge * Memory optimized: * Current generation: cache.r3.large , cache.r3.xlarge , cache.r3.2xlarge , cache.r3.4xlarge , cache.r3.8xlarge * Previous generation: cache.m2.xlarge , cache.m2.2xlarge , cache.m2.4xlarge Notes: * All T2 instances are created in an Amazon Virtual Private Cloud (Amazon VPC). * Redis backuprestore is not supported for Redis (cluster mode disabled) T1 and T2 instances. Backuprestore is supported on Redis (cluster mode enabled) T2 instances. * Redis Append-only files (AOF) functionality is not supported for T1 or T2 instances. For a complete listing of node types and specifications, see Amazon ElastiCache Product Features and Details and either Cache Node Type-Specific Parameters for Memcached or Cache Node Type-Specific Parameters for Redis .

cccSecurityGroupIds - One or more VPC security groups associated with the cache cluster. Use this parameter only when you are creating a cache cluster in an Amazon Virtual Private Cloud (Amazon VPC).

cccSnapshotARNs - A single-element string list containing an Amazon Resource Name (ARN) that uniquely identifies a Redis RDB snapshot file stored in Amazon S3. The snapshot file is used to populate the node group (shard). The Amazon S3 object name in the ARN cannot contain any commas. Example of an Amazon S3 ARN: arn:aws:s3:::my_bucket/snapshot1.rdb

cccAutoMinorVersionUpgrade - This parameter is currently disabled.

cccCacheParameterGroupName - The name of the parameter group to associate with this cache cluster. If this argument is omitted, the default parameter group for the specified engine is used. You cannot use any parameter group which has cluster-enabled=yes when creating a cluster.

cccSnapshotWindow - The daily time range (in UTC) during which ElastiCache begins taking a daily snapshot of your node group (shard). Example: 05:00-09:00 If you do not specify this parameter, ElastiCache automatically chooses an appropriate time range. Note: This parameter is only valid if the Engine parameter is redis .

cccAuthToken - Reserved parameter. The password used to access a password protected server. Password constraints: * Must be only printable ASCII characters. * Must be at least 16 characters and no more than 128 characters in length. * Cannot contain any of the following characters: /, '"', or "@". For more information, see AUTH password at Redis.

cccEngine - The name of the cache engine to be used for this cache cluster. Valid values for this parameter are: memcached | redis

cccPreferredAvailabilityZones - A list of the Availability Zones in which cache nodes are created. The order of the zones in the list is not important. This option is only supported on Memcached. If you want all the nodes in the same Availability Zone, use PreferredAvailabilityZone instead, or repeat the Availability Zone multiple times in the list. Default: System chosen Availability Zones.

cccPreferredMaintenanceWindow - Specifies the weekly time range during which maintenance on the cache cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period. Valid values for ddd are: Specifies the weekly time range during which maintenance on the cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period. Valid values for ddd are: * sun * mon * tue * wed * thu * fri * sat Example: sun:23:00-mon:01:30

cccCacheSubnetGroupName - The name of the subnet group to be used for the cache cluster. Use this parameter only when you are creating a cache cluster in an Amazon Virtual Private Cloud (Amazon VPC). Important: If you're going to launch your cluster in an Amazon VPC, you need to create a subnet group before you start creating a cluster. For more information, see Subnets and Subnet Groups .

cccPreferredAvailabilityZone - The EC2 Availability Zone in which the cache cluster is created. All nodes belonging to this Memcached cache cluster are placed in the preferred Availability Zone. If you want to create your nodes across multiple Availability Zones, use PreferredAvailabilityZones . Default: System chosen Availability Zone.

cccSnapshotRetentionLimit - The number of days for which ElastiCache retains automatic snapshots before deleting them. For example, if you set SnapshotRetentionLimit to 5, a snapshot taken today is retained for 5 days before being deleted. Default: 0 (i.e., automatic backups are disabled for this cache cluster).

cccAZMode - Specifies whether the nodes in this Memcached cluster are created in a single Availability Zone or created across multiple Availability Zones in the cluster's region. This parameter is only supported for Memcached cache clusters. If the AZMode and PreferredAvailabilityZones are not specified, ElastiCache assumes single-az mode.

cccSnapshotName - The name of a Redis snapshot from which to restore data into the new node group (shard). The snapshot status changes to restoring while the new node group (shard) is being created.

cccReplicationGroupId - Important: Due to current limitations on Redis (cluster mode disabled), this operation or parameter is not supported on Redis (cluster mode enabled) replication groups. The ID of the replication group to which this cache cluster should belong. If this parameter is specified, the cache cluster is added to the specified replication group as a read replica; otherwise, the cache cluster is a standalone primary that is not part of any replication group. If the specified replication group is Multi-AZ enabled and the Availability Zone is not specified, the cache cluster is created in Availability Zones that provide the best spread of read replicas across Availability Zones.

cccNotificationTopicARN - The Amazon Resource Name (ARN) of the Amazon Simple Notification Service (SNS) topic to which notifications are sent.

cccNumCacheNodes - The initial number of cache nodes that the cache cluster has. For clusters running Redis, this value must be 1. For clusters running Memcached, this value must be between 1 and 20. If you need more than 20 nodes for your Memcached cluster, please fill out the ElastiCache Limit Increase Request form at http://aws.amazon.com/contact-us/elasticache-node-limit-request/ .

cccTags - A list of cost allocation tags to be added to this resource. A tag is a key-value pair. A tag key must be accompanied by a tag value.

cccPort - The port number on which each of the cache nodes accepts connections.

cccCacheSecurityGroupNames - A list of security group names to associate with this cache cluster. Use this parameter only when you are creating a cache cluster outside of an Amazon Virtual Private Cloud (Amazon VPC).

cccCacheClusterId - The node group (shard) identifier. This parameter is stored as a lowercase string. Constraints: * A name must contain from 1 to 20 alphanumeric characters or hyphens. * The first character must be a letter. * A name cannot end with a hyphen or contain two consecutive hyphens.

data CreateCacheCluster

Represents the input of a CreateCacheCluster operation.

See: createCacheCluster smart constructor.

Request Lenses

cccEngineVersion :: Lens' CreateCacheCluster (Maybe Text)

The version number of the cache engine to be used for this cache cluster. To view the supported cache engine versions, use the DescribeCacheEngineVersions operation. Important: You can upgrade to a newer engine version (see Selecting a Cache Engine and Version ), but you cannot downgrade to an earlier engine version. If you want to use an earlier engine version, you must delete the existing cache cluster or replication group and create it anew with the earlier engine version.

cccCacheNodeType :: Lens' CreateCacheCluster (Maybe Text)

The compute and memory capacity of the nodes in the node group (shard). Valid node types are as follows: * General purpose: * Current generation: cache.t2.micro , cache.t2.small , cache.t2.medium , cache.m3.medium , cache.m3.large , cache.m3.xlarge , cache.m3.2xlarge , cache.m4.large , cache.m4.xlarge , cache.m4.2xlarge , cache.m4.4xlarge , cache.m4.10xlarge * Previous generation: cache.t1.micro , cache.m1.small , cache.m1.medium , cache.m1.large , cache.m1.xlarge * Compute optimized: cache.c1.xlarge * Memory optimized: * Current generation: cache.r3.large , cache.r3.xlarge , cache.r3.2xlarge , cache.r3.4xlarge , cache.r3.8xlarge * Previous generation: cache.m2.xlarge , cache.m2.2xlarge , cache.m2.4xlarge Notes: * All T2 instances are created in an Amazon Virtual Private Cloud (Amazon VPC). * Redis backuprestore is not supported for Redis (cluster mode disabled) T1 and T2 instances. Backuprestore is supported on Redis (cluster mode enabled) T2 instances. * Redis Append-only files (AOF) functionality is not supported for T1 or T2 instances. For a complete listing of node types and specifications, see Amazon ElastiCache Product Features and Details and either Cache Node Type-Specific Parameters for Memcached or Cache Node Type-Specific Parameters for Redis .

cccSecurityGroupIds :: Lens' CreateCacheCluster [Text]

One or more VPC security groups associated with the cache cluster. Use this parameter only when you are creating a cache cluster in an Amazon Virtual Private Cloud (Amazon VPC).

cccSnapshotARNs :: Lens' CreateCacheCluster [Text]

A single-element string list containing an Amazon Resource Name (ARN) that uniquely identifies a Redis RDB snapshot file stored in Amazon S3. The snapshot file is used to populate the node group (shard). The Amazon S3 object name in the ARN cannot contain any commas. Example of an Amazon S3 ARN: arn:aws:s3:::my_bucket/snapshot1.rdb

cccAutoMinorVersionUpgrade :: Lens' CreateCacheCluster (Maybe Bool)

This parameter is currently disabled.

cccCacheParameterGroupName :: Lens' CreateCacheCluster (Maybe Text)

The name of the parameter group to associate with this cache cluster. If this argument is omitted, the default parameter group for the specified engine is used. You cannot use any parameter group which has cluster-enabled=yes when creating a cluster.

cccSnapshotWindow :: Lens' CreateCacheCluster (Maybe Text)

The daily time range (in UTC) during which ElastiCache begins taking a daily snapshot of your node group (shard). Example: 05:00-09:00 If you do not specify this parameter, ElastiCache automatically chooses an appropriate time range. Note: This parameter is only valid if the Engine parameter is redis .

cccAuthToken :: Lens' CreateCacheCluster (Maybe Text)

Reserved parameter. The password used to access a password protected server. Password constraints: * Must be only printable ASCII characters. * Must be at least 16 characters and no more than 128 characters in length. * Cannot contain any of the following characters: /, '"', or "@". For more information, see AUTH password at Redis.

cccEngine :: Lens' CreateCacheCluster (Maybe Text)

The name of the cache engine to be used for this cache cluster. Valid values for this parameter are: memcached | redis

cccPreferredAvailabilityZones :: Lens' CreateCacheCluster [Text]

A list of the Availability Zones in which cache nodes are created. The order of the zones in the list is not important. This option is only supported on Memcached. If you want all the nodes in the same Availability Zone, use PreferredAvailabilityZone instead, or repeat the Availability Zone multiple times in the list. Default: System chosen Availability Zones.

cccPreferredMaintenanceWindow :: Lens' CreateCacheCluster (Maybe Text)

Specifies the weekly time range during which maintenance on the cache cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period. Valid values for ddd are: Specifies the weekly time range during which maintenance on the cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period. Valid values for ddd are: * sun * mon * tue * wed * thu * fri * sat Example: sun:23:00-mon:01:30

cccCacheSubnetGroupName :: Lens' CreateCacheCluster (Maybe Text)

The name of the subnet group to be used for the cache cluster. Use this parameter only when you are creating a cache cluster in an Amazon Virtual Private Cloud (Amazon VPC). Important: If you're going to launch your cluster in an Amazon VPC, you need to create a subnet group before you start creating a cluster. For more information, see Subnets and Subnet Groups .

cccPreferredAvailabilityZone :: Lens' CreateCacheCluster (Maybe Text)

The EC2 Availability Zone in which the cache cluster is created. All nodes belonging to this Memcached cache cluster are placed in the preferred Availability Zone. If you want to create your nodes across multiple Availability Zones, use PreferredAvailabilityZones . Default: System chosen Availability Zone.

cccSnapshotRetentionLimit :: Lens' CreateCacheCluster (Maybe Int)

The number of days for which ElastiCache retains automatic snapshots before deleting them. For example, if you set SnapshotRetentionLimit to 5, a snapshot taken today is retained for 5 days before being deleted. Default: 0 (i.e., automatic backups are disabled for this cache cluster).

cccAZMode :: Lens' CreateCacheCluster (Maybe AZMode)

Specifies whether the nodes in this Memcached cluster are created in a single Availability Zone or created across multiple Availability Zones in the cluster's region. This parameter is only supported for Memcached cache clusters. If the AZMode and PreferredAvailabilityZones are not specified, ElastiCache assumes single-az mode.

cccSnapshotName :: Lens' CreateCacheCluster (Maybe Text)

The name of a Redis snapshot from which to restore data into the new node group (shard). The snapshot status changes to restoring while the new node group (shard) is being created.

cccReplicationGroupId :: Lens' CreateCacheCluster (Maybe Text)

Important: Due to current limitations on Redis (cluster mode disabled), this operation or parameter is not supported on Redis (cluster mode enabled) replication groups. The ID of the replication group to which this cache cluster should belong. If this parameter is specified, the cache cluster is added to the specified replication group as a read replica; otherwise, the cache cluster is a standalone primary that is not part of any replication group. If the specified replication group is Multi-AZ enabled and the Availability Zone is not specified, the cache cluster is created in Availability Zones that provide the best spread of read replicas across Availability Zones.

cccNotificationTopicARN :: Lens' CreateCacheCluster (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon Simple Notification Service (SNS) topic to which notifications are sent.

cccNumCacheNodes :: Lens' CreateCacheCluster (Maybe Int)

The initial number of cache nodes that the cache cluster has. For clusters running Redis, this value must be 1. For clusters running Memcached, this value must be between 1 and 20. If you need more than 20 nodes for your Memcached cluster, please fill out the ElastiCache Limit Increase Request form at http://aws.amazon.com/contact-us/elasticache-node-limit-request/ .

cccTags :: Lens' CreateCacheCluster [Tag]

A list of cost allocation tags to be added to this resource. A tag is a key-value pair. A tag key must be accompanied by a tag value.

cccPort :: Lens' CreateCacheCluster (Maybe Int)

The port number on which each of the cache nodes accepts connections.

cccCacheSecurityGroupNames :: Lens' CreateCacheCluster [Text]

A list of security group names to associate with this cache cluster. Use this parameter only when you are creating a cache cluster outside of an Amazon Virtual Private Cloud (Amazon VPC).

cccCacheClusterId :: Lens' CreateCacheCluster Text

The node group (shard) identifier. This parameter is stored as a lowercase string. Constraints: * A name must contain from 1 to 20 alphanumeric characters or hyphens. * The first character must be a letter. * A name cannot end with a hyphen or contain two consecutive hyphens.

Destructuring the Response

createCacheClusterResponse

Arguments

:: Int
cccrsResponseStatus

-> CreateCacheClusterResponse

Creates a value of CreateCacheClusterResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cccrsCacheCluster - Undocumented member.

cccrsResponseStatus - -- | The response status code.

data CreateCacheClusterResponse

See: createCacheClusterResponse smart constructor.

Response Lenses

cccrsCacheCluster :: Lens' CreateCacheClusterResponse (Maybe CacheCluster)

Undocumented member.

cccrsResponseStatus :: Lens' CreateCacheClusterResponse Int

- | The response status code.