Network.AWS.ElastiCache.CreateCacheParameterGroup

Description

Creates a new Amazon ElastiCache cache parameter group. An ElastiCache cache parameter group is a collection of parameters and their values that are applied to all of the nodes in any cache cluster or replication group using the CacheParameterGroup.

A newly created CacheParameterGroup is an exact duplicate of the default parameter group for the CacheParameterGroupFamily. To customize the newly created CacheParameterGroup you can change the values of specific parameters. For more information, see:

ModifyCacheParameterGroup in the ElastiCache API Reference.

Parameters and Parameter Groups in the ElastiCache User Guide.

Creating a Request

createCacheParameterGroup

Arguments

:: Text
ccpgCacheParameterGroupName

-> Text
ccpgCacheParameterGroupFamily

-> Text
ccpgDescription

-> CreateCacheParameterGroup

Creates a value of CreateCacheParameterGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ccpgCacheParameterGroupName - A user-specified name for the cache parameter group.

ccpgCacheParameterGroupFamily - The name of the cache parameter group family that the cache parameter group can be used with. Valid values are: memcached1.4 | redis2.6 | redis2.8 | redis3.2

ccpgDescription - A user-specified description for the cache parameter group.

data CreateCacheParameterGroup

Represents the input of a CreateCacheParameterGroup operation.

See: createCacheParameterGroup smart constructor.

Request Lenses

ccpgCacheParameterGroupName :: Lens' CreateCacheParameterGroup Text

A user-specified name for the cache parameter group.

ccpgCacheParameterGroupFamily :: Lens' CreateCacheParameterGroup Text

The name of the cache parameter group family that the cache parameter group can be used with. Valid values are: memcached1.4 | redis2.6 | redis2.8 | redis3.2

ccpgDescription :: Lens' CreateCacheParameterGroup Text

A user-specified description for the cache parameter group.

Destructuring the Response

createCacheParameterGroupResponse

Arguments

:: Int
ccpgrsResponseStatus

-> CreateCacheParameterGroupResponse

Creates a value of CreateCacheParameterGroupResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ccpgrsCacheParameterGroup - Undocumented member.

ccpgrsResponseStatus - -- | The response status code.

data CreateCacheParameterGroupResponse

See: createCacheParameterGroupResponse smart constructor.

Response Lenses

ccpgrsCacheParameterGroup :: Lens' CreateCacheParameterGroupResponse (Maybe CacheParameterGroup)

Undocumented member.

ccpgrsResponseStatus :: Lens' CreateCacheParameterGroupResponse Int

- | The response status code.