Network.AWS.ElastiCache.CreateCacheSecurityGroup

Description

Creates a new cache security group. Use a cache security group to control access to one or more cache clusters.

Cache security groups are only used when you are creating a cache cluster outside of an Amazon Virtual Private Cloud (Amazon VPC). If you are creating a cache cluster inside of a VPC, use a cache subnet group instead. For more information, see CreateCacheSubnetGroup .

Creating a Request

createCacheSecurityGroup

Arguments

:: Text
ccsgCacheSecurityGroupName

-> Text
ccsgDescription

-> CreateCacheSecurityGroup

Creates a value of CreateCacheSecurityGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ccsgCacheSecurityGroupName - A name for the cache security group. This value is stored as a lowercase string. Constraints: Must contain no more than 255 alphanumeric characters. Cannot be the word Default. Example: mysecuritygroup

ccsgDescription - A description for the cache security group.

data CreateCacheSecurityGroup

Represents the input of a CreateCacheSecurityGroup operation.

See: createCacheSecurityGroup smart constructor.

Request Lenses

ccsgCacheSecurityGroupName :: Lens' CreateCacheSecurityGroup Text

A name for the cache security group. This value is stored as a lowercase string. Constraints: Must contain no more than 255 alphanumeric characters. Cannot be the word Default. Example: mysecuritygroup

ccsgDescription :: Lens' CreateCacheSecurityGroup Text

A description for the cache security group.

Destructuring the Response

createCacheSecurityGroupResponse

Arguments

:: Int
ccsgrsResponseStatus

-> CreateCacheSecurityGroupResponse

Creates a value of CreateCacheSecurityGroupResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ccsgrsCacheSecurityGroup - Undocumented member.

ccsgrsResponseStatus - -- | The response status code.

data CreateCacheSecurityGroupResponse

See: createCacheSecurityGroupResponse smart constructor.

Response Lenses

ccsgrsCacheSecurityGroup :: Lens' CreateCacheSecurityGroupResponse (Maybe CacheSecurityGroup)

Undocumented member.

ccsgrsResponseStatus :: Lens' CreateCacheSecurityGroupResponse Int

- | The response status code.