Network.AWS.ElastiCache.CreateCacheSecurityGroup

Description

The CreateCacheSecurityGroup action creates a new cache security group. Use a cache security group to control access to one or more cache clusters.

Cache security groups are only used when you are creating a cache cluster outside of an Amazon Virtual Private Cloud (VPC). If you are creating a cache cluster inside of a VPC, use a cache subnet group instead. For more information, see CreateCacheSubnetGroup.

http://docs.aws.amazon.com/AmazonElastiCache/latest/APIReference/API_CreateCacheSecurityGroup.html

Request

data CreateCacheSecurityGroup

Request constructor

createCacheSecurityGroup

Arguments

:: Text
ccsgCacheSecurityGroupName

-> Text
ccsgDescription

-> CreateCacheSecurityGroup

CreateCacheSecurityGroup constructor.

The fields accessible through corresponding lenses are:

ccsgCacheSecurityGroupName :: Text

ccsgDescription :: Text

Request lenses

ccsgCacheSecurityGroupName :: Lens' CreateCacheSecurityGroup Text

A name for the cache security group. This value is stored as a lowercase string.

Constraints: Must contain no more than 255 alphanumeric characters. Cannot be the word Default.

Example: mysecuritygroup

ccsgDescription :: Lens' CreateCacheSecurityGroup Text

A description for the cache security group.

Response

data CreateCacheSecurityGroupResponse

Response constructor

createCacheSecurityGroupResponse :: CreateCacheSecurityGroupResponse

CreateCacheSecurityGroupResponse constructor.

The fields accessible through corresponding lenses are:

ccsgrCacheSecurityGroup :: Maybe CacheSecurityGroup

Response lenses

ccsgrCacheSecurityGroup :: Lens' CreateCacheSecurityGroupResponse (Maybe CacheSecurityGroup)