Network.AWS.ElastiCache.CreateSnapshot

Description

Creates a copy of an entire cache cluster or replication group at a specific moment in time.

Creating a Request

createSnapshot

Arguments

:: Text
csSnapshotName

-> CreateSnapshot

Creates a value of CreateSnapshot with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

csCacheClusterId - The identifier of an existing cache cluster. The snapshot is created from this cache cluster.

csReplicationGroupId - The identifier of an existing replication group. The snapshot is created from this replication group.

csSnapshotName - A name for the snapshot being created.

data CreateSnapshot

Represents the input of a CreateSnapshot operation.

See: createSnapshot smart constructor.

Request Lenses

csCacheClusterId :: Lens' CreateSnapshot (Maybe Text)

The identifier of an existing cache cluster. The snapshot is created from this cache cluster.

csReplicationGroupId :: Lens' CreateSnapshot (Maybe Text)

The identifier of an existing replication group. The snapshot is created from this replication group.

csSnapshotName :: Lens' CreateSnapshot Text

A name for the snapshot being created.

Destructuring the Response

createSnapshotResponse

Arguments

:: Int
crersResponseStatus

-> CreateSnapshotResponse

Creates a value of CreateSnapshotResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

crersSnapshot - Undocumented member.

crersResponseStatus - -- | The response status code.

data CreateSnapshotResponse

See: createSnapshotResponse smart constructor.

Response Lenses

crersSnapshot :: Lens' CreateSnapshotResponse (Maybe Snapshot)

Undocumented member.

crersResponseStatus :: Lens' CreateSnapshotResponse Int

- | The response status code.