Network.AWS.ElastiCache.DeleteCacheSecurityGroup

Description

Deletes a cache security group.

Creating a Request

deleteCacheSecurityGroup

Arguments

:: Text
dcsgCacheSecurityGroupName

-> DeleteCacheSecurityGroup

Creates a value of DeleteCacheSecurityGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcsgCacheSecurityGroupName - The name of the cache security group to delete.

data DeleteCacheSecurityGroup

Represents the input of a DeleteCacheSecurityGroup operation.

See: deleteCacheSecurityGroup smart constructor.

Request Lenses

dcsgCacheSecurityGroupName :: Lens' DeleteCacheSecurityGroup Text

The name of the cache security group to delete.

Destructuring the Response

deleteCacheSecurityGroupResponse :: DeleteCacheSecurityGroupResponse

Creates a value of DeleteCacheSecurityGroupResponse with the minimum fields required to make a request.

data DeleteCacheSecurityGroupResponse

See: deleteCacheSecurityGroupResponse smart constructor.