Network.AWS.ElastiCache.DeleteCacheSecurityGroup

Description

The DeleteCacheSecurityGroup action deletes a cache security group.

You cannot delete a cache security group if it is associated with any cache clusters.

http://docs.aws.amazon.com/AmazonElastiCache/latest/APIReference/API_DeleteCacheSecurityGroup.html

Request

data DeleteCacheSecurityGroup

Request constructor

deleteCacheSecurityGroup

Arguments

:: Text
dcsgCacheSecurityGroupName

-> DeleteCacheSecurityGroup

DeleteCacheSecurityGroup constructor.

The fields accessible through corresponding lenses are:

dcsgCacheSecurityGroupName :: Text

Request lenses

dcsgCacheSecurityGroupName :: Lens' DeleteCacheSecurityGroup Text

The name of the cache security group to delete.

You cannot delete the default security group.

Response

data DeleteCacheSecurityGroupResponse

Response constructor

deleteCacheSecurityGroupResponse :: DeleteCacheSecurityGroupResponse

DeleteCacheSecurityGroupResponse constructor.