Network.AWS.ElastiCache.DeleteReplicationGroup

Description

The DeleteReplicationGroup action deletes an existing replication group. By default, this action deletes the entire replication group, including the primary cluster and all of the read replicas. You can optionally delete only the read replicas, while retaining the primary cluster.

When you receive a successful response from this action, Amazon ElastiCache immediately begins deleting the selected resources; you cannot cancel or revert this action.

http://docs.aws.amazon.com/AmazonElastiCache/latest/APIReference/API_DeleteReplicationGroup.html

Request

data DeleteReplicationGroup

Request constructor

deleteReplicationGroup

Arguments

:: Text
drgReplicationGroupId

-> DeleteReplicationGroup

DeleteReplicationGroup constructor.

The fields accessible through corresponding lenses are:

drgFinalSnapshotIdentifier :: Maybe Text

drgReplicationGroupId :: Text

drgRetainPrimaryCluster :: Maybe Bool

Request lenses

drgFinalSnapshotIdentifier :: Lens' DeleteReplicationGroup (Maybe Text)

The name of a final node group snapshot. ElastiCache creates the snapshot from the primary node in the cluster, rather than one of the replicas; this is to ensure that it captures the freshest data. After the final snapshot is taken, the cluster is immediately deleted.

drgReplicationGroupId :: Lens' DeleteReplicationGroup Text

The identifier for the cluster to be deleted. This parameter is not case sensitive.

drgRetainPrimaryCluster :: Lens' DeleteReplicationGroup (Maybe Bool)

If set to true, all of the read replicas will be deleted, but the primary node will be retained.

Response

data DeleteReplicationGroupResponse

Response constructor

deleteReplicationGroupResponse :: DeleteReplicationGroupResponse

DeleteReplicationGroupResponse constructor.

The fields accessible through corresponding lenses are:

drgrReplicationGroup :: Maybe ReplicationGroup

Response lenses

drgrReplicationGroup :: Lens' DeleteReplicationGroupResponse (Maybe ReplicationGroup)