Network.AWS.ElastiCache.DescribeCacheClusters

Description

Returns information about all provisioned cache clusters if no cache cluster identifier is specified, or about a specific cache cluster if a cache cluster identifier is supplied.

By default, abbreviated information about the cache clusters is returned. You can use the optional ShowCacheNodeInfo flag to retrieve detailed information about the cache nodes associated with the cache clusters. These details include the DNS address and port for the cache node endpoint.

If the cluster is in the creating state, only cluster-level information is displayed until all of the nodes are successfully provisioned.

If the cluster is in the deleting state, only cluster-level information is displayed.

If cache nodes are currently being added to the cache cluster, node endpoint information and creation time for the additional nodes are not displayed until they are completely provisioned. When the cache cluster state is available , the cluster is ready for use.

If cache nodes are currently being removed from the cache cluster, no endpoint information for the removed nodes is displayed.

This operation returns paginated results.

Creating a Request

describeCacheClusters :: DescribeCacheClusters

Creates a value of DescribeCacheClusters with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dShowCacheClustersNotInReplicationGroups - An optional flag that can be included in the DescribeCacheCluster request to show only nodes (API/CLI: clusters) that are not members of a replication group. In practice, this mean Memcached and single node Redis clusters.

dCacheClusterId - The user-supplied cluster identifier. If this parameter is specified, only information about that specific cache cluster is returned. This parameter isn't case sensitive.

dMarker - An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

dMaxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: minimum 20; maximum 100.

dShowCacheNodeInfo - An optional flag that can be included in the DescribeCacheCluster request to retrieve information about the individual cache nodes.

data DescribeCacheClusters

Represents the input of a DescribeCacheClusters operation.

See: describeCacheClusters smart constructor.

Request Lenses

dShowCacheClustersNotInReplicationGroups :: Lens' DescribeCacheClusters (Maybe Bool)

An optional flag that can be included in the DescribeCacheCluster request to show only nodes (API/CLI: clusters) that are not members of a replication group. In practice, this mean Memcached and single node Redis clusters.

dCacheClusterId :: Lens' DescribeCacheClusters (Maybe Text)

The user-supplied cluster identifier. If this parameter is specified, only information about that specific cache cluster is returned. This parameter isn't case sensitive.

dMarker :: Lens' DescribeCacheClusters (Maybe Text)

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

dMaxRecords :: Lens' DescribeCacheClusters (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: minimum 20; maximum 100.

dShowCacheNodeInfo :: Lens' DescribeCacheClusters (Maybe Bool)

An optional flag that can be included in the DescribeCacheCluster request to retrieve information about the individual cache nodes.

Destructuring the Response

describeCacheClustersResponse

Arguments

:: Int
drsResponseStatus

-> DescribeCacheClustersResponse

Creates a value of DescribeCacheClustersResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drsCacheClusters - A list of cache clusters. Each item in the list contains detailed information about one cache cluster.

drsMarker - Provides an identifier to allow retrieval of paginated results.

drsResponseStatus - -- | The response status code.

data DescribeCacheClustersResponse

Represents the output of a DescribeCacheClusters operation.

See: describeCacheClustersResponse smart constructor.

Response Lenses

drsCacheClusters :: Lens' DescribeCacheClustersResponse [CacheCluster]

A list of cache clusters. Each item in the list contains detailed information about one cache cluster.

drsMarker :: Lens' DescribeCacheClustersResponse (Maybe Text)

Provides an identifier to allow retrieval of paginated results.

drsResponseStatus :: Lens' DescribeCacheClustersResponse Int

- | The response status code.