Network.AWS.ElastiCache.DescribeCacheParameterGroups

Description

Returns a list of cache parameter group descriptions. If a cache parameter group name is specified, the list contains only the descriptions for that group.

This operation returns paginated results.

Creating a Request

describeCacheParameterGroups :: DescribeCacheParameterGroups

Creates a value of DescribeCacheParameterGroups with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcpgCacheParameterGroupName - The name of a specific cache parameter group to return details for.

dcpgMarker - An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

dcpgMaxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: minimum 20; maximum 100.

data DescribeCacheParameterGroups

Represents the input of a DescribeCacheParameterGroups operation.

See: describeCacheParameterGroups smart constructor.

Request Lenses

dcpgCacheParameterGroupName :: Lens' DescribeCacheParameterGroups (Maybe Text)

The name of a specific cache parameter group to return details for.

dcpgMarker :: Lens' DescribeCacheParameterGroups (Maybe Text)

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

dcpgMaxRecords :: Lens' DescribeCacheParameterGroups (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: minimum 20; maximum 100.

Destructuring the Response

describeCacheParameterGroupsResponse

Arguments

:: Int
dcpgrsResponseStatus

-> DescribeCacheParameterGroupsResponse

Creates a value of DescribeCacheParameterGroupsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcpgrsCacheParameterGroups - A list of cache parameter groups. Each element in the list contains detailed information about one cache parameter group.

dcpgrsMarker - Provides an identifier to allow retrieval of paginated results.

dcpgrsResponseStatus - -- | The response status code.

data DescribeCacheParameterGroupsResponse

Represents the output of a DescribeCacheParameterGroups operation.

See: describeCacheParameterGroupsResponse smart constructor.

Response Lenses

dcpgrsCacheParameterGroups :: Lens' DescribeCacheParameterGroupsResponse [CacheParameterGroup]

A list of cache parameter groups. Each element in the list contains detailed information about one cache parameter group.

dcpgrsMarker :: Lens' DescribeCacheParameterGroupsResponse (Maybe Text)

Provides an identifier to allow retrieval of paginated results.

dcpgrsResponseStatus :: Lens' DescribeCacheParameterGroupsResponse Int

- | The response status code.