Network.AWS.ElastiCache.DescribeCacheParameters

Description

Returns the detailed parameter list for a particular cache parameter group.

This operation returns paginated results.

Creating a Request

describeCacheParameters

Arguments

:: Text
dcpCacheParameterGroupName

-> DescribeCacheParameters

Creates a value of DescribeCacheParameters with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcpMarker - An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

dcpMaxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: minimum 20; maximum 100.

dcpSource - The parameter types to return. Valid values: user | system | engine-default

dcpCacheParameterGroupName - The name of a specific cache parameter group to return details for.

data DescribeCacheParameters

Represents the input of a DescribeCacheParameters operation.

See: describeCacheParameters smart constructor.

Request Lenses

dcpMarker :: Lens' DescribeCacheParameters (Maybe Text)

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

dcpMaxRecords :: Lens' DescribeCacheParameters (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: minimum 20; maximum 100.

dcpSource :: Lens' DescribeCacheParameters (Maybe Text)

The parameter types to return. Valid values: user | system | engine-default

dcpCacheParameterGroupName :: Lens' DescribeCacheParameters Text

The name of a specific cache parameter group to return details for.

Destructuring the Response

describeCacheParametersResponse

Arguments

:: Int
dcprsResponseStatus

-> DescribeCacheParametersResponse

Creates a value of DescribeCacheParametersResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcprsCacheNodeTypeSpecificParameters - A list of parameters specific to a particular cache node type. Each element in the list contains detailed information about one parameter.

dcprsMarker - Provides an identifier to allow retrieval of paginated results.

dcprsParameters - A list of Parameter instances.

dcprsResponseStatus - -- | The response status code.

data DescribeCacheParametersResponse

Represents the output of a DescribeCacheParameters operation.

See: describeCacheParametersResponse smart constructor.

Response Lenses

dcprsCacheNodeTypeSpecificParameters :: Lens' DescribeCacheParametersResponse [CacheNodeTypeSpecificParameter]

A list of parameters specific to a particular cache node type. Each element in the list contains detailed information about one parameter.

dcprsMarker :: Lens' DescribeCacheParametersResponse (Maybe Text)

Provides an identifier to allow retrieval of paginated results.

dcprsParameters :: Lens' DescribeCacheParametersResponse [Parameter]

A list of Parameter instances.

dcprsResponseStatus :: Lens' DescribeCacheParametersResponse Int

- | The response status code.