Network.AWS.ElastiCache.DescribeCacheSecurityGroups

Description

Returns a list of cache security group descriptions. If a cache security group name is specified, the list contains only the description of that group.

This operation returns paginated results.

Creating a Request

describeCacheSecurityGroups :: DescribeCacheSecurityGroups

Creates a value of DescribeCacheSecurityGroups with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcsgsCacheSecurityGroupName - The name of the cache security group to return details for.

dcsgsMarker - An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

dcsgsMaxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: minimum 20; maximum 100.

data DescribeCacheSecurityGroups

Represents the input of a DescribeCacheSecurityGroups operation.

See: describeCacheSecurityGroups smart constructor.

Request Lenses

dcsgsCacheSecurityGroupName :: Lens' DescribeCacheSecurityGroups (Maybe Text)

The name of the cache security group to return details for.

dcsgsMarker :: Lens' DescribeCacheSecurityGroups (Maybe Text)

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

dcsgsMaxRecords :: Lens' DescribeCacheSecurityGroups (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: minimum 20; maximum 100.

Destructuring the Response

describeCacheSecurityGroupsResponse

Arguments

:: Int
dcsgsrsResponseStatus

-> DescribeCacheSecurityGroupsResponse

Creates a value of DescribeCacheSecurityGroupsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcsgsrsCacheSecurityGroups - A list of cache security groups. Each element in the list contains detailed information about one group.

dcsgsrsMarker - Provides an identifier to allow retrieval of paginated results.

dcsgsrsResponseStatus - -- | The response status code.

data DescribeCacheSecurityGroupsResponse

Represents the output of a DescribeCacheSecurityGroups operation.

See: describeCacheSecurityGroupsResponse smart constructor.

Response Lenses

dcsgsrsCacheSecurityGroups :: Lens' DescribeCacheSecurityGroupsResponse [CacheSecurityGroup]

A list of cache security groups. Each element in the list contains detailed information about one group.

dcsgsrsMarker :: Lens' DescribeCacheSecurityGroupsResponse (Maybe Text)

Provides an identifier to allow retrieval of paginated results.

dcsgsrsResponseStatus :: Lens' DescribeCacheSecurityGroupsResponse Int

- | The response status code.