Network.AWS.ElastiCache.DescribeReservedCacheNodes

Description

Returns information about reserved cache nodes for this account, or about a specified reserved cache node.

This operation returns paginated results.

Creating a Request

describeReservedCacheNodes :: DescribeReservedCacheNodes

Creates a value of DescribeReservedCacheNodes with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drcnCacheNodeType - The cache node type filter value. Use this parameter to show only those reservations matching the specified cache node type. Valid node types are as follows: * General purpose: * Current generation: cache.t2.micro , cache.t2.small , cache.t2.medium , cache.m3.medium , cache.m3.large , cache.m3.xlarge , cache.m3.2xlarge , cache.m4.large , cache.m4.xlarge , cache.m4.2xlarge , cache.m4.4xlarge , cache.m4.10xlarge * Previous generation: cache.t1.micro , cache.m1.small , cache.m1.medium , cache.m1.large , cache.m1.xlarge * Compute optimized: cache.c1.xlarge * Memory optimized: * Current generation: cache.r3.large , cache.r3.xlarge , cache.r3.2xlarge , cache.r3.4xlarge , cache.r3.8xlarge * Previous generation: cache.m2.xlarge , cache.m2.2xlarge , cache.m2.4xlarge Notes: * All T2 instances are created in an Amazon Virtual Private Cloud (Amazon VPC). * Redis backuprestore is not supported for Redis (cluster mode disabled) T1 and T2 instances. Backuprestore is supported on Redis (cluster mode enabled) T2 instances. * Redis Append-only files (AOF) functionality is not supported for T1 or T2 instances. For a complete listing of node types and specifications, see Amazon ElastiCache Product Features and Details and either Cache Node Type-Specific Parameters for Memcached or Cache Node Type-Specific Parameters for Redis .

drcnProductDescription - The product description filter value. Use this parameter to show only those reservations matching the specified product description.

drcnMarker - An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

drcnMaxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: minimum 20; maximum 100.

drcnReservedCacheNodeId - The reserved cache node identifier filter value. Use this parameter to show only the reservation that matches the specified reservation ID.

drcnOfferingType - The offering type filter value. Use this parameter to show only the available offerings matching the specified offering type. Valid values: "Light Utilization"|"Medium Utilization"|"Heavy Utilization"

drcnDuration - The duration filter value, specified in years or seconds. Use this parameter to show only reservations for this duration. Valid Values: 1 | 3 | 31536000 | 94608000

drcnReservedCacheNodesOfferingId - The offering identifier filter value. Use this parameter to show only purchased reservations matching the specified offering identifier.

data DescribeReservedCacheNodes

Represents the input of a DescribeReservedCacheNodes operation.

See: describeReservedCacheNodes smart constructor.

Request Lenses

drcnCacheNodeType :: Lens' DescribeReservedCacheNodes (Maybe Text)

The cache node type filter value. Use this parameter to show only those reservations matching the specified cache node type. Valid node types are as follows: * General purpose: * Current generation: cache.t2.micro , cache.t2.small , cache.t2.medium , cache.m3.medium , cache.m3.large , cache.m3.xlarge , cache.m3.2xlarge , cache.m4.large , cache.m4.xlarge , cache.m4.2xlarge , cache.m4.4xlarge , cache.m4.10xlarge * Previous generation: cache.t1.micro , cache.m1.small , cache.m1.medium , cache.m1.large , cache.m1.xlarge * Compute optimized: cache.c1.xlarge * Memory optimized: * Current generation: cache.r3.large , cache.r3.xlarge , cache.r3.2xlarge , cache.r3.4xlarge , cache.r3.8xlarge * Previous generation: cache.m2.xlarge , cache.m2.2xlarge , cache.m2.4xlarge Notes: * All T2 instances are created in an Amazon Virtual Private Cloud (Amazon VPC). * Redis backuprestore is not supported for Redis (cluster mode disabled) T1 and T2 instances. Backuprestore is supported on Redis (cluster mode enabled) T2 instances. * Redis Append-only files (AOF) functionality is not supported for T1 or T2 instances. For a complete listing of node types and specifications, see Amazon ElastiCache Product Features and Details and either Cache Node Type-Specific Parameters for Memcached or Cache Node Type-Specific Parameters for Redis .

drcnProductDescription :: Lens' DescribeReservedCacheNodes (Maybe Text)

The product description filter value. Use this parameter to show only those reservations matching the specified product description.

drcnMarker :: Lens' DescribeReservedCacheNodes (Maybe Text)

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

drcnMaxRecords :: Lens' DescribeReservedCacheNodes (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: minimum 20; maximum 100.

drcnReservedCacheNodeId :: Lens' DescribeReservedCacheNodes (Maybe Text)

The reserved cache node identifier filter value. Use this parameter to show only the reservation that matches the specified reservation ID.

drcnOfferingType :: Lens' DescribeReservedCacheNodes (Maybe Text)

The offering type filter value. Use this parameter to show only the available offerings matching the specified offering type. Valid values: "Light Utilization"|"Medium Utilization"|"Heavy Utilization"

drcnDuration :: Lens' DescribeReservedCacheNodes (Maybe Text)

The duration filter value, specified in years or seconds. Use this parameter to show only reservations for this duration. Valid Values: 1 | 3 | 31536000 | 94608000

drcnReservedCacheNodesOfferingId :: Lens' DescribeReservedCacheNodes (Maybe Text)

The offering identifier filter value. Use this parameter to show only purchased reservations matching the specified offering identifier.

Destructuring the Response

describeReservedCacheNodesResponse

Arguments

:: Int
drcnrsResponseStatus

-> DescribeReservedCacheNodesResponse

Creates a value of DescribeReservedCacheNodesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drcnrsMarker - Provides an identifier to allow retrieval of paginated results.

drcnrsReservedCacheNodes - A list of reserved cache nodes. Each element in the list contains detailed information about one node.

drcnrsResponseStatus - -- | The response status code.

data DescribeReservedCacheNodesResponse

Represents the output of a DescribeReservedCacheNodes operation.

See: describeReservedCacheNodesResponse smart constructor.

Response Lenses

drcnrsMarker :: Lens' DescribeReservedCacheNodesResponse (Maybe Text)

Provides an identifier to allow retrieval of paginated results.

drcnrsReservedCacheNodes :: Lens' DescribeReservedCacheNodesResponse [ReservedCacheNode]

A list of reserved cache nodes. Each element in the list contains detailed information about one node.

drcnrsResponseStatus :: Lens' DescribeReservedCacheNodesResponse Int

- | The response status code.