Network.AWS.ElastiCache.DescribeSnapshots

Description

Returns information about cache cluster or replication group snapshots. By default, DescribeSnapshots lists all of your snapshots; it can optionally describe a single snapshot, or just the snapshots associated with a particular cache cluster.

This operation returns paginated results.

Creating a Request

describeSnapshots :: DescribeSnapshots

Creates a value of DescribeSnapshots with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsCacheClusterId - A user-supplied cluster identifier. If this parameter is specified, only snapshots associated with that specific cache cluster are described.

dsMarker - An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

dsMaxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved. Default: 50 Constraints: minimum 20; maximum 50.

dsSnapshotName - A user-supplied name of the snapshot. If this parameter is specified, only this snapshot are described.

dsShowNodeGroupConfig - A Boolean value which if true, the node group (shard) configuration is included in the snapshot description.

dsReplicationGroupId - A user-supplied replication group identifier. If this parameter is specified, only snapshots associated with that specific replication group are described.

dsSnapshotSource - If set to system , the output shows snapshots that were automatically created by ElastiCache. If set to user the output shows snapshots that were manually created. If omitted, the output shows both automatically and manually created snapshots.

data DescribeSnapshots

Represents the input of a DescribeSnapshotsMessage operation.

See: describeSnapshots smart constructor.

Request Lenses

dsCacheClusterId :: Lens' DescribeSnapshots (Maybe Text)

A user-supplied cluster identifier. If this parameter is specified, only snapshots associated with that specific cache cluster are described.

dsMarker :: Lens' DescribeSnapshots (Maybe Text)

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

dsMaxRecords :: Lens' DescribeSnapshots (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved. Default: 50 Constraints: minimum 20; maximum 50.

dsSnapshotName :: Lens' DescribeSnapshots (Maybe Text)

A user-supplied name of the snapshot. If this parameter is specified, only this snapshot are described.

dsShowNodeGroupConfig :: Lens' DescribeSnapshots (Maybe Bool)

A Boolean value which if true, the node group (shard) configuration is included in the snapshot description.

dsReplicationGroupId :: Lens' DescribeSnapshots (Maybe Text)

A user-supplied replication group identifier. If this parameter is specified, only snapshots associated with that specific replication group are described.

dsSnapshotSource :: Lens' DescribeSnapshots (Maybe Text)

If set to system , the output shows snapshots that were automatically created by ElastiCache. If set to user the output shows snapshots that were manually created. If omitted, the output shows both automatically and manually created snapshots.

Destructuring the Response

describeSnapshotsResponse

Arguments

:: Int
dssrsResponseStatus

-> DescribeSnapshotsResponse

Creates a value of DescribeSnapshotsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dssrsSnapshots - A list of snapshots. Each item in the list contains detailed information about one snapshot.

dssrsMarker - An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

dssrsResponseStatus - -- | The response status code.

data DescribeSnapshotsResponse

Represents the output of a DescribeSnapshots operation.

See: describeSnapshotsResponse smart constructor.

Response Lenses

dssrsSnapshots :: Lens' DescribeSnapshotsResponse [Snapshot]

A list of snapshots. Each item in the list contains detailed information about one snapshot.

dssrsMarker :: Lens' DescribeSnapshotsResponse (Maybe Text)

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

dssrsResponseStatus :: Lens' DescribeSnapshotsResponse Int

- | The response status code.