Network.AWS.ElastiCache.ModifyCacheCluster

Description

Modifies the settings for a cache cluster. You can use this operation to change one or more cluster configuration parameters by specifying the parameters and the new values.

Creating a Request

modifyCacheCluster

Arguments

:: Text
mccCacheClusterId

-> ModifyCacheCluster

Creates a value of ModifyCacheCluster with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mccEngineVersion - The upgraded version of the cache engine to be run on the cache nodes. Important: You can upgrade to a newer engine version (see Selecting a Cache Engine and Version ), but you cannot downgrade to an earlier engine version. If you want to use an earlier engine version, you must delete the existing cache cluster and create it anew with the earlier engine version.

mccCacheNodeType - A valid cache node type that you want to scale this cache cluster up to.

mccSecurityGroupIds - Specifies the VPC Security Groups associated with the cache cluster. This parameter can be used only with clusters that are created in an Amazon Virtual Private Cloud (Amazon VPC).

mccAutoMinorVersionUpgrade - This parameter is currently disabled.

mccCacheParameterGroupName - The name of the cache parameter group to apply to this cache cluster. This change is asynchronously applied as soon as possible for parameters when the ApplyImmediately parameter is specified as true for this request.

mccSnapshotWindow - The daily time range (in UTC) during which ElastiCache begins taking a daily snapshot of your cache cluster.

mccNewAvailabilityZones - The list of Availability Zones where the new Memcached cache nodes are created. This parameter is only valid when NumCacheNodes in the request is greater than the sum of the number of active cache nodes and the number of cache nodes pending creation (which may be zero). The number of Availability Zones supplied in this list must match the cache nodes being added in this request. This option is only supported on Memcached clusters. Scenarios: * Scenario 1: You have 3 active nodes and wish to add 2 nodes. Specify NumCacheNodes=5 (3 + 2) and optionally specify two Availability Zones for the two new nodes. * Scenario 2: You have 3 active nodes and 2 nodes pending creation (from the scenario 1 call) and want to add 1 more node. Specify NumCacheNodes=6 ((3 + 2) + 1) and optionally specify an Availability Zone for the new node. * Scenario 3: You want to cancel all pending operations. Specify NumCacheNodes=3 to cancel all pending operations. The Availability Zone placement of nodes pending creation cannot be modified. If you wish to cancel any nodes pending creation, add 0 nodes by setting NumCacheNodes to the number of current nodes. If cross-az is specified, existing Memcached nodes remain in their current Availability Zone. Only newly created nodes can be located in different Availability Zones. For guidance on how to move existing Memcached nodes to different Availability Zones, see the Availability Zone Considerations section of Cache Node Considerations for Memcached . Impact of new add/remove requests upon pending requests * Scenario-1 * Pending Action: Delete * New Request: Delete * Result: The new delete, pending or immediate, replaces the pending delete. * Scenario-2 * Pending Action: Delete * New Request: Create * Result: The new create, pending or immediate, replaces the pending delete. * Scenario-3 * Pending Action: Create * New Request: Delete * Result: The new delete, pending or immediate, replaces the pending create. * Scenario-4 * Pending Action: Create * New Request: Create * Result: The new create is added to the pending create. Important: Important: If the new create request is Apply Immediately - Yes , all creates are performed immediately. If the new create request is Apply Immediately - No , all creates are pending.

mccPreferredMaintenanceWindow - Specifies the weekly time range during which maintenance on the cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period. Valid values for ddd are: * sun * mon * tue * wed * thu * fri * sat Example: sun:23:00-mon:01:30

mccCacheNodeIdsToRemove - A list of cache node IDs to be removed. A node ID is a numeric identifier (0001, 0002, etc.). This parameter is only valid when NumCacheNodes is less than the existing number of cache nodes. The number of cache node IDs supplied in this parameter must match the difference between the existing number of cache nodes in the cluster or pending cache nodes, whichever is greater, and the value of NumCacheNodes in the request. For example: If you have 3 active cache nodes, 7 pending cache nodes, and the number of cache nodes in this ModifyCacheCluser call is 5, you must list 2 (7 - 5) cache node IDs to remove.

mccSnapshotRetentionLimit - The number of days for which ElastiCache retains automatic cache cluster snapshots before deleting them. For example, if you set SnapshotRetentionLimit to 5, a snapshot that was taken today is retained for 5 days before being deleted.

mccNotificationTopicStatus - The status of the Amazon SNS notification topic. Notifications are sent only if the status is active . Valid values: active | inactive

mccAZMode - Specifies whether the new nodes in this Memcached cache cluster are all created in a single Availability Zone or created across multiple Availability Zones. Valid values: single-az | cross-az . This option is only supported for Memcached cache clusters.

mccApplyImmediately - If true , this parameter causes the modifications in this request and any pending modifications to be applied, asynchronously and as soon as possible, regardless of the PreferredMaintenanceWindow setting for the cache cluster. If false , changes to the cache cluster are applied on the next maintenance reboot, or the next failure reboot, whichever occurs first. Important: If you perform a ModifyCacheCluster before a pending modification is applied, the pending modification is replaced by the newer modification. Valid values: true | false Default: false

mccNotificationTopicARN - The Amazon Resource Name (ARN) of the Amazon SNS topic to which notifications are sent.

mccNumCacheNodes - The number of cache nodes that the cache cluster should have. If the value for NumCacheNodes is greater than the sum of the number of current cache nodes and the number of cache nodes pending creation (which may be zero), more nodes are added. If the value is less than the number of existing cache nodes, nodes are removed. If the value is equal to the number of current cache nodes, any pending add or remove requests are canceled. If you are removing cache nodes, you must use the CacheNodeIdsToRemove parameter to provide the IDs of the specific cache nodes to remove. For clusters running Redis, this value must be 1. For clusters running Memcached, this value must be between 1 and 20.

mccCacheSecurityGroupNames - A list of cache security group names to authorize on this cache cluster. This change is asynchronously applied as soon as possible. You can use this parameter only with clusters that are created outside of an Amazon Virtual Private Cloud (Amazon VPC). Constraints: Must contain no more than 255 alphanumeric characters. Must not be Default.

mccCacheClusterId - The cache cluster identifier. This value is stored as a lowercase string.

data ModifyCacheCluster

Represents the input of a ModifyCacheCluster operation.

See: modifyCacheCluster smart constructor.

Request Lenses

mccEngineVersion :: Lens' ModifyCacheCluster (Maybe Text)

The upgraded version of the cache engine to be run on the cache nodes. Important: You can upgrade to a newer engine version (see Selecting a Cache Engine and Version ), but you cannot downgrade to an earlier engine version. If you want to use an earlier engine version, you must delete the existing cache cluster and create it anew with the earlier engine version.

mccCacheNodeType :: Lens' ModifyCacheCluster (Maybe Text)

A valid cache node type that you want to scale this cache cluster up to.

mccSecurityGroupIds :: Lens' ModifyCacheCluster [Text]

Specifies the VPC Security Groups associated with the cache cluster. This parameter can be used only with clusters that are created in an Amazon Virtual Private Cloud (Amazon VPC).

mccAutoMinorVersionUpgrade :: Lens' ModifyCacheCluster (Maybe Bool)

This parameter is currently disabled.

mccCacheParameterGroupName :: Lens' ModifyCacheCluster (Maybe Text)

The name of the cache parameter group to apply to this cache cluster. This change is asynchronously applied as soon as possible for parameters when the ApplyImmediately parameter is specified as true for this request.

mccSnapshotWindow :: Lens' ModifyCacheCluster (Maybe Text)

The daily time range (in UTC) during which ElastiCache begins taking a daily snapshot of your cache cluster.

mccNewAvailabilityZones :: Lens' ModifyCacheCluster [Text]

The list of Availability Zones where the new Memcached cache nodes are created. This parameter is only valid when NumCacheNodes in the request is greater than the sum of the number of active cache nodes and the number of cache nodes pending creation (which may be zero). The number of Availability Zones supplied in this list must match the cache nodes being added in this request. This option is only supported on Memcached clusters. Scenarios: * Scenario 1: You have 3 active nodes and wish to add 2 nodes. Specify NumCacheNodes=5 (3 + 2) and optionally specify two Availability Zones for the two new nodes. * Scenario 2: You have 3 active nodes and 2 nodes pending creation (from the scenario 1 call) and want to add 1 more node. Specify NumCacheNodes=6 ((3 + 2) + 1) and optionally specify an Availability Zone for the new node. * Scenario 3: You want to cancel all pending operations. Specify NumCacheNodes=3 to cancel all pending operations. The Availability Zone placement of nodes pending creation cannot be modified. If you wish to cancel any nodes pending creation, add 0 nodes by setting NumCacheNodes to the number of current nodes. If cross-az is specified, existing Memcached nodes remain in their current Availability Zone. Only newly created nodes can be located in different Availability Zones. For guidance on how to move existing Memcached nodes to different Availability Zones, see the Availability Zone Considerations section of Cache Node Considerations for Memcached . Impact of new add/remove requests upon pending requests * Scenario-1 * Pending Action: Delete * New Request: Delete * Result: The new delete, pending or immediate, replaces the pending delete. * Scenario-2 * Pending Action: Delete * New Request: Create * Result: The new create, pending or immediate, replaces the pending delete. * Scenario-3 * Pending Action: Create * New Request: Delete * Result: The new delete, pending or immediate, replaces the pending create. * Scenario-4 * Pending Action: Create * New Request: Create * Result: The new create is added to the pending create. Important: Important: If the new create request is Apply Immediately - Yes , all creates are performed immediately. If the new create request is Apply Immediately - No , all creates are pending.

mccPreferredMaintenanceWindow :: Lens' ModifyCacheCluster (Maybe Text)

Specifies the weekly time range during which maintenance on the cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period. Valid values for ddd are: * sun * mon * tue * wed * thu * fri * sat Example: sun:23:00-mon:01:30

mccCacheNodeIdsToRemove :: Lens' ModifyCacheCluster [Text]

A list of cache node IDs to be removed. A node ID is a numeric identifier (0001, 0002, etc.). This parameter is only valid when NumCacheNodes is less than the existing number of cache nodes. The number of cache node IDs supplied in this parameter must match the difference between the existing number of cache nodes in the cluster or pending cache nodes, whichever is greater, and the value of NumCacheNodes in the request. For example: If you have 3 active cache nodes, 7 pending cache nodes, and the number of cache nodes in this ModifyCacheCluser call is 5, you must list 2 (7 - 5) cache node IDs to remove.

mccSnapshotRetentionLimit :: Lens' ModifyCacheCluster (Maybe Int)

The number of days for which ElastiCache retains automatic cache cluster snapshots before deleting them. For example, if you set SnapshotRetentionLimit to 5, a snapshot that was taken today is retained for 5 days before being deleted.

mccNotificationTopicStatus :: Lens' ModifyCacheCluster (Maybe Text)

The status of the Amazon SNS notification topic. Notifications are sent only if the status is active . Valid values: active | inactive

mccAZMode :: Lens' ModifyCacheCluster (Maybe AZMode)

Specifies whether the new nodes in this Memcached cache cluster are all created in a single Availability Zone or created across multiple Availability Zones. Valid values: single-az | cross-az . This option is only supported for Memcached cache clusters.

mccApplyImmediately :: Lens' ModifyCacheCluster (Maybe Bool)

If true , this parameter causes the modifications in this request and any pending modifications to be applied, asynchronously and as soon as possible, regardless of the PreferredMaintenanceWindow setting for the cache cluster. If false , changes to the cache cluster are applied on the next maintenance reboot, or the next failure reboot, whichever occurs first. Important: If you perform a ModifyCacheCluster before a pending modification is applied, the pending modification is replaced by the newer modification. Valid values: true | false Default: false

mccNotificationTopicARN :: Lens' ModifyCacheCluster (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon SNS topic to which notifications are sent.

mccNumCacheNodes :: Lens' ModifyCacheCluster (Maybe Int)

The number of cache nodes that the cache cluster should have. If the value for NumCacheNodes is greater than the sum of the number of current cache nodes and the number of cache nodes pending creation (which may be zero), more nodes are added. If the value is less than the number of existing cache nodes, nodes are removed. If the value is equal to the number of current cache nodes, any pending add or remove requests are canceled. If you are removing cache nodes, you must use the CacheNodeIdsToRemove parameter to provide the IDs of the specific cache nodes to remove. For clusters running Redis, this value must be 1. For clusters running Memcached, this value must be between 1 and 20.

mccCacheSecurityGroupNames :: Lens' ModifyCacheCluster [Text]

A list of cache security group names to authorize on this cache cluster. This change is asynchronously applied as soon as possible. You can use this parameter only with clusters that are created outside of an Amazon Virtual Private Cloud (Amazon VPC). Constraints: Must contain no more than 255 alphanumeric characters. Must not be Default.

mccCacheClusterId :: Lens' ModifyCacheCluster Text

The cache cluster identifier. This value is stored as a lowercase string.

Destructuring the Response

modifyCacheClusterResponse

Arguments

:: Int
mccrsResponseStatus

-> ModifyCacheClusterResponse

Creates a value of ModifyCacheClusterResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mccrsCacheCluster - Undocumented member.

mccrsResponseStatus - -- | The response status code.

data ModifyCacheClusterResponse

See: modifyCacheClusterResponse smart constructor.

Response Lenses

mccrsCacheCluster :: Lens' ModifyCacheClusterResponse (Maybe CacheCluster)

Undocumented member.

mccrsResponseStatus :: Lens' ModifyCacheClusterResponse Int

- | The response status code.