Network.AWS.ElastiCache.ModifyReplicationGroup

Description

Modifies the settings for a replication group.

Important: Due to current limitations on Redis (cluster mode disabled), this operation or parameter is not supported on Redis (cluster mode enabled) replication groups.

Creating a Request

modifyReplicationGroup

Arguments

:: Text
mrgReplicationGroupId

-> ModifyReplicationGroup

Creates a value of ModifyReplicationGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mrgAutomaticFailoverEnabled - Determines whether a read replica is automatically promoted to read/write primary if the existing primary encounters a failure. Valid values: true | false

mrgEngineVersion - The upgraded version of the cache engine to be run on the cache clusters in the replication group. Important: You can upgrade to a newer engine version (see Selecting a Cache Engine and Version ), but you cannot downgrade to an earlier engine version. If you want to use an earlier engine version, you must delete the existing replication group and create it anew with the earlier engine version.

mrgCacheNodeType - A valid cache node type that you want to scale this replication group to.

mrgSnapshottingClusterId - The cache cluster ID that is used as the daily snapshot source for the replication group. This parameter cannot be set for Redis (cluster mode enabled) replication groups.

mrgSecurityGroupIds - Specifies the VPC Security Groups associated with the cache clusters in the replication group. This parameter can be used only with replication group containing cache clusters running in an Amazon Virtual Private Cloud (Amazon VPC).

mrgAutoMinorVersionUpgrade - This parameter is currently disabled.

mrgCacheParameterGroupName - The name of the cache parameter group to apply to all of the clusters in this replication group. This change is asynchronously applied as soon as possible for parameters when the ApplyImmediately parameter is specified as true for this request.

mrgReplicationGroupDescription - A description for the replication group. Maximum length is 255 characters.

mrgSnapshotWindow - The daily time range (in UTC) during which ElastiCache begins taking a daily snapshot of the node group (shard) specified by SnapshottingClusterId . Example: 05:00-09:00 If you do not specify this parameter, ElastiCache automatically chooses an appropriate time range.

mrgPrimaryClusterId - For replication groups with a single primary, if this parameter is specified, ElastiCache promotes the specified cluster in the specified replication group to the primary role. The nodes of all other clusters in the replication group are read replicas.

mrgPreferredMaintenanceWindow - Specifies the weekly time range during which maintenance on the cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period. Valid values for ddd are: * sun * mon * tue * wed * thu * fri * sat Example: sun:23:00-mon:01:30

mrgNodeGroupId - The name of the Node Group (called shard in the console).

mrgSnapshotRetentionLimit - The number of days for which ElastiCache retains automatic node group (shard) snapshots before deleting them. For example, if you set SnapshotRetentionLimit to 5, a snapshot that was taken today is retained for 5 days before being deleted. Important If the value of SnapshotRetentionLimit is set to zero (0), backups are turned off.

mrgNotificationTopicStatus - The status of the Amazon SNS notification topic for the replication group. Notifications are sent only if the status is active . Valid values: active | inactive

mrgApplyImmediately - If true , this parameter causes the modifications in this request and any pending modifications to be applied, asynchronously and as soon as possible, regardless of the PreferredMaintenanceWindow setting for the replication group. If false , changes to the nodes in the replication group are applied on the next maintenance reboot, or the next failure reboot, whichever occurs first. Valid values: true | false Default: false

mrgNotificationTopicARN - The Amazon Resource Name (ARN) of the Amazon SNS topic to which notifications are sent.

mrgCacheSecurityGroupNames - A list of cache security group names to authorize for the clusters in this replication group. This change is asynchronously applied as soon as possible. This parameter can be used only with replication group containing cache clusters running outside of an Amazon Virtual Private Cloud (Amazon VPC). Constraints: Must contain no more than 255 alphanumeric characters. Must not be Default .

mrgReplicationGroupId - The identifier of the replication group to modify.

data ModifyReplicationGroup

Represents the input of a ModifyReplicationGroups operation.

See: modifyReplicationGroup smart constructor.

Request Lenses

mrgAutomaticFailoverEnabled :: Lens' ModifyReplicationGroup (Maybe Bool)

Determines whether a read replica is automatically promoted to read/write primary if the existing primary encounters a failure. Valid values: true | false

mrgEngineVersion :: Lens' ModifyReplicationGroup (Maybe Text)

The upgraded version of the cache engine to be run on the cache clusters in the replication group. Important: You can upgrade to a newer engine version (see Selecting a Cache Engine and Version ), but you cannot downgrade to an earlier engine version. If you want to use an earlier engine version, you must delete the existing replication group and create it anew with the earlier engine version.

mrgCacheNodeType :: Lens' ModifyReplicationGroup (Maybe Text)

A valid cache node type that you want to scale this replication group to.

mrgSnapshottingClusterId :: Lens' ModifyReplicationGroup (Maybe Text)

The cache cluster ID that is used as the daily snapshot source for the replication group. This parameter cannot be set for Redis (cluster mode enabled) replication groups.

mrgSecurityGroupIds :: Lens' ModifyReplicationGroup [Text]

Specifies the VPC Security Groups associated with the cache clusters in the replication group. This parameter can be used only with replication group containing cache clusters running in an Amazon Virtual Private Cloud (Amazon VPC).

mrgAutoMinorVersionUpgrade :: Lens' ModifyReplicationGroup (Maybe Bool)

This parameter is currently disabled.

mrgCacheParameterGroupName :: Lens' ModifyReplicationGroup (Maybe Text)

The name of the cache parameter group to apply to all of the clusters in this replication group. This change is asynchronously applied as soon as possible for parameters when the ApplyImmediately parameter is specified as true for this request.

mrgReplicationGroupDescription :: Lens' ModifyReplicationGroup (Maybe Text)

A description for the replication group. Maximum length is 255 characters.

mrgSnapshotWindow :: Lens' ModifyReplicationGroup (Maybe Text)

The daily time range (in UTC) during which ElastiCache begins taking a daily snapshot of the node group (shard) specified by SnapshottingClusterId . Example: 05:00-09:00 If you do not specify this parameter, ElastiCache automatically chooses an appropriate time range.

mrgPrimaryClusterId :: Lens' ModifyReplicationGroup (Maybe Text)

For replication groups with a single primary, if this parameter is specified, ElastiCache promotes the specified cluster in the specified replication group to the primary role. The nodes of all other clusters in the replication group are read replicas.

mrgPreferredMaintenanceWindow :: Lens' ModifyReplicationGroup (Maybe Text)

Specifies the weekly time range during which maintenance on the cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period. Valid values for ddd are: * sun * mon * tue * wed * thu * fri * sat Example: sun:23:00-mon:01:30

mrgNodeGroupId :: Lens' ModifyReplicationGroup (Maybe Text)

The name of the Node Group (called shard in the console).

mrgSnapshotRetentionLimit :: Lens' ModifyReplicationGroup (Maybe Int)

The number of days for which ElastiCache retains automatic node group (shard) snapshots before deleting them. For example, if you set SnapshotRetentionLimit to 5, a snapshot that was taken today is retained for 5 days before being deleted. Important If the value of SnapshotRetentionLimit is set to zero (0), backups are turned off.

mrgNotificationTopicStatus :: Lens' ModifyReplicationGroup (Maybe Text)

The status of the Amazon SNS notification topic for the replication group. Notifications are sent only if the status is active . Valid values: active | inactive

mrgApplyImmediately :: Lens' ModifyReplicationGroup (Maybe Bool)

If true , this parameter causes the modifications in this request and any pending modifications to be applied, asynchronously and as soon as possible, regardless of the PreferredMaintenanceWindow setting for the replication group. If false , changes to the nodes in the replication group are applied on the next maintenance reboot, or the next failure reboot, whichever occurs first. Valid values: true | false Default: false

mrgNotificationTopicARN :: Lens' ModifyReplicationGroup (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon SNS topic to which notifications are sent.

mrgCacheSecurityGroupNames :: Lens' ModifyReplicationGroup [Text]

A list of cache security group names to authorize for the clusters in this replication group. This change is asynchronously applied as soon as possible. This parameter can be used only with replication group containing cache clusters running outside of an Amazon Virtual Private Cloud (Amazon VPC). Constraints: Must contain no more than 255 alphanumeric characters. Must not be Default .

mrgReplicationGroupId :: Lens' ModifyReplicationGroup Text

The identifier of the replication group to modify.

Destructuring the Response

modifyReplicationGroupResponse

Arguments

:: Int
mrgrsResponseStatus

-> ModifyReplicationGroupResponse

Creates a value of ModifyReplicationGroupResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mrgrsReplicationGroup - Undocumented member.

mrgrsResponseStatus - -- | The response status code.

data ModifyReplicationGroupResponse

See: modifyReplicationGroupResponse smart constructor.

Response Lenses

mrgrsReplicationGroup :: Lens' ModifyReplicationGroupResponse (Maybe ReplicationGroup)

Undocumented member.

mrgrsResponseStatus :: Lens' ModifyReplicationGroupResponse Int

- | The response status code.