Network.AWS.ElastiCache.ModifyReplicationGroup

Description

The ModifyReplicationGroup action modifies the settings for a replication group.

http://docs.aws.amazon.com/AmazonElastiCache/latest/APIReference/API_ModifyReplicationGroup.html

Request

data ModifyReplicationGroup

Request constructor

modifyReplicationGroup

Arguments

:: Text
mrgReplicationGroupId

-> ModifyReplicationGroup

ModifyReplicationGroup constructor.

The fields accessible through corresponding lenses are:

mrgApplyImmediately :: Maybe Bool

mrgAutoMinorVersionUpgrade :: Maybe Bool

mrgAutomaticFailoverEnabled :: Maybe Bool

mrgCacheParameterGroupName :: Maybe Text

mrgCacheSecurityGroupNames :: [Text]

mrgEngineVersion :: Maybe Text

mrgNotificationTopicArn :: Maybe Text

mrgNotificationTopicStatus :: Maybe Text

mrgPreferredMaintenanceWindow :: Maybe Text

mrgPrimaryClusterId :: Maybe Text

mrgReplicationGroupDescription :: Maybe Text

mrgReplicationGroupId :: Text

mrgSecurityGroupIds :: [Text]

mrgSnapshotRetentionLimit :: Maybe Int

mrgSnapshotWindow :: Maybe Text

mrgSnapshottingClusterId :: Maybe Text

Request lenses

mrgApplyImmediately :: Lens' ModifyReplicationGroup (Maybe Bool)

If true, this parameter causes the modifications in this request and any pending modifications to be applied, asynchronously and as soon as possible, regardless of the PreferredMaintenanceWindow setting for the replication group.

If false, then changes to the nodes in the replication group are applied on the next maintenance reboot, or the next failure reboot, whichever occurs first.

Valid values: true | false

Default: false

mrgAutoMinorVersionUpgrade :: Lens' ModifyReplicationGroup (Maybe Bool)

This parameter is currently disabled.

mrgAutomaticFailoverEnabled :: Lens' ModifyReplicationGroup (Maybe Bool)

Whether a read replica will be automatically promoted to read/write primary if the existing primary encounters a failure.

Valid values: true | false

ElastiCache Multi-AZ replication groups are not supported on:

Redis versions earlier than 2.8.6. T1 and T2 cache node types.

mrgCacheParameterGroupName :: Lens' ModifyReplicationGroup (Maybe Text)

The name of the cache parameter group to apply to all of the clusters in this replication group. This change is asynchronously applied as soon as possible for parameters when the ApplyImmediately parameter is specified as true for this request.

mrgCacheSecurityGroupNames :: Lens' ModifyReplicationGroup [Text]

A list of cache security group names to authorize for the clusters in this replication group. This change is asynchronously applied as soon as possible.

This parameter can be used only with replication group containing cache clusters running outside of an Amazon Virtual Private Cloud (VPC).

Constraints: Must contain no more than 255 alphanumeric characters. Must not be Default.

mrgEngineVersion :: Lens' ModifyReplicationGroup (Maybe Text)

The upgraded version of the cache engine to be run on the cache clusters in the replication group.

mrgNotificationTopicArn :: Lens' ModifyReplicationGroup (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon SNS topic to which notifications will be sent.

The Amazon SNS topic owner must be same as the replication group owner.

mrgNotificationTopicStatus :: Lens' ModifyReplicationGroup (Maybe Text)

The status of the Amazon SNS notification topic for the replication group. Notifications are sent only if the status is active.

Valid values: active | inactive

mrgPreferredMaintenanceWindow :: Lens' ModifyReplicationGroup (Maybe Text)

Specifies the weekly time range during which maintenance on the cache cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period. Valid values for ddd are:

sun mon tue wed thu fri sat Example: 'sun:05:00-sun:09:00'

mrgPrimaryClusterId :: Lens' ModifyReplicationGroup (Maybe Text)

If this parameter is specified, ElastiCache will promote each of the cache clusters in the specified replication group to the primary role. The nodes of all other cache clusters in the replication group will be read replicas.

mrgReplicationGroupDescription :: Lens' ModifyReplicationGroup (Maybe Text)

A description for the replication group. Maximum length is 255 characters.

mrgReplicationGroupId :: Lens' ModifyReplicationGroup Text

The identifier of the replication group to modify.

mrgSecurityGroupIds :: Lens' ModifyReplicationGroup [Text]

Specifies the VPC Security Groups associated with the cache clusters in the replication group.

This parameter can be used only with replication group containing cache clusters running in an Amazon Virtual Private Cloud (VPC).

mrgSnapshotRetentionLimit :: Lens' ModifyReplicationGroup (Maybe Int)

The number of days for which ElastiCache will retain automatic node group snapshots before deleting them. For example, if you set SnapshotRetentionLimit to 5, then a snapshot that was taken today will be retained for 5 days before being deleted.

Important If the value of SnapshotRetentionLimit is set to zero (0), backups are turned off.

mrgSnapshotWindow :: Lens' ModifyReplicationGroup (Maybe Text)

The daily time range (in UTC) during which ElastiCache will begin taking a daily snapshot of the node group specified by SnapshottingClusterId.

Example: '05:00-09:00'

If you do not specify this parameter, then ElastiCache will automatically choose an appropriate time range.

mrgSnapshottingClusterId :: Lens' ModifyReplicationGroup (Maybe Text)

The cache cluster ID that will be used as the daily snapshot source for the replication group.

Response

data ModifyReplicationGroupResponse

Response constructor

modifyReplicationGroupResponse :: ModifyReplicationGroupResponse

ModifyReplicationGroupResponse constructor.

The fields accessible through corresponding lenses are:

mrgrReplicationGroup :: Maybe ReplicationGroup

Response lenses

mrgrReplicationGroup :: Lens' ModifyReplicationGroupResponse (Maybe ReplicationGroup)