Network.AWS.ElastiCache.RebootCacheCluster

Description

Reboots some, or all, of the cache nodes within a provisioned cache cluster. This operation applies any modified cache parameter groups to the cache cluster. The reboot operation takes place as soon as possible, and results in a momentary outage to the cache cluster. During the reboot, the cache cluster status is set to REBOOTING.

The reboot causes the contents of the cache (for each cache node being rebooted) to be lost.

When the reboot is complete, a cache cluster event is created.

Creating a Request

rebootCacheCluster

Arguments

:: Text
rccCacheClusterId

-> RebootCacheCluster

Creates a value of RebootCacheCluster with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rccCacheClusterId - The cache cluster identifier. This parameter is stored as a lowercase string.

rccCacheNodeIdsToReboot - A list of cache node IDs to reboot. A node ID is a numeric identifier (0001, 0002, etc.). To reboot an entire cache cluster, specify all of the cache node IDs.

data RebootCacheCluster

Represents the input of a RebootCacheCluster operation.

See: rebootCacheCluster smart constructor.

Request Lenses

rccCacheClusterId :: Lens' RebootCacheCluster Text

The cache cluster identifier. This parameter is stored as a lowercase string.

rccCacheNodeIdsToReboot :: Lens' RebootCacheCluster [Text]

A list of cache node IDs to reboot. A node ID is a numeric identifier (0001, 0002, etc.). To reboot an entire cache cluster, specify all of the cache node IDs.

Destructuring the Response

rebootCacheClusterResponse

Arguments

:: Int
rccrsResponseStatus

-> RebootCacheClusterResponse

Creates a value of RebootCacheClusterResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rccrsCacheCluster - Undocumented member.

rccrsResponseStatus - -- | The response status code.

data RebootCacheClusterResponse

See: rebootCacheClusterResponse smart constructor.

Response Lenses

rccrsCacheCluster :: Lens' RebootCacheClusterResponse (Maybe CacheCluster)

Undocumented member.

rccrsResponseStatus :: Lens' RebootCacheClusterResponse Int

- | The response status code.