Network.AWS.ElastiCache.RevokeCacheSecurityGroupIngress

Description

Revokes ingress from a cache security group. Use this operation to disallow access from an Amazon EC2 security group that had been previously authorized.

Creating a Request

revokeCacheSecurityGroupIngress

Arguments

:: Text
rcsgiCacheSecurityGroupName

-> Text
rcsgiEC2SecurityGroupName

-> Text
rcsgiEC2SecurityGroupOwnerId

-> RevokeCacheSecurityGroupIngress

Creates a value of RevokeCacheSecurityGroupIngress with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rcsgiCacheSecurityGroupName - The name of the cache security group to revoke ingress from.

rcsgiEC2SecurityGroupName - The name of the Amazon EC2 security group to revoke access from.

rcsgiEC2SecurityGroupOwnerId - The AWS account number of the Amazon EC2 security group owner. Note that this is not the same thing as an AWS access key ID - you must provide a valid AWS account number for this parameter.

data RevokeCacheSecurityGroupIngress

Represents the input of a RevokeCacheSecurityGroupIngress operation.

See: revokeCacheSecurityGroupIngress smart constructor.

Request Lenses

rcsgiCacheSecurityGroupName :: Lens' RevokeCacheSecurityGroupIngress Text

The name of the cache security group to revoke ingress from.

rcsgiEC2SecurityGroupName :: Lens' RevokeCacheSecurityGroupIngress Text

The name of the Amazon EC2 security group to revoke access from.

rcsgiEC2SecurityGroupOwnerId :: Lens' RevokeCacheSecurityGroupIngress Text

The AWS account number of the Amazon EC2 security group owner. Note that this is not the same thing as an AWS access key ID - you must provide a valid AWS account number for this parameter.

Destructuring the Response

revokeCacheSecurityGroupIngressResponse

Arguments

:: Int
rcsgirsResponseStatus

-> RevokeCacheSecurityGroupIngressResponse

Creates a value of RevokeCacheSecurityGroupIngressResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rcsgirsCacheSecurityGroup - Undocumented member.

rcsgirsResponseStatus - -- | The response status code.

data RevokeCacheSecurityGroupIngressResponse

See: revokeCacheSecurityGroupIngressResponse smart constructor.

Response Lenses

rcsgirsCacheSecurityGroup :: Lens' RevokeCacheSecurityGroupIngressResponse (Maybe CacheSecurityGroup)

Undocumented member.

rcsgirsResponseStatus :: Lens' RevokeCacheSecurityGroupIngressResponse Int

- | The response status code.