Network.AWS.ElastiCache.RevokeCacheSecurityGroupIngress

Description

The RevokeCacheSecurityGroupIngress action revokes ingress from a cache security group. Use this action to disallow access from an Amazon EC2 security group that had been previously authorized.

http://docs.aws.amazon.com/AmazonElastiCache/latest/APIReference/API_RevokeCacheSecurityGroupIngress.html

Request

data RevokeCacheSecurityGroupIngress

Request constructor

revokeCacheSecurityGroupIngress

Arguments

:: Text
rcsgiCacheSecurityGroupName

-> Text
rcsgiEC2SecurityGroupName

-> Text
rcsgiEC2SecurityGroupOwnerId

-> RevokeCacheSecurityGroupIngress

RevokeCacheSecurityGroupIngress constructor.

The fields accessible through corresponding lenses are:

rcsgiCacheSecurityGroupName :: Text

rcsgiEC2SecurityGroupName :: Text

rcsgiEC2SecurityGroupOwnerId :: Text

Request lenses

rcsgiCacheSecurityGroupName :: Lens' RevokeCacheSecurityGroupIngress Text

The name of the cache security group to revoke ingress from.

rcsgiEC2SecurityGroupName :: Lens' RevokeCacheSecurityGroupIngress Text

The name of the Amazon EC2 security group to revoke access from.

rcsgiEC2SecurityGroupOwnerId :: Lens' RevokeCacheSecurityGroupIngress Text

The AWS account number of the Amazon EC2 security group owner. Note that this is not the same thing as an AWS access key ID - you must provide a valid AWS account number for this parameter.

Response

data RevokeCacheSecurityGroupIngressResponse

Response constructor

revokeCacheSecurityGroupIngressResponse :: RevokeCacheSecurityGroupIngressResponse

RevokeCacheSecurityGroupIngressResponse constructor.

The fields accessible through corresponding lenses are:

rcsgirCacheSecurityGroup :: Maybe CacheSecurityGroup

Response lenses

rcsgirCacheSecurityGroup :: Lens' RevokeCacheSecurityGroupIngressResponse (Maybe CacheSecurityGroup)