Network.AWS.ElastiCache.V2014_07_15.DeleteCacheSubnetGroup

Description

The DeleteCacheSubnetGroup operation deletes a cache subnet group. You cannot delete a cache subnet group if it is associated with any cache clusters. https:/elasticache.amazonaws.com ?Action=DeleteCacheSubnetGroup &CacheSubnetGroupName=mysubnetgroup &Version=2014-03-24 &SignatureVersion=4 &SignatureMethod=HmacSHA256 &Timestamp=20140401T192317Z &X-Amz-Credential= 5d013245-4172-11df-8520-e7e1e602a915.

Request

data DeleteCacheSubnetGroup

Represents the input of a DeleteCacheSubnetGroup operation.

Request constructor

mkDeleteCacheSubnetGroup

Arguments

:: Text
dcsg1CacheSubnetGroupName

-> DeleteCacheSubnetGroup

Smart constructor for the minimum required parameters to construct a valid DeleteCacheSubnetGroup request.

The fields accessible through corresponding lenses are:

CacheSubnetGroupName :: Text

Request lenses

dcsg1CacheSubnetGroupName :: Lens' DeleteCacheSubnetGroup Text

The name of the cache subnet group to delete. Constraints: Must contain no more than 255 alphanumeric characters or hyphens.

Response

data DeleteCacheSubnetGroupResponse

Response constructor

mkDeleteCacheSubnetGroupResponse :: DeleteCacheSubnetGroupResponse

Smart constructor for the minimum required parameters to construct a valid DeleteCacheSubnetGroupResponse response.

This constructor is provided for convenience and testing purposes.