Network.AWS.ElastiCache.V2014_07_15.DescribeCacheSecurityGroups

Description

The DescribeCacheSecurityGroups operation returns a list of cache security group descriptions. If a cache security group name is specified, the list will contain only the description of that group. https:/elasticache.us-east-1.amazonaws.com ?Action=DescribeCacheSecurityGroups &MaxRecords=100 &Version=2014-03-24 &SignatureVersion=4 &SignatureMethod=HmacSHA256 &Timestamp=20140401T192317Z &X-Amz-Credential= default 123456789012 default mycachesecuritygroup 123456789012 My Security Group a95360ae-b7fc-11e0-9326-b7275b9d4a6c.

Documentation

describeCacheSecurityGroups :: DescribeCacheSecurityGroups

Minimum specification for a DescribeCacheSecurityGroups request.

data DescribeCacheSecurityGroups

Constructors

DescribeCacheSecurityGroups

Fields

_dcsgoMaxRecords :: Maybe Integer

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: minimum 20; maximum 100.

_dcsgoCacheSecurityGroupName :: Maybe Text

The name of the cache security group to return details for.

_dcsgoMarker :: Maybe Text

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

dcsgoMaxRecords :: Lens' DescribeCacheSecurityGroups (Maybe Integer)

dcsgoMarker :: Lens' DescribeCacheSecurityGroups (Maybe Text)

dcsgoCacheSecurityGroupName :: Lens' DescribeCacheSecurityGroups (Maybe Text)

data DescribeCacheSecurityGroupsResponse

Constructors

DescribeCacheSecurityGroupsResponse

Fields

_csgrCacheSecurityGroups :: [CacheSecurityGroup]

A list of cache security groups. Each element in the list contains detailed information about one group.

_csgrMarker :: Maybe Text

Provides an identifier to allow retrieval of paginated results.

csgrMarker :: Lens' DescribeCacheSecurityGroupsResponse (Maybe Text)

csgrCacheSecurityGroups :: Lens' DescribeCacheSecurityGroupsResponse [CacheSecurityGroup]