Network.AWS.ElastiCache.V2014_07_15.DescribeReservedCacheNodes

Description

The DescribeReservedCacheNodes operation returns information about reserved cache nodes for this account, or about a specified reserved cache node. https:/elasticache.amazonaws.com ?Action=DescribeReservedCacheNodes &ReservedCacheNodeId=customerSpecifiedID &Version=2014-03-24 &SignatureVersion=4 &SignatureMethod=HmacSHA256 &Timestamp=20140401T192317Z &X-Amz-Credential= Medium Utilization memcached 649fd0c8-cf6d-47a0-bfa6-060f8e75e95f payment-failed myreservationid 1 2010-12-15T00:25:14.131Z 31536000 227.5 0.046 cache.m1.small c695119b-2961-11e1-bd06-6fe008f046c3.

Request

data DescribeReservedCacheNodes

Represents the input of a DescribeReservedCacheNodes operation.

Request constructor

mkDescribeReservedCacheNodes :: DescribeReservedCacheNodes

Smart constructor for the minimum required parameters to construct a valid DescribeReservedCacheNodes request.

The fields accessible through corresponding lenses are:

ReservedCacheNodeId :: Maybe Text

ReservedCacheNodesOfferingId :: Maybe Text

CacheNodeType :: Maybe Text

Duration :: Maybe Text

ProductDescription :: Maybe Text

OfferingType :: Maybe Text

MaxRecords :: Maybe Integer

Marker :: Maybe Text

Request lenses

drcnReservedCacheNodeId :: Lens' DescribeReservedCacheNodes (Maybe Text)

The reserved cache node identifier filter value. Use this parameter to show only the reservation that matches the specified reservation ID.

drcnReservedCacheNodesOfferingId :: Lens' DescribeReservedCacheNodes (Maybe Text)

The offering identifier filter value. Use this parameter to show only purchased reservations matching the specified offering identifier.

drcnCacheNodeType :: Lens' DescribeReservedCacheNodes (Maybe Text)

The cache node type filter value. Use this parameter to show only those reservations matching the specified cache node type.

drcnDuration :: Lens' DescribeReservedCacheNodes (Maybe Text)

The duration filter value, specified in years or seconds. Use this parameter to show only reservations for this duration. Valid Values: 1 | 3 | 31536000 | 94608000.

drcnProductDescription :: Lens' DescribeReservedCacheNodes (Maybe Text)

The product description filter value. Use this parameter to show only those reservations matching the specified product description.

drcnOfferingType :: Lens' DescribeReservedCacheNodes (Maybe Text)

The offering type filter value. Use this parameter to show only the available offerings matching the specified offering type. Valid values: "Light Utilization" | "Medium Utilization" | "Heavy Utilization".

drcnMaxRecords :: Lens' DescribeReservedCacheNodes (Maybe Integer)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: minimum 20; maximum 100.

drcnMarker :: Lens' DescribeReservedCacheNodes (Maybe Text)

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Response

data DescribeReservedCacheNodesResponse

Represents the output of a DescribeReservedCacheNodes operation.

Response constructor

mkDescribeReservedCacheNodesResponse :: DescribeReservedCacheNodesResponse

Smart constructor for the minimum required parameters to construct a valid DescribeReservedCacheNodesResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Marker :: Maybe Text

ReservedCacheNodes :: [ReservedCacheNode]

Response lenses

drcnrMarker :: Lens' DescribeReservedCacheNodesResponse (Maybe Text)

Provides an identifier to allow retrieval of paginated results.

drcnrReservedCacheNodes :: Lens' DescribeReservedCacheNodesResponse [ReservedCacheNode]

A list of reserved cache nodes. Each element in the list contains detailed information about one node.