Network.AWS.ElastiCache.V2014_07_15.DescribeSnapshots

Description

The DescribeSnapshots operation returns information about cache cluster snapshots. By default, DescribeSnapshots lists all of your snapshots; it can optionally describe a single snapshot, or just the snapshots associated with a particular cache cluster. https:/elasticache.us-east-1.amazonaws.com ?Action=DescribeSnapshots &MaxRecords=50 &Version=2014-03-24 &SignatureVersion=4 &SignatureMethod=HmacSHA256 &Timestamp=20140401T192317Z &X-Amz-Credential= my-redis-primary 6379 cache.m1.small default.redis2.8 redis us-east-1d 2014-04-01T18:46:57.972Z 2.8.6 manual true wed:09:00-wed:10:00 my-manual-snapshot 5 2014-04-01T18:54:12Z 2014-04-01T18:46:57.972Z 0001 3 MB creating 1 07:30-08:30 51b0b25e-b9cf-11e3-8a16-7978bb24ffdf.

Request

data DescribeSnapshots

Represents the input of a DescribeSnapshotsMessage operation.

Request constructor

mkDescribeSnapshots :: DescribeSnapshots

Smart constructor for the minimum required parameters to construct a valid DescribeSnapshots request.

Request lenses

ds1CacheClusterId :: Lens' DescribeSnapshots (Maybe Text)

A user-supplied cluster identifier. If this parameter is specified, only snapshots associated with that specific cache cluster will be described.

ds1SnapshotName :: Lens' DescribeSnapshots (Maybe Text)

A user-supplied name of the snapshot. If this parameter is specified, only this snapshot will be described.

ds1SnapshotSource :: Lens' DescribeSnapshots (Maybe Text)

If set to system, the output shows snapshots that were automatically created by ElastiCache. If set to user the output shows snapshots that were manually created. If omitted, the output shows both automatically and manually created snapshots.

ds1Marker :: Lens' DescribeSnapshots (Maybe Text)

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

ds1MaxRecords :: Lens' DescribeSnapshots (Maybe Integer)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved. Default: 50 Constraints: minimum 20; maximum 50.

Response

data DescribeSnapshotsResponse

Represents the output of a DescribeSnapshots operation.

Response lenses

dsrrMarker :: Lens' DescribeSnapshotsResponse (Maybe Text)

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

dsrrSnapshots :: Lens' DescribeSnapshotsResponse [Snapshot]

A list of snapshots. Each item in the list contains detailed information about one snapshot.